==> scripts/1_processing_cleaned up.Rmd <==
---
title: "Combining WC, WP, LFM, and RWC data into one datasheet"
author: "Indra Boving"
date: "3/21/2024"
output: html_document
---

##Setup

```{r setup, include=FALSE}
library(data.table)
library(janitor)
library(here)
library(tidyverse)
library(lubridate)
library(readxl)
library(gridExtra)
#library(calecopal)
library(rstatix)
library(plotly)

select = dplyr::select
 
#source("http://goo.gl/UUyEzD") #outlier KD (original function)

source(here::here("scripts","scripts_functions", "figure_info.R")) #if making pretty figs 
```

#Env. data: 

From dendra: https://dendra.science/orgs/ucnrs/datastreams?faceted=true&scheme=dq&selectStationId=58e68cacdf5ce600012602d6 
```{r}
shift_env <- read_csv(here::here("data", "sedgwick-during-shift.csv"), show_col_types = FALSE) %>% 
  clean_names() %>% 
  mutate(adt_time = parse_date_time(time,  "m/d/y H:M"), 
         date = as.Date(adt_time), 
         month = month(date))
```

```{r}
## Data file version (so it's not hard coded in every read_excel() call)
datver <- "20230724"
dataversion <- paste0("Data_", datver)
```

#------------------
#Rerun processing:

Run all scripts for processing to make sure those are updated: 

(NOTE: if changing datver, will have to change in each of these scripts)

```{r, warning=FALSE, eval=FALSE}
 source(here::here("scripts", "0_processing-scripts", "leaf_area_processing.R"))
 source(here("scripts","0_processing-scripts", "alas_processing.R"))
 source(here::here("scripts","0_processing-scripts", "lfm_processing.R"))
datver_wp <- "20231128" #this was hard coded into md_pd_pheno_processing, pulled out here to make sure this wasn't an error
source(here::here("scripts","0_processing-scripts", "md_pd_pheno_processing.R"))
 source(here::here("scripts","0_processing-scripts", "wc_processing.R"))
 source(here::here("scripts","0_processing-scripts", "swc_processing.R"))

#use below if you want to do them all at once.
#list.files("processing-scripts", full.names = TRUE) %>% walk(source)
```

#Tree info:

```{r}
trees_sites_df <- wc_df <- read_csv(here("processed-data", paste0("wc_alldates_",datver,".csv")), show_col_types = FALSE) %>% 
  ungroup() %>% 
  select(tree, plot, site, species) %>% 
  distinct()

write_csv(trees_sites_df, here("processed-data", "trees_sites.csv"))
```

#Combine DFs: 

###New LWA:
Using total leaf masses (ww and dw) and total scan (sum of all individually scanned leaves, when that was relevant)

Calculate LMA and see they make sense and if they dont then those are things we're missing. 

Problem where there are two envelopes for some tree/time combos in week 37

```{r}
#This has the updated dry scans, from which we can get an updated LWA to compare with our LWA approximations: 
#From thwh 'SHIFT LWC Areas & Masses' sheet
datver_lwa <- "20240321" #where the lwa info is located

#Raw Area data: 
lwa_raw0 <- read_csv(here(paste0("Data_",datver_lwa), "shift_lwa_new.csv"), show_col_types = FALSE) %>% 
  clean_names() %>% 
  mutate(tree_id = case_when(tree_id %in% c(1475) ~ 1478, #screwed up tree ids
                          tree_id %in% c(2082) ~ 2382, 
                          tree_id %in% c(2385) ~ 2085, 
                          TRUE ~ as.numeric(tree_id))) %>% 
  #filter(is.na(bulk_or_individual_leaf)) %>%
  mutate(bulk_or_individual_leaf_old = bulk_or_individual_leaf) %>% #make sure bulk vs. individual leaves are correctly coded; this is important because bulk leaves will not be summarized over a single date. 
  group_by(dry_scan_area, tree_id) %>%
  mutate(bulk_or_individual_leaf = case_when(
    is.na(dry_scan_area) ~ NA_character_,
    n() > 1 | bulk_or_individual_leaf_old %in% c("bulk") ~ "bulk",
    TRUE ~ "leaf"
  )) %>%
  ungroup() %>% 
  mutate(dry_scan_area = case_when(
    date %in% c("2022-04-04") & tree_id %in% c(2086) & pd_or_md %in% c("md") ~ 10.109, #entered wrong in datasheet (entered as summed bulk, but was missing one leaf entry)
    TRUE ~ as.numeric(dry_scan_area))
  ) %>% 
  # filter(!(date %in% c("9/12/22", "9/15/22", "9/16/22") & tree_id %in% c(1478, 2006, 2012,2015) & notes %in% c("y0, area from rstudio"))) %>% 
  mutate(tree = as.numeric(tree_id))# %>% 
 # filter(date %in% c("4/25/22", "4/27/22"))

##Add in species and site and plot info! 

lwa_raw <- merge(trees_sites_df, lwa_raw0, by = c("tree"))

##Clean up raw area data so that we are clear which scans are for individual leaves and which are for bulk samples, convert the dry scans using the scaling coefficient we measured by rescanning some already-scanned dry leaves
lwa_new_df1 <- lwa_raw %>% 
    clean_names() %>% 
  mutate(tree = tree_id, 
         time = pd_or_md,
         date_wc = lubridate::mdy(date), 
         week = week(date_wc), 
         rep = leaf_number) %>% 
  select(tree, date_wc, time, week, rep, lwc_leaf, lwc_bulk, dry_scan_area, bulk_or_individual_leaf, species) %>% 
  group_by(tree) %>% 
  fill(species, .direction = "downup") %>% 
  ungroup() %>% 
  mutate(area_cm2 = case_when(
    species %in% c("blue oak") & week <= 17 ~ dry_scan_area*1.166899, 
    species %in% c("live oak") & week <= 17 ~ dry_scan_area*1.126367, 
    species %in% c("blue oak") & week > 17 ~ dry_scan_area*1.068375, 
    species %in% c("live oak") & week > 17 ~ dry_scan_area*1.058706, 
    TRUE ~ as.numeric(dry_scan_area)
    )) %>%  #1.166899 = dry to wet mass conversion based on rescanned leaves from al:as collectd on 4/4/22), in leaf_area_processes.R script
  distinct()

#### For all leaves that were scanned in bulk, rename that scanned area as "sum_area_cm2":
lwa_bulk_df_all <- lwa_new_df1 %>% 
  filter(bulk_or_individual_leaf %in% c("bulk"), 
         !(week == 37)) %>% #week 37 (september samplings) have two entries for some reason! I dont know why, but we'll do those separately.
  mutate(sum_area_cm2 = area_cm2) %>% 
  select(-rep, -lwc_leaf) %>%
  distinct()

#sometimes bulk isn't actually bulk
lwa_bulk_multiples <- lwa_bulk_df_all %>% 
  group_by(tree, week, time) %>% 
  filter(n() > 1) |>
  mutate(sum_area_cm2 = sum(area_cm2))

#sometimes bulk is actually bulk!
lwa_bulk_singles <- lwa_bulk_df_all %>% 
  group_by(tree, week, time) %>% 
  filter(n() == 1)
  
 lwa_bulk_df <- bind_rows(lwa_bulk_singles, lwa_bulk_multiples)

####week 37, where they're listed as bulk but dont have a good reason to be separated. 
lwa_wk_37 <- lwa_new_df1 %>% 
  filter(week == 37) %>% 
  group_by(tree, week, time, date_wc) %>% 
   #select(-lwc_leaf) %>% 
  mutate(sum_area_cm2 = sum(area_cm2, na.rm = T)) %>%  #sum of area of all the leaves for a given day and time
 # select(tree, date_wc, week, sum_area_cm2, time) %>% 
  ungroup() %>% 
#  select(sum_area_cm2, tree, week, time, date_wc) %>% 
  distinct()

lwa_wk_37_screwedup <- lwa_wk_37 %>% 
  filter(tree %in% c(2085, 2005, 2006, 2015, 2088, 2013, 2004, 2007, 2370, 2012))

#### For all individual leaves that were scanned separately, find the sum of them for each tree/week/time, and call this "sum_area_cm2":

lwa_leaves_df <- lwa_new_df1 %>% 
  filter(bulk_or_individual_leaf %in% c("leaf"),
         !(week == 37)) %>%
  group_by(tree, week, time, date_wc) %>% 
   #select(-lwc_leaf) %>% 
  mutate(sum_area_cm2 = sum(area_cm2, na.rm = T)) %>%  #sum of area of all the leaves for a given day and time
 # select(tree, date_wc, week, sum_area_cm2, time) %>% 
  ungroup() %>% 
#  select(sum_area_cm2, tree, week, time, date_wc) %>% 
  distinct()


##Combine the bulk and the leaves summed into one df: 
lwa_new_df2 <- bind_rows(lwa_leaves_df, lwa_bulk_df, lwa_wk_37) %>% 
  select(-lwc_leaf, 
         -lwc_bulk, 
         #-dry_scan_area, #-bulk_or_individual_leaf, 
         -area_cm2, 
         -rep) %>% 
  distinct() %>% 
  drop_na(sum_area_cm2)

#--------

#this is new scans from trees we didn't get to the first time around, should all be bulk. 

#Round 1: 
entered_scans <- read_csv(here::here("data", "leaf_area_scans", "updated_dry_scans_shift_20240504.csv")) %>% 
  clean_names()

entered_scans_721 <- entered_scans %>% 
  filter(date %in% c("7/21/22")) %>% 
  separate(source, into = c("species", "tree", "scan_info"), sep = "_") %>% 
  select(tree, total_area, date, time) %>% 
  mutate(date = "7/19/22") %>% #make this match teh dates that were entered
  mutate(time = case_when(
    tree %in% c(1478) & total_area == 14.608 ~ "pd", 
    tree %in% c(2015) & total_area ==  23.015 ~ "pd", #one of these 2015s is actually a PD but labled as MD see which is an oddball when running stats/figs
   # tree %in% c(2015) & total_area == 10.586 ~ "pd", 
    TRUE ~ as.character(time)
  ))

entered_scans_523 <- entered_scans %>% 
  filter(date %in% c("5/23/22")) %>% 
  separate(source, into = c("tree", "time", "scan_info"), sep = "_") %>% 
  select(tree, total_area, date, time)# %>% 
 # mutate(date = "5/25/22") #make this match teh dates that were entered

entered_scans_525 <- entered_scans %>% 
  filter(date %in% c("5/25/22")) %>% 
  separate(source, into = c("tree", "time", "scan_info"), sep = "_") %>% 
  select(tree, total_area, date, time)

entered_scans_916 <- entered_scans %>% 
  filter(date %in% c("9/16/22")) %>% 
  separate(source, into = c("tree", "time", "scan_info"), sep = "_") %>% 
  select(tree, total_area, date, time) %>% 
  mutate(date = "9/15/22") #make this match teh dates that were entered

entered_scans_all <- bind_rows(entered_scans_523, entered_scans_525, entered_scans_721, entered_scans_916) %>% 
  mutate(time = case_when(
    time %in% c("MD") ~ "md", 
    time %in% c("PD") ~ "pd", 
    TRUE ~ as.character(time)
  )) %>% 
  mutate(date_wc = mdy(date), 
         week = week(date_wc),
         sum_area_cm2 = total_area, 
         bulk_or_individual_leaf = "bulk", 
         tree = as.numeric(tree)) %>% 
  select(-total_area, -date) %>% 
  merge(trees_sites_df, by = c("tree")) %>% 
  mutate(sum_area_cm2 = case_when(
    species %in% c("blue oak") & week <= 17 ~  sum_area_cm2*1.166899, 
    species %in% c("live oak") & week <= 17 ~  sum_area_cm2*1.126367, 
    species %in% c("blue oak") & week > 17 ~  sum_area_cm2*1.068375, 
    species %in% c("live oak") & week > 17 ~  sum_area_cm2*1.058706, 
    TRUE ~ as.numeric(sum_area_cm2))) 


##Round 2: 
entered_scans_2 <- read_csv(here::here("data", "leaf_area_scans", "total_area_330_425_54.csv")) %>% 
  clean_names() %>% 
  rename("time" = "pd_or_md")

entered_scans_330 <- entered_scans_2 %>% 
  filter(date %in% c("2022-03-30")) %>% 
  separate(source, into = c("species", "tree", "scan_info"), sep = "_") %>% 
 # mutate(time = "pd") %>% 
  select(tree, total_area, date, time) 

entered_scans_504 <- entered_scans_2 %>% 
  filter(date %in% c("2022-05-04")) %>% 
  separate(source, into = c("species", "tree", "scan_info"), sep = "_") %>% 
 # mutate(time = "md") %>% 
  select(tree, total_area, date, time) 

entered_scans_425_pd <- entered_scans_2 %>%
  filter(date %in% c("2022-04-25"), 
         time == "pd") %>%
  separate(source, into = c("species", "tree", "scan_info"), sep = "_") %>%
  select(tree, total_area, date, time)

entered_scans_425_md <- entered_scans_2 %>%
  filter(date %in% c("2022-04-25"), 
         time == "md") %>%
  separate(source, into = c("species", "tree", "scan_info"), sep = "_") %>%
  select(tree, total_area, date, time)

entered_scans_all2 <- bind_rows(entered_scans_330, entered_scans_504, 
                                entered_scans_425_md, entered_scans_425_pd
                                ) %>% 
  mutate(date_wc = ymd(date), 
         week = week(date_wc),
         sum_area_cm2 = total_area, 
         bulk_or_individual_leaf = "bulk", 
         tree = as.numeric(tree)) %>% 
  select(-total_area, -date) %>% 
  merge(trees_sites_df, by = c("tree")) %>% 
  mutate(sum_area_cm2 = case_when(
    species %in% c("blue oak") & week <= 17 ~  sum_area_cm2*1.166899, 
    species %in% c("live oak") & week <= 17 ~  sum_area_cm2*1.126367, 
    species %in% c("blue oak") & week > 17 ~  sum_area_cm2*1.068375, 
    species %in% c("live oak") & week > 17 ~  sum_area_cm2*1.058706, 
    TRUE ~ as.numeric(sum_area_cm2))) 

#coombine newer scans and older scans, this DF will have all scanned leaves. 
lwa_new_df <- bind_rows(lwa_new_df2, entered_scans_all, entered_scans_all2) %>% 
  distinct() %>% 
  drop_na(sum_area_cm2) %>% 
  select(-dry_scan_area) %>% 
  distinct() %>% 
  select(-plot, -site) %>% 
  mutate(week = case_when(
    week == 38 ~ 37, 
    TRUE ~ as.numeric(week)
  )) %>% 
  mutate() %>% 
  mutate(date_wc = case_when(
    date_wc %in% c("2022-05-25") ~ as.Date("2022-05-23"), 
    date_wc %in% c("2022-09-16") ~ as.Date("2022-09-15"), 
    TRUE ~ as.Date(date_wc))) %>% 
  select(-week) %>% 
  arrange(date_wc, tree)

#these are problems ones to go back to!
lwa_problems <- lwa_new_df %>% 
  group_by(tree, date_wc, time) %>% 
  filter(n() > 1)
```

#####WC:
Now this is to get the dry weight info in a condition with which to attach it to the updated scans: 

Water content data looks like this: 

md_bulk_wet = as.numeric(md_bulk_wet), 
lwc_md_bulk = ((md_bulk_wet-md_bulk_dry)/md_bulk_dry)
lwc_pd_bulk = ((pd_bulk_wet - pd_bulk_dry)/pd_bulk_dry)
lwc_md_leaf = ((ww_g_md - dw_g_md)/dw_g_md)
lwc_pd_leaf = ((ww_g_pd - dw_g_pd)/dw_g_pd)) 

Things to note: 
Week 9: We only have individual leaf lwc, not bulk. Not many trees have BOTH PD and MD values. 
Week 10: Some trees have multiple reps for bulk and no leaf, maybe this is fine since we separate bulk and leaves later and get distict rows? 
Week 11: looks like we only got PDs for a lot of trees here; might be resolved later by combining weeks? 
Week 12: 2022-03-25; Switched to not doing leaves, only doing bulk  
Week 13: All PD measurement from 2022-03-30
Week 14: 2022-04-04 and 2022-04-06; back to both leaf and bulk; seems to be a lot fo trees with only PD
Week 15: 2022-04-11 and 2022-04-13; looks like we got both bulk and leaves, and did a better job of getting both PD and MD
Week 16:NA
Week 17: 2022-04-25; just bulk, bth PD and MD present
Week 18: 2022-05-04; just MDs
Wewk 19: 2022-05-09; mix of PD and MD, all bulk
Week 20: NA
Week 21: 2022-05-23, some trees missing MDs
Week 29: 2022-07-19, decent job getting all, all bulk
Week 37: 2022-09-12 & 2022-09-15 all bulk; looks good except for some random missing info 

Tree 2384 seems to be missing a lot; did it get confused with another tree? 

```{r}
wc_df <- read_csv(here("processed-data", paste0("wc_alldates_",datver,".csv")), show_col_types = FALSE) %>% #contains LWC data
  mutate(date = lubridate::ymd(date))%>% 
  select(-...1) %>% 
  mutate(date_wc = date) %>% 
  select(-date, -notes) %>% 
  # filter_at(vars(lwc_bulk, dm_bulk_g, wm_bulk_g, lwc_leaf, dm_leaf_g, wm_leaf_g),any_vars(!is.na(.))) %>% 
  #filter(!(duplicated(lwc_bulk))) %>% 
  filter(!(week == 40)) %>% #these are Serpentine trees where we just have midddays, remove from this analysis
  group_by(week) %>% 
  fill(c("date_wc"), .direction = "downup") %>% #some weeks dont have dates, fill those in
  #filter(is.na(date_wc)) %>% 
  ungroup() %>% 
  #filter(!(week %in% 37 & tree %in% c(2382, 2383, 2384, 1478, 2015, 2012, 2006))) %>% #these have duplicates, need to look at envelopes. 
  distinct() %>% 
  mutate(wm_leaf_g = case_when(
   wm_leaf_g %in% c(1.6180) ~ (0.1618),
   TRUE ~ as.numeric(wm_leaf_g)
  ),
 lwc_leaf = case_when(
   tree %in% c(2345) & week %in% c(10) & rep %in% c(2) & time %in% c("md")~ (0.96867),
   TRUE ~ as.numeric(lwc_leaf )
  )
  )

###First the water weights (how much water is in bulk leaves). 

#####For this, each tree, time (pd or md) and week has 1 value that is in the sum of all the bulk + leaf measurements

wc_lwa_long <- wc_df %>% 
  #filter(week == 13) %>% 
  mutate(water_leaves = wm_leaf_g - dm_leaf_g, 
         water_bulk = wm_bulk_g - dm_bulk_g) %>% 
  select(-rep, -dm_bulk_g, -dm_leaf_g, -lwc_bulk, -lwc_leaf, -wm_bulk_g, -wm_leaf_g) %>% 
  pivot_longer(cols = c(water_leaves, water_bulk)) %>% 
  distinct() %>% 
 # select(-lwc_bulk, -lwc_leaf) %>% 
  drop_na(value) %>% 
  group_by(tree, time, week, date_wc) %>%  #sum of all the weights of leaves for a given date/time
  mutate(sum_water_wt_g = sum(value, na.rm = T)) %>% 
  select(1:6, date_wc, sum_water_wt_g) %>% 
  distinct() %>% 
  drop_na(date_wc)

###Next the dry weights (how much dry matter on a gram basis is in bulk leaves). For this: 
#####- each tree, time (pd or md) and week has 1 value that is in the sum of all the bulk + leaf measurements

wc_lma_long <- wc_df %>% 
  select(-lwc_bulk, -lwc_leaf, -rep,-wm_bulk_g, -wm_leaf_g) %>% 
  pivot_longer(cols = c(dm_leaf_g, dm_bulk_g)) %>% 
  drop_na(value) %>% 
  distinct() %>% 
  ungroup() %>% 
  group_by(tree, time, week, date_wc) %>%  #sum of all the weights of leaves for a given date/time
  mutate(sum_dm_g = sum(value, na.rm = T)) %>% 
  #select(1:6, date_wc, sum_dm_g, lwc_bulk, lwc_leaf) %>% 
  select(-value, -name)  %>% 
    distinct()

#Attach those two together, calculate a new LWC metric based on these combined values (should be similar to previously calculated LWC, which was either bulk or leaves or average of the two)

######New LWC measurement df: 
wc_long_lwa_all1 <- merge(wc_lma_long, wc_lwa_long) %>% #combine dry mass datafrome with water weight dataframe
  mutate(lwc_new = sum_water_wt_g/sum_dm_g) %>% #make a new lwc_df
  distinct()

####Old LWC measurement df: 
wc_long_lwa_old <- wc_df %>% 
  select(tree, rep, week, time, date_wc, lwc_bulk, lwc_leaf, dm_bulk_g, wm_bulk_g, dm_leaf_g, wm_leaf_g)


####Combine the newly calculated ones and the old ones together and see if there is a mismatch 
wc_long_lwa_all2 <- merge(wc_long_lwa_all1, wc_long_lwa_old, 
                          by = c("date_wc", "tree", "time", "week")) %>% 
  mutate(mismatch_bulk = lwc_new - lwc_bulk, 
         mismatch_leaf = lwc_new - lwc_leaf) %>% 
 # group_by(tree, date_wc, time, week) %>% 
#  rowwise() %>% 
 # mutate(lwc_mean = mean(lwc_bulk, lwc_leaf, na.rm = TRUE)) %>% 
 # ungroup() %>% 
  mutate(lwc_wrong = case_when(
    mismatch_bulk > .5 ~ "off", 
    mismatch_bulk < -0.5 ~ "off", 
    TRUE ~ "probably correct"
  )) %>% 
  group_by(tree, date_wc, time) %>% 
  fill(c("plot", "site", "species"), .direction = "downup") %>% 
  distinct() %>% 
  mutate(date_wc = case_when(
    date_wc %in% c("2022-09-16") ~ as.Date("2022-09-15"), 
    TRUE ~ as.Date(date_wc)
  ))

#whats up with specific weeks? 
wc_long_lwa_wk_37 <- wc_long_lwa_all2  %>% 
  filter(week == 37)
wc_long_lwa_wk_13 <- wc_long_lwa_all2  %>% 
  filter(week == 13)

#####These are ones that we should be skeptical of: 
mismatched_lwc_lwa <- wc_long_lwa_all2 %>% 
  filter(!(week == 19)) %>% #week 19 is just wonky because using new leaf weights
  select(-lwc_leaf, -mismatch_leaf) %>% 
  distinct() %>% 
  filter(lwc_wrong == "off")

plotly::ggplotly(wc_long_lwa_all2 %>%  #masses look correct
                 #filter(week == 33) %>%
                  # drop_na(sum_area_cm2) %>%
  ggplot(aes(y = lwc_new,
             x = lwc_bulk,
             color = species,
             shape = lwc_wrong, 
             label = tree)) +
  geom_point()+
  facet_wrap(~week, scales = "free") +
    geom_abline()
  ) 

##Add summed water weight, LWC, and dry weight data to the LWA df that contains area data: 
wc_lwa_df <- merge(wc_long_lwa_all2, lwa_new_df, 
                   by = c("tree", "time", "date_wc", "species"), 
                   all = T) %>% 
  # select(tree, time, rep, week, date_wc, dry_scan_area, dm_leaf_g, wm_leaf_g, dm_bulk_g, wm_bulk_g) %>% 
  filter(tree > 100, # we dont care about shrubs right now
         sum_dm_g > 0, #these cant be correct (no sums below zero make sense)
         !(date_wc %in% c("2022-02-28"))) %>% #we'll ignore this date, since we didn't scan these leaves
  distinct() %>% 
  mutate(lwa_g_cm2_new = sum_water_wt_g/sum_area_cm2, 
         lma_g_cm2_new = sum_dm_g/sum_area_cm2) %>% 
  select(-lwc_leaf, -mismatch_leaf) %>% 
  distinct() %>% 
  arrange(date_wc, tree) #%>% 
  #drop_na(species)

#Visualize to see if it makes sense

plotly::ggplotly(wc_lwa_df  %>%  #masses look correct
                 #filter(week == 33) %>% 
                   drop_na(sum_area_cm2) %>% 
  ggplot(aes(y = lma_g_cm2_new, 
             x = lwa_g_cm2_new, 
             color = species, 
             shape = time, 
             label = tree)) +
  geom_point()+
  facet_wrap(~date_wc*week))

plotly::ggplotly(wc_lwa_df  %>% 
  #filter(week == 37) %>% 
  drop_na(sum_area_cm2) %>% 
  ggplot(aes(y = sum_dm_g, 
             x = sum_area_cm2, 
             color = species,
             shape = time, 
             label = tree)) +
  geom_point()+
  facet_wrap(~date_wc*week))

plotly::ggplotly(wc_lwa_df  %>% 
  #filter(week == 37) %>% 
  drop_na(sum_area_cm2) %>% 
  ggplot(aes(y = lwa_g_cm2_new, 
             x = lwc_new, 
             color = species,
             shape = time,
             label = tree)) +
  geom_point()+
  facet_wrap(~date_wc*week))

#Check for weirdos in the LWC data (where there are mismatches)
plotly::ggplotly(wc_lwa_df %>% 
  #filter(week == 37) %>% 
  ggplot(aes(y = lwc_new, #new LWC calc
             x = lwc_bulk, #old LWC calc
             color = species,
             shape = lwc_wrong, 
             label = tree)) +
  geom_point()+
  facet_wrap(~time, scales = "free") +
  geom_abline())


plotly::ggplotly(wc_lwa_df  %>% 
  #filter(week == 33) %>% 
  ggplot(aes(y = lwc_new, #new LWC calc
             x = lwc_bulk, #old LWC calc
             color = species,
             shape = lwc_wrong, 
             label = tree)) +
  geom_point()+
  facet_wrap(~week, scales = "free") +
  geom_abline())
```



```{r, eval = F}
#THIS IS THE DF THAT IS GETTING MISSINGS ENTERED INTO: 
#--------
 wc_lwa_df %>% 
  select(tree, week, date_wc, time, species, dm_bulk_g, wm_bulk_g, dry_scan_area, sum_area_cm2) %>%
  distinct() %>% 
  drop_na(species) %>% 
  mutate(scan_done = case_when(
    dry_scan_area > 0 ~ "done", 
    TRUE ~ as.character("scan needed")
  )) %>% 
  write_csv(here::here("processed-data", "lwa_toenter_20240422.csv"))
```

######*REMOVING WC and LWA OUTLIERS: 
```{r}
#remove obviously off values, make NA: 
lwa_df <- wc_lwa_df %>% 
 # select(-rep) %>% 
  #distinct()
  
  #LWA new: any values with wrong areas and wrong summed masses
  mutate(lwa_g_cm2_new = case_when(
        week %in% c(13) & tree %in% c(2009, 2370) & time %in% c("pd") ~ NA_real_, 
    date_wc %in% c("2022-04-04") & tree %in% c(2086) & time %in% c("md") ~ NA_real_, 
     date_wc %in% c("2022-03-08") & tree %in% c(2345) & time %in% c("md") ~ NA_real_, 
    date_wc %in% c("2022-03-08") & tree %in% c(2377) & time %in% c("md") ~ NA_real_, 
    lwc_wrong %in% c("off") ~ NA_real_,
    week %in% c(19) & tree %in% c(2020) & time %in% c("pd") ~ NA_real_, #LWA is way too high, check scan? 
    tree %in% c(2383, 2384) & week %in% c(37) ~ NA_real_,
    tree %in% c(2346, 2381) & week %in% c(15) & time %in% c("md") ~ NA_real_,
    tree %in% c(2370) & week %in% c(15) & time %in% c("pd") ~ NA_real_,
    tree %in% c(2009) & week %in% c(13) ~ NA_real_, #this area is wrong, much too low - lost a leaf? 
    tree %in% c(2086) & week %in% c(14) & time %in% c('md') ~ NA_real_,
   #  tree %in% c(2382) & week %in% c(19) & time %in% c('pd')~ NA_real_, #not an issue with area, just with weights
  tree %in% c(2343) & week %in% c(15) & time %in% c('md') ~ NA_real_, 
  tree %in% c(2009) & week %in% c(29) & lwa_g_cm2_new > 0.02 ~ NA_real_, #two rows, one is off so lets delete that one
  tree %in% c(2347) & week %in% c(29) & lwa_g_cm2_new > 0.02 ~ NA_real_,#two rows, one is off so lets delete that one
  week %in% 37 & tree %in% c(2382, 2383, 2384, 1478, 2015, 2012, 2006) & time %in% c("pd") ~ NA_real_,  #areas area off for these
    TRUE ~ as.numeric(lwa_g_cm2_new)
  ))%>% 
  
  
#LMA new: any values with wrong areas
  mutate(lma_g_cm2_new = case_when(
    lwc_wrong %in% c("off") ~ NA_real_,
    week %in% c(13) & tree %in% c(2009, 2370) & time %in% c("pd") ~ NA_real_, 
    date_wc %in% c("2022-04-04") & tree %in% c(2086) & time %in% c("md") ~ NA_real_, 
     date_wc %in% c("2022-03-08") & tree %in% c(2345) & time %in% c("md") ~ NA_real_, 
    date_wc %in% c("2022-03-08") & tree %in% c(2377) & time %in% c("md") ~ NA_real_, 
    week %in% c(19) & tree %in% c(2020) & time %in% c("pd") ~ NA_real_, #predawn LWA is weirdly super high
    week %in% c(21) & tree %in% c(2378) & time %in% c("pd") ~ NA_real_, #predawn LWA is weirdly super high
    tree %in% c(2383, 2384) & week %in% c(37) ~ NA_real_,
    tree %in% c(2346, 2381) & week %in% c(15) & time %in% c("md") ~ NA_real_,
    tree %in% c(2370) & week %in% c(15) & time %in% c("pd") ~ NA_real_,
    tree %in% c(2009) & week %in% c(13) ~ NA_real_, #this area is wrong!
    tree %in% c(2009) & week %in% c(29) & lwa_g_cm2_new > 0.02 ~ NA_real_, #two rows, one is off so lets delete that one
      tree %in% c(2347) & week %in% c(29) & lwa_g_cm2_new > 0.02 ~ NA_real_, #two rows, one is off so lets delete that one
    tree %in% c(2086) & week %in% c(14) & time %in% c('md') ~ NA_real_,
    week %in% 37 & tree %in% c(2382, 2383, 2384, 1478, 2015, 2012, 2006) & time %in% c("pd") ~ NA_real_,  #areas area off for these
    # tree %in% c(2382) & week %in% c(19) & time %in% c('pd')~ NA_real_,
    tree %in% c(2343) & week %in% c(15) & time %in% c('md') ~ NA_real_,
    TRUE ~ as.numeric(lma_g_cm2_new)
  )) %>% 
  
  
#LWC: issues with weights
  mutate(lwc_new = case_when(
    lwc_wrong %in% c("off") ~ NA_real_,
    tree %in% c(2383, 2384) & week %in% c(37) ~ NA_real_,
    tree %in% c(2086) & week %in% c(14) & time %in% c('md') ~ NA_real_,
     tree %in% c(2382) & week %in% c(19) & time %in% c('pd')~ NA_real_,
    tree %in% c(2343) & week %in% c(15) & time %in% c('md') ~ NA_real_,
     tree %in% c(2370) & week %in% c(29) & time %in% c("md") ~ NA_real_,
      tree %in% c(2331) & week %in% c(15) & time %in% c("pd") ~ NA_real_,
    tree %in% c(2091) & week %in% c(19) & time %in% c("md") ~ NA_real_,
    tree %in% c(2354) & week %in% c(15) & time %in% c("md") ~ NA_real_,
    tree %in% c(2345) & week %in% c(10) & time %in% c("md") ~ NA_real_, 
    TRUE ~ as.numeric(lwc_new)
  )) %>% 
  
  
#Issue with weights, 
  mutate(lwc_bulk = case_when(
    lwc_wrong %in% c("off") ~ NA_real_,
    tree %in% c(2383, 2384) & week %in% c(37) ~ NA_real_,
    tree %in% c(2086) & week %in% c(14) & time %in% c('md') ~ NA_real_,
     tree %in% c(2382) & week %in% c(19) & time %in% c('pd')~ NA_real_,
    tree %in% c(2343) & week %in% c(15) & time %in% c('md') ~ NA_real_,
      tree %in% c(2370) & week %in% c(29) & time %in% c("md") ~ NA_real_,
    tree %in% c(2331) & week %in% c(15) & time %in% c("pd") ~ NA_real_,
    tree %in% c(2091) & week %in% c(19) & time %in% c("md") ~ NA_real_,
    tree %in% c(2354) & week %in% c(15) & time %in% c("md") ~ NA_real_,
    tree %in% c(2345) & week %in% c(10) & time %in% c("md") ~ NA_real_, 
    
    TRUE ~ as.numeric(lwc_bulk)
  )) %>% 
  select(-mismatch_bulk, -lwc_wrong) %>% 
  distinct() %>% 
  mutate(week = case_when(
    week %in% c(38) ~ 37, 
    TRUE ~ as.numeric(week)
  )) %>% 
  mutate(date = date_wc) %>% 
#when we didn't measure exactly on the same day, but want predawns and middays to overlap, change the dates (need to do this for MPas too)
  mutate(date = case_when( 
    date %in% c("2022-05-25") ~ as.Date("2022-05-23"), 
    date %in% c("2022-04-27") ~ as.Date("2022-04-25"), 
    date %in% c("2022-09-16") ~ as.Date("2022-09-15"), 
    TRUE ~ as.Date(date))) %>% 
  mutate(tree = as.numeric(tree), 
         tree = case_when(
     tree == 2127 ~ 2327, #weirdly named trees
     tree == 2309 ~ 2379, 
     tree == 2353 ~ 2352, 
     tree == 2374 ~ 2376, 
     tree == 2364 ~ 2360, 
     tree == 2004 & date %in% c("2022-07-19") ~ 2005,
     tree == 2014 & date %in% c("2022-09-15") ~ 2013,
     species %in% c("ARCA") & site %in% c("Cucu Mesa") ~ 12,
     TRUE ~ as.numeric(tree))) %>% 
  mutate(week = case_when(
    date %in% c("2022-03-30") ~ 14, #this was being listed as week 13, but I think it should actually be 14 based on 4/4 and 4/6 samplings having some midday trees that correspond to the predawns measured on this day. 
    TRUE ~ week(date))) 

investigating_wk10 <- lwa_df %>% 
  filter(week == 10) %>% 
  distinct()

unique(investigating_wk10$tree)

plotly::ggplotly(lwa_df %>% 
  #filter(week == 33) %>% 
  ggplot(aes(y = lwc_new, #new LWC calc
             x = lwc_bulk, #old LWC calc
             color = species,
             shape = time,
             #shape = lwc_wrong, 
             label = tree)) +
  geom_point()+
  facet_wrap(~week, scales = "free") +
  geom_abline())

plotly::ggplotly(lwa_df %>% 
  #filter(week == 33) %>% 
  ggplot(aes(y = lma_g_cm2_new, #new LWC calc
             x = lwc_bulk, #old LWC calc
             color = species,
             shape = time,
             #shape = lwc_wrong, 
             label = tree)) +
  geom_point()+
  facet_wrap(~week, scales = "free"))

plotly::ggplotly(lwa_df %>% 
  #filter(week == 33) %>% 
  ggplot(aes(y = lma_g_cm2_new, #new LWC calc
             x = lwa_g_cm2_new, #old LWC calc
             color = species,
             shape = time,
             #shape = lwc_wrong, 
             label = tree)) +
  geom_point()+
  facet_wrap(~week*date))

##These actually all look good now! Yay! lwc_wrong also doesn't stand out as supppper wierd except in the plot comparing lwc_new to our older lwc_bulk, so we'll keep them in for now.
```


###WP:
```{r}
#Read in raw water potential data, and change weird trees: 
wp_df <- read_csv(here("processed-data", paste0("wp_alldates_long_",datver,".csv")), show_col_types = FALSE)  %>%
  select(-...1) %>% 
#weirdly named trees
  mutate(tree = as.numeric(tree), 
         tree = case_when(
     tree == 2127 ~ 2327, 
     tree == 2309 ~ 2379, 
     tree == 2353 ~ 2352, 
     tree == 2374 ~ 2376, 
     tree == 2364 ~ 2360, 
     tree == 2004 & date %in% c("2022-07-19") ~ 2005,
     tree == 2014 & date %in% c("2022-09-15") ~ 2013,
     species %in% c("ARCA") & site %in% c("Cucu Mesa") ~ 12,
     TRUE ~ as.numeric(tree)),
  date_wp = date) %>% 
  select(-tag, -plot_number, -date ) %>% 
#when we didn't measure exactly on the same day, but want predawns and middays to overlap, change the dates (need to do this for LWC too)
  mutate(date_wp = case_when(
    date_wp %in% c("2022-05-25") ~ as.Date("2022-05-23"),  
   # date_wp %in% c("2022-03-25") ~ as.Date("2022-03-23"), 
    date_wp %in% c("2022-09-16") ~ as.Date("2022-09-15"), 
   date_wp %in% c("2022-04-27") ~ as.Date("2022-04-25"), 
    TRUE ~ as.Date(date_wp))) %>% 
  distinct()

#Look at it to see what measurements are oddballs: 

#LL and MiddleEarth
plotly::ggplotly(wp_df %>% 
                   filter(plot == "MiddleEarth") %>% 
                   ggplot(aes(y = mpa, 
                     x = as.factor(tree), 
                     color = time)) +
  geom_point() + 
  facet_wrap(~as.factor(date_wp)))

#Hobbiton
plotly::ggplotly(wp_df %>% 
                   filter(plot == "Hobbiton") %>% 
                   ggplot(aes(y = mpa, 
                     x = as.factor(tree), 
                     color = time)) +
  geom_point() + 
  facet_wrap(~as.factor(date_wp)))

#LL and MiddleEarth
plotly::ggplotly(wp_df %>% 
                   filter(plot == "Rohan") %>% 
                   ggplot(aes(y = mpa, 
                     x = as.factor(tree), 
                     color = time)) +
  geom_point() + 
  facet_wrap(~as.factor(date_wp)))

plotly::ggplotly(wp_df %>% 
                   filter(plot == "Weathertop") %>% 
                   ggplot(aes(y = mpa, 
                     x = as.factor(tree), 
                     color = time)) +
  geom_point() + 
  facet_wrap(~as.factor(date_wp)))

#LL and MiddleEarth
plotly::ggplotly(wp_df %>% 
                   filter(plot == "Ridge") %>% 
                   ggplot(aes(y = mpa, 
                     x = as.factor(tree), 
                     color = time)) +
  geom_point() + 
  facet_wrap(~as.factor(date_wp)))

#LL and MiddleEarth
plotly::ggplotly(wp_df %>% 
                   filter(plot == "High") %>% 
                   ggplot(aes(y = mpa, 
                     x = as.factor(tree), 
                     color = time)) +
  geom_point() + 
  facet_wrap(~as.factor(date_wp)))

#Cucu, low
plotly::ggplotly(wp_df %>% 
                   filter(plot == "Low" & site == "Cucu Mesa") %>% 
                   ggplot(aes(y = mpa, 
                     x = as.factor(tree), 
                     color = time)) +
  geom_point() + 
  facet_wrap(~as.factor(date_wp)))

#Piney
plotly::ggplotly(wp_df %>% 
                   filter(site == "PineNeedle") %>% 
                   ggplot(aes(y = mpa, 
                     x = as.factor(tree), 
                     color = time)) +
  geom_point() + 
  facet_wrap(~as.factor(date_wp)))

#Chamise
plotly::ggplotly(wp_df %>% 
                   filter(site == "Chamise") %>% 
                   ggplot(aes(y = mpa, 
                     x = as.factor(tree), 
                     color = time)) +
  geom_point() + 
  facet_wrap(~as.factor(date_wp)))
```
```{r}
plotly::ggplotly(wp_df %>% 
                   filter(species == "live oak"
                          #week %in% c(12, 14, 15)
                          ) %>% 
                   group_by(tree, time, date_wp) %>% 
                   mutate(mpa = mean(mpa, na.rm = T)) %>% 
  mutate(tree_num = as.numeric(tree)) %>% 
  ggplot(aes(y =mpa, 
             x = as.character(tree), 
             shape = species,
             color = time)) +
  geom_point(alpha = .3) + 
  facet_wrap(~date_wp*week, nrow = 5))
```

Remove outliers from individual mpa measurements: 

######*REMOVING MPA OUTLIERS: 
```{r}
wp_df_2 <- wp_df %>% 
  
#MiddleEarth
  filter(!(tree %in% c(2354) & date_wp %in% c("2022-02-28") & time %in% c("md") & mpa %in% c(2.63, 0.962))) %>% 
  filter(!(tree %in% c(2377) & date_wp %in% c("2022-02-28") & time %in% c("md") & mpa %in% c(2.71))) %>% 
  filter(!(tree %in% c(2352, 3454) & date_wp %in% c("2022-03-08") & time %in% c("md") & mpa %in% c(3.5, 2.9))) %>% 
  filter(!(tree %in% c(3454) & date_wp %in% c("2022-03-08") & time %in% c("pd") & mpa > 2)) %>% 
  filter(!(tree %in% c(2366) & date_wp %in% c("2022-03-08") & time %in% c("md") & mpa == 3)) %>% 
  filter(!(tree %in% c(2376) & date_wp %in% c("2022-03-08") & time %in% c("md") & mpa %in% c(4.306)))%>% 
  filter(!(tree %in% c(2377) & date_wp %in% c("2022-03-08") & time %in% c("pd"))) %>% #notes say these were hard to read and they're insanely high
  filter(!(tree %in% c(2380) & date_wp %in% c("2022-03-08") & time %in% c("pd") & mpa < .6)) %>%
  filter(!(tree %in% c(2365) & date_wp %in% c("2022-03-15") & time %in% c("pd") & mpa < .5)) %>% 
  filter(!(tree %in% c(2365) & date_wp %in% c("2022-03-25") & time %in% c("pd") & mpa > 2)) %>% #notes say PDs were weird because taken eve of next day
  filter(!(tree %in% c(2377) & date_wp %in% c("2022-03-25") & time %in% c("md") & mpa >= 2.55)) %>% #notes say these were "old leaves ugly"
  filter(!(tree %in% c(2377) & date_wp %in% c("2022-04-11") & time %in% c("pd") & mpa == 0.3 )) %>% 
  filter(!(tree %in% c(2354) & date_wp %in% c("2022-04-13") & time %in% c("md") & mpa > 3.5)) %>% #super negative relative to all other measurements
  #filter(!(tree %in% c(2379) & date_wp %in% c("2022-04-13") & time %in% c("md") & mpa > 3)) %>% #mpas are split around 2 and around 3; not notes which to trust. Take average? 
  filter(!(tree %in% c(2352) & date_wp %in% c("2022-05-09") & time %in% c("pd") & mpa > 3 )) %>% #1/5 pds is >3, rest are < 2
  filter(!(tree %in% c(2376) & date_wp %in% c("2022-08-18") & time %in% c("pd") & mpa > 3)) %>% 
  filter(!(tree %in% c(2354) & date_wp %in% c("2022-09-15") & time %in% c("pd") & mpa == 2.9)) %>% 
  
#Hobbiton:
   filter(!(tree %in% c(2347) & date_wp %in% c("2022-02-28") & time %in% c("pd") & mpa > 2 )) %>% 
  filter(!(tree %in% c(2347) & date_wp %in% c("2022-03-08") & time %in% c("md") & mpa > 3 )) %>% 
   filter(!(tree %in% c(2346) & date_wp %in% c("2022-03-15") & time %in% c("pd") & mpa > 3 )) %>% 
   filter(!(tree %in% c(2347) & date_wp %in% c("2022-03-15") & time %in% c("md") & mpa > 3 )) %>% 
  mutate(time = case_when(
    tree %in% c(2345) & date_wp %in% c("2022-03-25") & mpa %in% c(1.800, 2.700, 3.100, 3.450) ~ "md", 
    tree %in% c(2345) & date_wp %in% c("2022-03-25") & mpa %in% c(1.850, 2.050, 1.300, 1.500, 2.500, 2.400, 2.4500) ~ "pd", #predawn and middays entered in wrong spot? 
    TRUE ~ as.character(time))
  ) %>% 
  filter(!(tree %in% c(2301) & date_wp %in% c("2022-04-04") & time %in% c("pd") & mpa > 3)) %>%
  filter(!(tree %in% c(2345) & date_wp %in% c("2022-08-18") & time %in% c("pd") & mpa > 3)) %>% #note says these were y0 and they're also WAY more negative than the middays. 

#Weathertop
  filter(!(tree %in% c(2381) & date_wp %in% c("2022-03-02") & time %in% c("md") & mpa > 2)) %>%
  filter(!(tree %in% c(2381) & date_wp %in% c("2022-03-02") & time %in% c("md") & mpa < .5)) %>%
  filter(!(tree %in% c(2381) & date_wp %in% c("2022-03-15") & time %in% c("md") & mpa < .5)) %>%
    filter(!(tree %in% c(2371) & date_wp %in% c("2022-03-30") & time %in% c("pd") & mpa < .5)) %>%
  filter(!(tree %in% c(2384) & date_wp %in% c("2022-04-06") & time %in% c("md") & mpa > 2.5)) %>% #notes say "injured leaf" and "? broke"
  filter(!(tree %in% c(2382) & date_wp %in% c("2022-04-11") & time %in% c("pd") & mpa > 2 )) %>% #way more neg than MDs, and we did a ton of reps likely due to such weird values
  filter(!(tree %in% c(2010) & date_wp %in% c("2022-07-09") & time %in% c("md") & mpa == 3.00)) %>%
  filter(!(tree %in% c(2383) & date_wp %in% c("2022-09-15") & time %in% c("pd") & mpa < 0.4 )) %>%

#ridge
  filter(!(tree %in% c(2331) & date_wp %in% c("2022-07-19") & time %in% c("pd") & mpa > 3)) %>%
    
#Cucu, high
  filter(!(tree %in% c(2006) & date_wp %in% c("2022-09-15") & time %in% c("pd") & mpa > 3)) %>%
  
#Cucu, low
  filter(!(tree %in% c(2015) & date_wp %in% c("2022-07-09") & time %in% c("pd") & mpa > 4 )) %>%
  filter(!(tree %in% c(2012) & date_wp %in% c("2022-05-23") & time %in% c("md") & mpa < 2.7 )) %>%
  
#Piney (In general a LOT of variation in PDs and MDs at Piney, probable because of the Y1 vs. Y0 differences, which are not well noted except in the notebooks)
  filter(!(tree %in% c(2086) & date_wp %in% c("2022-04-04") & time %in% c("md"))) %>% #not sure what this tree is, we didnt seen to collect middays at Piney on this day; just predawns. Are we missing a sheet with the Piney middays? 
  filter(!(tree %in% c(2089) & date_wp %in% c("2022-04-27") & time %in% c("md") & mpa < 2 )) %>%
  filter(!(tree %in% c(2092) & date_wp %in% c("2022-05-23") & time %in% c("md") & mpa < 2 )) %>%
  filter(!(tree %in% c(2092) & date_wp %in% c("2022-09-15") & time %in% c("pd") & mpa > 1.5)) %>%
  
#Chamise: 
  filter(!(tree %in% c(2026) & date_wp %in% c("2022-04-04") & time %in% c("md") & mpa > 3.4 )) %>%
  filter(!(tree %in% c(2029) & date_wp %in% c("2022-05-09") & time %in% c("pd") & mpa > 2)) %>%
  mutate(date_wp = ymd(date_wp)) %>% 
  mutate(week = case_when(
    date_wp %in% c("2022-03-30") ~ 14, #this was being listed as week 13, but I think it should actually be 14 based on 4/4 and 4/6 samplings having some midday trees that correspond to the predawns measured on this day. 
    TRUE ~ week(date_wp))) %>% 
  arrange(date_wp, tree, time)
```

Check that we didn't lose more MPas than planned: 
```{r}
plotly::ggplotly(wp_df_2 %>% 
                   filter(species == "live oak"
                          #week %in% c(12, 14, 15)
                          ) %>% 
                   group_by(tree, time, date_wp) %>% 
                   mutate(mpa = mean(mpa, na.rm = T)) %>% 
  mutate(tree_num = as.numeric(tree)) %>% 
  ggplot(aes(y =mpa, 
             x = as.character(tree), 
             shape = species,
             color = time)) +
  geom_point(alpha = .3) + 
  ggtitle("After removing MPa outliers") +
  facet_wrap(~date_wp*week, nrow = 5))
```

```{r}
plotly::ggplotly(wp_df_2 %>% 
                   filter(plot == "MiddleEarth") %>% 
                   ggplot(aes(y = mpa, 
                     x = as.factor(tree), 
                     color = time)) +
  geom_point() + 
  facet_wrap(~as.factor(date_wp)))

#Hobbiton
plotly::ggplotly(wp_df_2 %>% 
                   filter(plot == "Hobbiton") %>% 
                   ggplot(aes(y = mpa, 
                     x = as.factor(tree), 
                     color = time)) +
  geom_point() + 
  facet_wrap(~as.factor(date_wp)))

#Rohan
plotly::ggplotly(wp_df_2 %>% 
                   filter(plot == "Rohan") %>% 
                   ggplot(aes(y = mpa, 
                     x = as.factor(tree), 
                     color = time)) +
  geom_point() + 
  facet_wrap(~as.factor(date_wp)))

#Weathertop
plotly::ggplotly(wp_df_2 %>% 
                   filter(plot == "Weathertop") %>% 
                   ggplot(aes(y = mpa, 
                     x = as.factor(tree), 
                     color = time)) +
  geom_point() + 
  facet_wrap(~as.factor(date_wp)))

#LL and MiddleEarth
plotly::ggplotly(wp_df_2 %>% 
                   filter(plot == "Ridge") %>% 
                   ggplot(aes(y = mpa, 
                     x = as.factor(tree), 
                     color = time)) +
  geom_point() + 
  facet_wrap(~as.factor(date_wp)))

#CUCU
plotly::ggplotly(wp_df_2 %>% 
                   filter(plot == "High") %>% 
                   ggplot(aes(y = mpa, 
                     x = as.factor(tree), 
                     color = time)) +
  geom_point() + 
  facet_wrap(~as.factor(date_wp)))


#Cucu, low
plotly::ggplotly(wp_df_2 %>% 
                   filter(plot == "Low" & site == "Cucu Mesa") %>% 
                   ggplot(aes(y = mpa, 
                     x = as.factor(tree), 
                     color = time)) +
  geom_point() + 
  facet_wrap(~as.factor(date_wp)))

#Piney
plotly::ggplotly(wp_df_2 %>% 
                   filter(site == "PineNeedle") %>% 
                   ggplot(aes(y = mpa, 
                     x = as.factor(tree), 
                     color = time)) +
  geom_point() + 
  facet_wrap(~as.factor(date_wp)))

#Chamise
plotly::ggplotly(wp_df_2 %>% 
                   filter(site == "Chamise") %>% 
                   ggplot(aes(y = mpa, 
                     x = as.factor(tree), 
                     color = time)) +
  geom_point() + 
  facet_wrap(~as.factor(date_wp)))

#All
plotly::ggplotly(wp_df_2 %>% 
                  # filter(site == "Chamise") %>% 
                   ggplot(aes(y = mpa, 
                     x = as.factor(tree), 
                     color = time)) +
  geom_point() + 
  facet_wrap(~as.factor(date_wp)*week))


#Pivot wider to view: 
wp_df2_wider <- wp_df_2 %>% 
  select(site, week, time, mpa, tree, date_wp) %>% 
                   pivot_wider(names_from = "time", 
                               values_from = "mpa", 
                               values_fn = function(x) mean(x, na.rm = TRUE)) %>% 
  filter(tree > 100)

# plotly::ggplotly(wp_df2_wider %>%  
#                    ggplot(aes(y = md, 
#                      x = pd, 
#                      color = site, 
#                      label = tree
#                      )) +
#   geom_point() + 
#     geom_abline() +
#   facet_wrap(~as.factor(week)))
#all PDs are less negative than MDs, thats good!!

#write out to send to Lee: 
write_csv(wp_df_2, here::here("processed-data", "wp_cleanedup_20240516.csv"))
#write out to send to Lee: 
write_csv(wp_df2_wider, here::here("processed-data", "wp_wide_cleanedup_20240516.csv"))
```

####WC + WP, LWA


```{r}
#Merge water potential data with the dataset we made above that has the rescanned LMA and LWA for LWC leaves; rep is one per tree/week time to MPa will also need to be summarized similarly.
wp_mean_df <- wp_df_2 %>% 
  group_by(tree, week, time, date_wp) %>% 
  mutate(mean_mpa = mean(mpa, na.rm = T), 
         sd_mpa = sd(mpa, na.rm = T)) %>% 
  select(-mpa) %>% 
  distinct() %>% 
  mutate(date = date_wp) %>% 
  ungroup() #%>% 
 # select(-week)

plotly::ggplotly(wp_mean_df %>% 
                   filter(species == "live oak"
                          #week %in% c(12, 14, 15)
                          ) %>% 
                   group_by(tree, time, date_wp) %>% 
                  # mutate(mpa = mean(mpa, na.rm = T)) %>% 
  mutate(tree_num = as.numeric(tree)) %>% 
  ggplot(aes(y =mean_mpa, 
             x = as.character(tree), 
             shape = species,
             color = time)) +
  geom_point(alpha = .3) + 
  ggtitle("Original, Just MPAs") +
  facet_wrap(~date_wp*week, nrow = 5))
```


```{r}
wp_wc_df2 <- merge(wp_mean_df, 
                   lwa_df %>% 
                     mutate(date = case_when(
                       date %in% c("2022-09-12") ~ as.Date("2022-09-15"),
                       TRUE ~ as.Date(date)
                     )), 
                  by = c("tree","time", "species", "site", "plot", "date"), 
                  all = T) %>% 
  mutate(week = lubridate::week(date)) %>% 
  select(-week.x, -week.y) %>% 
  # mutate(week = case_when(
  #   date %in% c("2022-03-30") ~ 14, #this was being listed as week 13, but I think it should actually be 14 based on 4/4 and 4/6 samplings having some midday trees that correspond to the predawns measured on this day. 
  #   TRUE ~ week(date))) %>% 
  #mutate(week = week(date)) %>% 
  filter(species != "ARCA", 
         species != "LEU") %>% 
  mutate(sum_water_wt_g = case_when(
    sum_water_wt_g %in% c(0) ~ NA, 
    TRUE ~ as.numeric(sum_water_wt_g)
  )) %>% 
  mutate(sum_area_cm2 = case_when(
    sum_area_cm2 %in% c(0) ~ NA, 
    TRUE ~ as.numeric(sum_area_cm2)
  ))%>% 
  mutate(lwa_g_cm2_new = case_when(
    lwa_g_cm2_new %in% c(0) ~ NA, 
    TRUE ~ as.numeric(lwa_g_cm2_new)
  )) %>% 
  mutate(lma_g_cm2_new = case_when(
    lma_g_cm2_new %in% c(0) ~ NA, 
    TRUE ~ as.numeric(lma_g_cm2_new)
  )) %>% 
  distinct() %>% 
#deal with LWA outliers:
  mutate(lwa_g_cm2_new = case_when(
    week %in% c(19) & tree %in% c(2384) & time %in% c("pd") ~ NA,
    week %in% c(19) & tree %in% c(2089) ~ NA,
     week %in% c(14) & tree %in% c(2013, 2014, 2008, 2370) & time %in% c("md") ~ NA,
    TRUE ~ as.numeric(lwa_g_cm2_new)
  )) %>% 
#LWA outliers are likely the same as the LWC outliers: 
     mutate(lwc_bulk = case_when(
    week %in% c(19) & tree %in% c(2384) & time %in% c("pd") ~ NA,
    week %in% c(19) & tree %in% c(2089) ~ NA,
     week %in% c(14) & tree %in% c(2013, 2014, 2008, 2370) & time %in% c("md") ~ NA,
    TRUE ~ as.numeric(lwc_bulk))) %>% 
  #remove specific points that have issues: (if no comment, should track down later to see what is wrong!)
 # filter(!(week %in% c(19) & tree %in% c(2089))) %>% #two bulk dry weights entered (0.5982 and 1.3073)
#  filter(!(week %in% c(37) & tree %in% c(2086) & time %in% c("pd"))) %>% 
 # filter(!(week %in% c(14) & tree %in% c(2013, 2014, 2008, 2370) & time %in% c("md"))) %>% #LWC is weird for these, might have been mislabled? 2370 just had one mpa rep
#  filter(!(week %in% c(21) & tree %in% c(2343) & time %in% c("pd"))) %>% 
 # filter(!(week %in% c(17) & tree %in% c(2089, 2093, 2091) & time %in% c("md"))) %>%
##  filter(!(week %in% c(19) & tree %in% c(2022, 2030, 2023) & time %in% c("pd"))) %>%
 # filter(!(week %in% c(19) & tree %in% c(2384) & time %in% c("pd"))) %>% #lwa is <0, must be wrong
  select(-rep) %>% 
  distinct() %>%
  filter(species != "ARCA", 
         species != "LEU") %>% 
  distinct() %>% 
#remove weird values before we combine leaves and bulk so that the combined represents the non-weird values
  mutate(lwc_bulk = case_when( 
    lwc_bulk > 2.5 ~ NA_real_, 
    lwc_bulk < 0 ~ NA_real_, 
    TRUE ~ as.numeric(lwc_bulk)
  )) %>% 
   # mutate(lwc_leaf = case_when(
   #   lwc_leaf >= 5 ~ NA_real_, #weird values
   #  lwc_leaf < 0 ~ NA_real_,
   #  TRUE ~ as.numeric(lwc_leaf))) %>% 
#combine reps to get mean mpa and lwc per tree; combine bulk and separate lwc: 
 group_by(tree, week, time, date_wc) %>%
 mutate(#mpa_mean = mean(mpa, na.rm = T),
        lwc_leaf = (wm_leaf_g - dm_leaf_g)/dm_leaf_g, 
        #sum_area_cm2 = sum(area_cm2), #sum leaf area across all reps
        lwc_leaf_mean = mean(lwc_leaf, na.rm=T),
        lwc_bulk_mean = mean(lwc_bulk, na.rm=T),
        lwc_bulk_leaf_diff = lwc_leaf_mean-lwc_bulk_mean) %>% #see if there are any huge differences between leaves and bulk
 # ungroup() %>% 
 # dplyr::select(-rep) %>%
 # distinct() %>% 
  group_by(tree, plot, site, species, time, week, date_wc) %>% 
#when there are bigish differences, replace the combined bulk+leaf with just the bulk; otherwise take the mean of both. 
  mutate(lwc_mean = case_when(
    lwc_bulk_leaf_diff < -.3 ~ lwc_bulk_mean, 
    lwc_bulk_leaf_diff > .32 ~ lwc_leaf_mean,
   # lwc_mean = NA ~ lfm,
    TRUE ~ mean(c_across(c('lwc_leaf_mean', 'lwc_bulk_mean')), na.rm=TRUE)
    )) %>% 
  select(-lwc_leaf_mean, -lwc_bulk_mean, -lwc_leaf, -lwc_bulk, -dm_leaf_g, -wm_leaf_g,-lwc_bulk_leaf_diff) %>% 
  arrange(tree, week, time)  %>% 
group_by(tree, week, time, date) %>%
  fill(c("dm_bulk_g", "wm_bulk_g"), .direction = "downup") %>% 
      distinct() %>% 
#since we didn't collect RWC every week, create a column we can use to link RWC to all of the closest LWC weeks (should probably linearly interpolate at some point)
  mutate(week_rwc_link = case_when(
    week %in% c(9, 10, 11) ~ 11, 
    week %in% c(12, 13) ~ 13, 
    week %in% c(14, 15, 16) ~ 15, 
    week %in% c(17, 18) ~ 17, 
    week %in% c(19) ~ 19, 
    week %in% c(20, 21) ~ 21, 
    week %in% c(22) ~ 21, 
    week %in% c(29, 30) ~ 30, 
    week %in% c(33) ~ 37, 
    TRUE ~ as.numeric(week)
  ),
#since we didn't collect alas every week, create a column we can use to link LMA to all of the closest LWC weeks (should probably linearly interpolate at some point):
  week_alas_link = case_when(
    week %in% c(9, 10) ~ 10, 
    week %in% c(11) ~ 11, 
    week %in% c(12, 13) ~ 13, 
    week %in% c(14, 15) ~ 14, 
    week %in% c(16, 17, 18) ~ 17, 
    week %in% c(19, 20, 21) ~ 21, 
    week %in% c(29, 30) ~ 29, 
    week %in% c(33) ~ 37, 
    week %in% c(33) ~ 40, 
    TRUE ~ as.numeric(week)
  )) %>% 
  mutate(timing = case_when( #define early and late seasons as before and after week 17 (April end week)
    week %in% c(9, 10, 11, 12, 13, 14, 15, 16) ~ "early", 
    week %in% c(17, 18, 19, 20, 21, 29, 33, 37, 38, 40) ~ "late"
  )) %>% 
 ## filter(!if_all(c(mpa_mean,lwc_mean), is.na)) %>% 
 # select(-lwc_leaf, -lwc_bulk, - lwc_bulk_leaf_diff) %>% 
  distinct() %>% 
  mutate(type = "leaf") %>% #call these all leaves, becuase they are not stems
  # mutate(date_wp = case_when(
  #   date_wp %in% c("2022-03-25") ~ as.Date("2022-03-23"), 
  #   TRUE ~ as.Date(date_wp)
  # )) %>% 
  # group_by(tree, time, week) %>% 
  # fill(c(dm_bulk_g, wm_bulk_g), .direction = c("down")) %>% 
    distinct() %>% 
  arrange(tree, week, time) %>% 
  mutate(lma_g_cm2_new = case_when( #make any values that were outside of the Al:As LMA to NA (likely a scanning issue/lost leaves)
    lma_g_cm2_new > 0.02 ~ NA_real_, 
     lma_g_cm2_new < 0.009 ~ NA_real_, 
    TRUE~ as.numeric(lma_g_cm2_new)
  )) %>% 
  #filter(date %in% c("2022-05-23")) %>% 
 # filter(week %in% c(12:14))%>% 
  select(tree, week, time, date, order(colnames(.))) %>% 
  mutate(tree = case_when(
    tree %in% c(2014) & week %in% c(37) ~ 2013, #mislabled?
    TRUE ~ as.numeric(tree)
  )) %>% 
    arrange(tree, week, time) %>% 
  ungroup()
```

```{r}
plotly::ggplotly(wp_wc_df2 %>% 
                   filter(species == "live oak"
                          #week %in% c(12, 14, 15)
                          ) %>% 
                   group_by(tree, time, date_wp) %>% 
                  # mutate(mpa = mean(mpa, na.rm = T)) %>% 
  mutate(tree_num = as.numeric(tree)) %>% 
  ggplot(aes(y =mean_mpa, 
             x = as.character(tree), 
             shape = species,
             color = time)) +
  geom_point(alpha = .3) + 
    ggtitle("After adding LWC and LWA") +
  facet_wrap(~date_wp*week, nrow = 5))
```


```{r}
#there are some issues with week 37, where we dont have scans from trees measured 9/15 but we do for trees measured 
wp_wc_wk37_probs <- wp_wc_df2 %>% 
  filter(week == 37) %>% 
  group_by(tree, time) %>% 
  filter(n() > 1) %>% 
  group_by(tree, week, time) %>% 
  fill(c("mean_mpa", "date_wp"), .direction = "downup")

wp_wc_wk37_noprobs <- wp_wc_df2 %>% 
  filter(week == 37) %>% 
  group_by(tree, time) %>% 
  filter(n() == 1) %>% 
  group_by(tree, week, time) %>% 
  fill(c("mean_mpa", "date_wp"), .direction = "downup")

wp_wc_wk37 <- bind_rows(wp_wc_wk37_noprobs, wp_wc_wk37_probs)

wp_wc_no_wk37 <- wp_wc_df2 %>% 
  filter(!(week == 37))

wp_wc_probs <- wp_wc_df2 %>% 
  group_by(tree, week, time) %>% 
  filter(n() > 2)

wp_wc_df3 <- bind_rows(wp_wc_wk37, wp_wc_no_wk37) %>% 
  mutate(date_wp = case_when(
    week %in% c(14) & time %in% c("md") & date_wp %in% c("<NA>", NA) ~ as.Date("2022-04-06"),
    week %in% c(37) & time %in% c("md") & date_wp %in% c("<NA>", NA) ~ as.Date("2022-09-15"),
    TRUE ~ as.Date(date_wp)
   ))

#Visualize: some measurements of LWA don't make sense
plotly::ggplotly(wp_wc_df3 %>% 
  ggplot(aes(y = mean_mpa, 
             x = lwa_g_cm2_new, 
             color = species, 
             shape = as.factor(week),
             label = tree)) +
  geom_point())

#Visualize: some measurements of LWA don't make sense
plotly::ggplotly(wp_wc_df3 %>% 
  ggplot(aes(y = mean_mpa, 
             x = lwc_mean, 
             color = species, 
             shape = as.factor(week),
             label = tree)) +
  geom_point())

#Visualize: some measurements of LMA don't make sense; these are the same that don't make sense for LWA, so probably a bigger issue
plotly::ggplotly(wp_wc_df3  %>% 
  ggplot(aes(y = mean_mpa, 
             x = lma_g_cm2_new, 
             color = species,
             shape = as.factor(week),
             label = tree)) +
  geom_point())


plotly::ggplotly(wp_wc_df3  %>% 
                   filter(week == 17 & tree %in% c(2009:2030)) %>% 
  ggplot(aes(y = mean_mpa, 
             x = lwa_g_cm2_new, 
             color = species, 
             shape = time,
             label = tree),
             tooltip = c("tree", "species", "time")) +
  geom_point() +
    geom_line(aes(group = tree)) +
    facet_wrap(~week*date, scales = "free")
  )

plotly::ggplotly(wp_wc_df3  %>% 
                 #week == 17) %>% 
  ggplot(aes(y = mean_mpa, 
             x = lwc_mean, 
             color = species, 
             shape = time,
             label = tree),
             tooltip = c("tree", "species", "time")) +
  geom_point() +
    facet_wrap(~week*date, scales = "free")
  )

write.csv(wp_wc_df3, here::here("processed-data", "wp_wc_df.csv"))

wp_wc_df3 %>% 
  group_by(species, week, date_wp, time) %>% 
  summarise(n = n()) %>% 
  pivot_wider(
    names_from = c(species),
    values_from = c(n)
  )

plotly::ggplotly(wp_wc_df3 %>% 
                   filter(species == "live oak"
                          #week %in% c(12, 14, 15)
                          ) %>% 
                   group_by(tree, time, date_wp) %>% 
                  # mutate(mpa = mean(mpa, na.rm = T)) %>% 
  mutate(tree_num = as.numeric(tree)) %>% 
  ggplot(aes(y =mean_mpa, 
             x = as.character(tree), 
             shape = species,
             color = time)) +
  geom_point(alpha = .3) + 
    ggtitle("After ") +
  facet_wrap(~date_wp*week, nrow = 5))
```

#####RWC:


Merge those. MPa and LWC are repeated because they align with SWC leaves and stems of different years. 

For moving on in combining, we'll just use Year 0 Leaves, should give us a df that is still one row per LWC and MPa per date/time combo 

```{r}
rwc_df_raw <- read.csv(here("processed-data", paste0("rwc_alldates_",datver,".csv"))) 


rwc_df0 <- rwc_df_raw %>% 
  mutate(tree = case_when(
    tree %in% c(2364, 2361) ~ 2360, 
    tree %in% c(2083) ~ 2383, 
    tree %in% c(2081) ~ 2381, 
    tree %in% c(2078) ~ 2378, 
    TRUE ~ as.numeric(tree)
  )) %>% 
  mutate(date = case_when(
   # date %in% c("2022-02-28", "2022-03-03") ~ as.Date("2022-02-28"), 
    date %in% c("2022-04-11", "2022-04-12", "2022-04-14") ~ as.Date("2022-04-12"), 
    date %in% c("2022-05-03", "2022-05-04", "2022-05-09") ~ as.Date("2022-05-03"),
    date %in% c("2022-05-24", "2022-05-25", "2022-05-27") ~ as.Date("2022-05-25"),
    date %in% c("2022-07-27", "2022-07-28") ~ as.Date("2022-07-27"),
    date %in% c("2022-09-12", "2022-09-15") ~ as.Date("2022-09-15"),
    TRUE ~ as.Date(date)
  )) %>% 
  mutate(year_rwc = case_when(
    date %in% c("2022-04-11") ~ 1, #assume we didn't know 
    year %in% c(NA_real_) ~ 0, #assume no year listed is year 0
    year %in% c(3) ~ 1, #assume year 3 is just year 1
    TRUE ~ as.numeric(year)
  )) %>% 
  mutate(week_rwc_link = case_when(
    week %in% c(22) ~ 21, #no mpas for week 22 (prob becuase of week split in week function), so make that 21
    TRUE ~ as.numeric(week))) %>% 
  group_by(tree, date, type, year_rwc) %>% 
#remove absurd values -- issue with wet leaves having drops on them? 
  mutate(swc_per_dry_g = case_when(
    swc_per_dry_g > 5 ~ NA_real_, 
    TRUE ~ as.numeric(swc_per_dry_g)
  )) %>% 
  mutate(mean_swc_per_dry_g = mean(swc_per_dry_g, na.rm = T), 
         mean_swc_g = mean(swc_g, na.rm = T), 
         swc_ww_g = swc_ww_g, 
         swc_dw_g = swc_dw_g) %>% 
  ungroup() %>% 
  mutate(date_rwc = date, 
         week = as.numeric(week), 
         type_rwc = type) %>% 
  select(-date) %>% 
  distinct() %>% 
  select(mean_swc_per_dry_g, mean_swc_g, tree, date_rwc, year_rwc , type_rwc, week_rwc_link,
         ) %>% 
  as.data.frame() 
  
#MERGE
rwc_df <- merge(wp_wc_df3,rwc_df0, by = c("tree", "week_rwc_link"),
                    all = T
                    ) %>% 
  mutate(type_rwc = case_when(
    type_rwc %in% c("leeaf") ~ "leaf", 
    type_rwc %in% c("setm") ~ "stem", 
    TRUE ~ as.character(type_rwc)
  )) %>% 
  distinct() %>% 
  mutate(week_wp = week) %>% 
  group_by(tree) %>% 
  fill(c("species"), .direction = "downup") %>% 
  drop_na(species) %>% 
 # filter(!(mean_swc_per_dry_g > 2 & species %in% c("live oak"))) %>% #Doing this removes a whole date, not sure what was up there!
  mutate(year_rwc = as.factor(year_rwc))  %>% 
  mutate(date_rwc = case_when(
    date_rwc %in% c("2022-05-23") ~ as.Date("2022-05-25"), 
    TRUE ~ as.Date(date_rwc)
  )) %>% 
  mutate(year_rwc = case_when(
    date_rwc %in% c("2022-03-15") & species %in% c("live oak") ~ "1", #assume we didn't know 
    year_rwc %in% c(NA_real_) ~ "0", #assume no year listed is year 0
    year_rwc %in% c("3") ~ "1", #assume year 3 is just year 1
    species %in% c("blue oak") ~ "0",
    TRUE ~ as.factor(year_rwc)
  )) 

write_csv(rwc_df, here::here("processed-data", "rwc_2022.csv"))

wp_wc_rwc_df_all <- rwc_df


plotly::ggplotly(wp_wc_rwc_df_all %>% 
                   filter(species == "live oak"
                          #week %in% c(12, 14, 15)
                          ) %>% 
                   group_by(tree, time, date_wp) %>% 
                  # mutate(mpa = mean(mpa, na.rm = T)) %>% 
  mutate(tree_num = as.numeric(tree)) %>% 
  ggplot(aes(y =mean_mpa, 
             x = as.character(tree), 
             shape = species,
             color = time)) +
  geom_point(alpha = .3) + 
    ggtitle("After adding RWC") +
  facet_wrap(~date_wp*week, nrow = 5))
```

###AL:As & inferred LWA: 

lwc = g water/g dry
lma = g dry/area cm2

```{r}
alas_df_raw <- read_csv(here::here("processed-data", paste0("alas_leaf_area_df_week",datver,".csv")), show_col_types = F) %>% 
  rename(tree = tree_id) %>% 
  mutate(tree = case_when( #weird trees, misentered
    tree %in% c(2048) ~ 2348,
     tree %in% c(2049) ~ 2349,
     tree %in% c(2050) ~ 2354,
     tree %in% c(2051) ~ 2351,
     tree %in% c(2052) ~ 2352,
     tree %in% c(2078) ~ 2378,
     tree %in% c(2080) ~ 2380,
     tree %in% c(2081) ~ 2381,
     tree %in% c(2082) ~ 2382,
    tree %in% c(2083) ~ 2383,
    tree %in% c(2393) ~ 2093,
    tree %in% c(2350) ~ 2354, # I remember doing this a lot :( 
    tree %in% c(2303) ~ 2330,
    tree %in% c(2434) ~ 2343,
    TRUE ~ as.numeric(tree)
  )) %>% 
  mutate(alas_measured = "yes")

leaf_number <- read_csv(here(here("Data_20230316","alas_leaf_number.csv")), show_col_types = FALSE) %>% 
  clean_names() %>% 
  mutate(date = mdy(date),
         week = week(date),
         tree = tree_id) %>% 
  select(week, tree, branch, sub_branch, year, leaf_count) %>% 
  mutate(tree = case_when(
    tree %in% c(2048) ~ 2348,
     tree %in% c(2049) ~ 2349,
     tree %in% c(2050) ~ 2354,
     tree %in% c(2051) ~ 2351,
     tree %in% c(2052) ~ 2352,
     tree %in% c(2078) ~ 2378,
     tree %in% c(2080) ~ 2380,
     tree %in% c(2081) ~ 2381,
     tree %in% c(2082) ~ 2382,
     tree %in% c(2083) ~ 2383,
    tree %in% c(2393) ~ 2093,
    tree %in% c(2350) ~ 2354,
    tree %in% c(2303) ~ 2330,
    tree %in% c(2434) ~ 2343,
    TRUE ~ as.numeric(tree)
  ))

alas_df <- merge(alas_df_raw, leaf_number, by = c("week", "tree", "branch", "year", "sub_branch")) %>% 
  distinct() %>% 
  select(-'...1')

#Are there weird duplicates? 
duplicates <- alas_df  |> #no duplicates it seems
  group_by_all() |>
  filter(n() > 1) |>
  ungroup()
duplicates

#Now, take the mean value for each tree, week, keeping years separate too. This dataset is best for joining with wc and mpa, as those also have single row per week, tree, year (because of bulk lwc metric)
alas_df_means <- alas_df %>% 
  group_by(week, tree, year) %>% 
  mutate(alas_cm2_per_mm2 = mean(alas_cm2_per_mm2, na.rm = T), 
         mean_leaf_count = mean(leaf_count, na.rm = T), 
         lma_g_cm2 = mean(lma_g_cm2, na.rm = T), 
         sla_cm_g = mean(sla_cm_g, na.rm = T), 
         ldm_g = mean(ldm_g, na.rm = T), 
         leaf_count = mean(leaf_count, na.rm = T)) %>% 
  ungroup() %>% 
  select(tree, week, species, alas_cm2_per_mm2, lma_g_cm2, sla_cm_g, ldm_g, leaf_count, year, alas_measured) %>% 
  distinct() %>% 
  mutate(week_alas_link = week, 
         type = "leaf") %>% 
  select(-week, -species) 


#Next, Merge with the OG df: merge alas_df_means with the df with wp, wc, and rwc-related values (NOTE: After this merge, SWC, LMA, and Al:As - and other related metrics -  will be duplicated since they need to be in rows for both predawn and midday to calculate inferred LWA)

wp_wc_rwc_alas_df1 <- merge(wp_wc_rwc_df_all, alas_df_means, by = c("week_alas_link", "tree", "type"), 
  all = T) %>% 
  group_by(week_alas_link, tree, year) %>% 
  mutate(lwa_g_cm2 = lwc_mean * lma_g_cm2,
         ldmc = ldm_g/(ldm_g + mean_swc_per_dry_g))  %>% 
  select(-timing, -plot, -site) %>% 
  distinct() %>% 
  filter() %>% 
  mutate(week = case_when(
    week == NA ~ week_alas_link, 
    TRUE ~ as.numeric(week)
  )) %>% 
  mutate(type = case_when(
    type == NA ~ type_rwc, 
    TRUE ~ as.character(type)
  )) %>% 
  arrange(tree, week, time)%>% 
  mutate(lwa_g_cm2_new = case_when( #Make weird looking new_lwa values NA, and rename the dataset since it's getting annoyingly long. 
    lwa_g_cm2_new > 0.05 ~ NA_real_, 
    lwa_g_cm2_new  > .3 ~ NA_real_,
    TRUE ~ as.numeric(lwa_g_cm2_new)
  )) %>% 
  group_by(tree) %>% 
  fill(c("species"), .direction = "downup")

#look at how new LWA and old LWA compare: 
wp_wc_rwc_alas_df1 %>% #week 19 is still weird!!! 
  filter(lwa_g_cm2_new < .3) %>% 
ggplot(aes(y = lwa_g_cm2_new, 
           x = lwa_g_cm2,
           color = as.factor(week))) +
  geom_point()+
  facet_wrap(~species) +
  geom_abline()

#look at how new LWA and old LWA compare: 
wp_wc_rwc_alas_df1 %>% #week 19 is still weird!!! 
  #filter(lwa_g_cm2_new < .3) %>% 
ggplot(aes(y = lwc_new, 
           x = lwc_mean,
           color = as.factor(species))) +
  geom_point()+
  facet_wrap(~week) +
  geom_abline()
#decent!
```
```{r}
plotly::ggplotly(wp_wc_rwc_alas_df1 %>% 
                   filter(species == "live oak"
                          #week %in% c(12, 14, 15)
                          ) %>% 
                   group_by(tree, time, date_wp) %>% 
                  # mutate(mpa = mean(mpa, na.rm = T)) %>% 
  mutate(tree_num = as.numeric(tree)) %>% 
  ggplot(aes(y =mean_mpa, 
             x = as.character(tree), 
             shape = species,
             color = time)) +
  geom_point(alpha = .3) + 
    ggtitle("After adding Al:As") +
  facet_wrap(~date_wp*week, nrow = 5))
```

Lil looksie of new vs. old inferred LWA: 
```{r}
#look at it again: 
wp_wc_rwc_alas_df1 %>% 
ggplot(aes(y = lwa_g_cm2_new, 
           x = lwa_g_cm2,
           color = species
           #color = as.factor(week)
           )) +
  geom_point()+
  facet_wrap(~week) +
  geom_abline()

wp_wc_rwc_alas_df1%>% 
ggplot(aes(y = lma_g_cm2_new, 
           x = lma_g_cm2,
           color = species
           #color = as.factor(week)
           )) +
  geom_point()+
  facet_wrap(~week) +
  geom_abline()

#look at it again (compare distributions of inferred LWA and LWA from new scans)
wp_wc_rwc_alas_df1 %>% 
  ggplot() +
  geom_histogram(aes(x = lwa_g_cm2_new)) +
  labs(title = "LWA from scans")

wp_wc_rwc_alas_df1 %>% 
  ggplot() +
  geom_histogram(aes(x = lwa_g_cm2)) +
  labs(title = "Inferred LWA from LMA + LWC")

#Make a new dataset that we can use to just look at LWA types: 
lwa_data <- wp_wc_rwc_alas_df1 %>% 
  select(tree, week, time,lwa_g_cm2, lwa_g_cm2_new) %>% 
  distinct() %>% 
  group_by(time, tree, week) %>% 
  mutate(mean_lwa_old = mean(lwa_g_cm2, na.rm = T)) %>% 
  select(-lwa_g_cm2) %>% 
  distinct() %>% 
  mutate(mismatch = mean_lwa_old - lwa_g_cm2_new)

mean(lwa_data$lwa_g_cm2_new, na.rm = T)
sd(lwa_data$lwa_g_cm2_new, na.rm = T)


lwa_mismatched <- lwa_data %>% 
  filter(mismatch > 0.005 | mismatch < -0.005)

#Somewhat similar!!! yay!
```

Combine full dataset with tree information: 

```{r}
wp_wc_rwc_alas_df <- merge(trees_sites_df, wp_wc_rwc_alas_df1, all = T) %>% 
  dplyr::group_by(tree, time, week) %>%
  fill(c(date_wp, mean_mpa, lwc_mean), .direction = "downup") %>%
  dplyr::ungroup() %>% 
  group_by(tree) %>% 
  fill(c(species, site, plot), .direction = "downup") %>% 
  ungroup() %>% 
  mutate(
    rwc = (lwc_mean/mean_swc_per_dry_g)*100,
    week = case_when(
      week %in% c(13) ~ 14, #combine dates that are very close together (week() just split at the wrong spot)
      week %in% c(11) ~ 10, #or where we seem to have collected some predawns one day and some middays another day
    #  week %in% c(18) ~ 19, #4 days apart, 
      TRUE ~ week
  )) %>%
  filter(!(week == 18)) %>% #remove week 18 because mostly did middays and repeated most of those trees 4 days later on the 19
  mutate(tree = as.character(tree))  %>% 
  dplyr::group_by(tree, week, time, type_rwc) %>% #combine years to get a year mean
  mutate(water_potential = mean(mean_mpa, na.rm = T), 
         ldmc_mean = mean(ldmc, na.rm = T),
         lma_g_cm2_mean= mean(lma_g_cm2, na.rm = T),
         mpa_mean = mean(mean_mpa, na.rm = T), 
         tree_factor = as.factor(tree), 
         lwc_mean = mean(lwc_mean, na.rm = T),
         alas_cm2_per_mm2_mean = mean(alas_cm2_per_mm2, na.rm = T), 
         lwa_g_cm2_mean = mean(lwa_g_cm2, na.rm = T), 
         lwa_g_cm2_new_mean = mean(lwa_g_cm2_new, na.rm = T), 
         rwc_mean = mean(rwc, na.rm = T), 
         swc_per_dry_g_mean = mean(mean_swc_per_dry_g, na.rm = T),
         lwc_new_mean = mean(lwc_new, na.rm = T)
         ) %>% 
  rename(year_alas = year) %>% 
  ungroup() %>% 
  dplyr::distinct() %>% 
  #filter(!week == 9)%>% 
  group_by(tree, week) %>% 
  fill(alas_cm2_per_mm2, date_wp, lma_g_cm2_mean, ldmc_mean, leaf_count,.direction = "downup") %>% 
  ungroup() %>% 
    group_by(tree, week, time, date_wp) %>% 
  fill(lma_g_cm2_new, lwa_g_cm2_new_mean ,.direction = "downup") %>% 
  ungroup() %>% 
  #filter(year_alas == 0) %>% 
  distinct() %>% 
  drop_na(time) %>% 
  arrange(tree, week, date_wc, time) %>% 
  #mutate(swc_per_dry_g_mean = mean_swc_per_dry_g) %>% 
    mutate(mpa_mean = mean_mpa) %>% 
  # select( -week_rwc_link, -alas_measured, -mpa_mean, -tree_factor, -dm_bulk_g, -wm_bulk_g, -sum_water_wt_g, -sum_dm_g, -bulk_or_individual_leaf, -sum_area_cm2, -type_rwc, -year_rwc, -week_wp, -mean_swc_g, -ldmc, -rwc, -lwa_g_cm2, -ldm_g, -year_alas, -mean_swc_per_dry_g, - mean_mpa, -lma_g_cm2, -alas_cm2_per_mm2, -lwa_g_cm2_new, -date_wc, -lwc_new) %>% 
select(tree, week, date_wp, time, site, plot,  order(colnames(.))) %>% 
  distinct() 
```


####Dataset with everything, both means and actual values: 
```{r}
write.csv(wp_wc_rwc_alas_df, here("processed-data", paste0("wp_wc_rwc_",datver,".csv"))) 
```



####Dataset for analysis of LWC, MPA, LWA for manuscript:

######*REMOVING GENERAL OUTLIERS:

This has only leaves, with one row per week/tree/time/year; it will have duplicates of Al:As related measurements, but most importantly LWA, LWA, and MPa should have unique rows. 

```{r}
analysis_bothspp_df <- wp_wc_rwc_alas_df %>% 
  filter(#species == "blue oak", 
         type == "leaf",
         #lwa_g_cm2 < 0.025, #some weirdly high values?
         #!(week == 29 & lwc_mean > 1) #one weird value in week 29)
         )  %>% 
  select(week, time, tree, plot, site, species, 
         water_potential,  
         lwc_mean, 
         #lwc_g_cm2,
         lma_g_cm2_mean, 
         #lma_g_cm2_new,
         lwa_g_cm2_new_mean,
         ldmc_mean, 
         alas_cm2_per_mm2_mean, 
         lwa_g_cm2_mean, 
         #lwa_g_cm2_new,
         #leaf_count, 
         swc_per_dry_g_mean,
         rwc_mean,
        # swc,
         date_wp, 
         week_alas_link) %>% 
  distinct() %>% 
  rename(lwa_g_cm2 = lwa_g_cm2_mean, 
         lma_g_cm2 = lma_g_cm2_mean,
         lwa_g_cm2_new = lwa_g_cm2_new_mean,
         alas_cm2_per_mm2 = alas_cm2_per_mm2_mean,
         ldmc = ldmc_mean, 
         rwc = rwc_mean) %>% 
  group_by(tree, time, week) %>% 
  # mutate(alas_cm2_per_mm2 = case_when(
  #   alas_cm2_per_mm2 > 100 ~ NA_real_, #remove alas outliers
  #   TRUE ~ as.numeric(alas_cm2_per_mm2)
  # )) %>% 
   mutate(tree = as.character(tree),
          #tree_week = paste(tree, week, sep = "_"),
          time_season = case_when(
    week <= 17 ~ "before leafout", 
    week > 17 ~ "after leafout", 
    TRUE ~ as.character("none")
  )) %>%  
  filter(!(week == 29 & lwc_mean < .451), #weird values, remove for now?
         !(week == 15 & lwc_mean < .7), 
         !(time %in% c("pd") & lwc_mean < .5), 
         #!(lwa_g_cm2_new < .01 & time %in% c("pd")),
         !(week == 33 & tree == 2347), 
        # !(week == 18), #seems off
         #lwa_g_cm2 < 0.025 #weird outliers due to LMA, probably
  ) %>%
  mutate(water_potential = -1*water_potential, 
         lwc = lwc_mean) %>% 
  mutate(lwa_g_cm2_new = case_when(
    is.infinite(lwa_g_cm2_new) ~ NA_real_, 
    TRUE ~ as.numeric(lwa_g_cm2_new)))  %>% 
  
#issues found later on in analysis: 
  mutate(lwc_mean = case_when(
  tree %in% c(2343) & week %in% c(15) & time %in% c('md') ~ NA_real_,
  tree %in% c(2347) & week %in% c(15) & time %in% c('md') ~ NA_real_,
  tree %in% c(2370) & week %in% c(15) & time %in% c('md') ~ NA_real_,
  TRUE ~ as.numeric(lwc_mean))) %>% 
  filter(!(tree %in% 2012)) %>% #this tree doesn't actually exist
  filter(!(week == 17 & tree %in% c(2010, 2379))) %>% #these trees have something wrong too
# filter(!(species %in% c("live oak") & week %in% c(14, 15, 33))) %>%  #these weeks only had 2 trees collected with both PD and MD live oaks; remove!
  mutate(date_wp = case_when(
    week == 37 ~ as.Date("2022-09-15"), 
    TRUE ~ as.Date(date_wp)
  ))
 # filter(!week %in% c(9)) #this week has v low sample sizes, lets just not use!
```

Sample sizes: 
```{r}
sample_sizes <- analysis_bothspp_df %>% 
  drop_na(lwc_mean) %>% 
  group_by(date_wp, time, species) %>% 
  summarise(n = n()) %>% 
  pivot_wider(names_from = "species", 
             values_from = "n") %>% 
  arrange(date_wp, time) %>% 
  write_csv(here::here("processed-data", "sample_sizes_postQAQC.csv"))
sample_sizes

sample_sizes %>%
  kableExtra::kbl(caption = "Sample Sizes") %>%
  kableExtra::kable_classic(full_width = F, html_font = "Cambria") %>% 
  kableExtra::save_kable(here::here("figures", "SI Figures", "sample_sizes_table"), device = jpeg)

investigating <-analysis_bothspp_df %>% 
  filter(week == 29) %>% 
  distinct()
```
#FINAL CSVs:

```{r}
write.csv(analysis_bothspp_df, here("processed-data", paste0("analysis_bothspp",datver,".csv"))) 

#just blue oaks: 
analysis_df <- analysis_bothspp_df %>% 
  filter(species %in% c("blue oak"))
  
write.csv(analysis_df, here("processed-data", paste0("analysis",datver,".csv"))) 

#just live oaks: 
analysis_quag_df <- analysis_bothspp_df %>% 
  filter(species %in% c("live oak"))
  
write.csv(analysis_quag_df, here("processed-data", paste0("analysis_quag",datver,".csv"))) 
```

```{r}
plotly::ggplotly(analysis_bothspp_df %>% 
                   filter(species == "live oak", 
                          #week %in% c(12, 14, 15)
                          ) %>% 
  mutate(tree_num = as.numeric(tree)) %>% 
  ggplot(aes(y = water_potential, 
             x = tree, 
             shape = species,
             color = time)) +
  geom_point(alpha = .3) + 
  facet_wrap(~date_wp*week, nrow = 5))

plotly::ggplotly(analysis_df %>% 
                   filter(species == "blue oak", 
                          time == "md",
                          week %in% c(14)
                          ) %>% 
  mutate(tree_num = as.numeric(tree)) %>% 
  ggplot(aes(x = lwc_mean, 
             y = -1*water_potential, 
             color = time, 
             label = tree)) +
  geom_point(alpha = .3) + 
  facet_wrap(~date_wp*week, nrow = 5))
```

######Just looking at the data

#######Sample size/week

- What are the counts per week?
- How often did we measure both pd and md for same tree?
```{r}
lwc_counts <- analysis_df %>% 
  drop_na(lwc_mean) %>% 
  select(lwc_mean, tree, time) %>% 
  group_by(week, time) %>% 
  count() %>% 
  rename(n_lwc = n)

lwa_counts <- analysis_df %>% 
  select(lwa_g_cm2_new, tree, time) %>% 
  drop_na(lwa_g_cm2_new) %>% 
  group_by(week, time) %>% 
  count() %>% 
  rename(n_lwa = n)

counts <- merge(lwa_counts, lwc_counts, 
                by = c("week","time")) %>% 
  mutate(paired_or_all = "all trees")

 #only paired samples: 

lwc_counts_paired <- analysis_df %>% 
  drop_na(lwc_mean) %>% 
  select(lwc_mean, tree, time, species) %>% 
  group_by(tree, week) %>% 
  filter(all(c("pd", "md") %in% time)) %>%
  ungroup() %>% 
  distinct() %>% 
  group_by(week, time, species) %>% 
  count() %>% 
  rename(n_lwc = n) %>% 
  filter(time == "md") %>% 
  select(-time)

lwa_counts_paired <- analysis_df %>% 
  drop_na(lwa_g_cm2_new) %>% 
  select(lwa_g_cm2_new, tree, time, species) %>% 
  group_by(tree, week) %>% 
  filter(all(c("pd", "md") %in% time)) %>%
  ungroup() %>% 
  distinct() %>% 
  group_by(week, time, species) %>% 
  count() %>% 
  rename(n_lwa = n)%>% 
  filter(time == "md") %>% 
  select(-time)

counts_paired <- merge(lwa_counts_paired, lwc_counts_paired, 
                by = c("week", "time", "species")) %>% 
  mutate(paired_or_all = "paired", 
         time = "measured both") 

counts_all <- bind_rows(counts_paired, 
                        counts) %>% 
  select(-paired_or_all) %>% 
  arrange(week)

write.csv(counts_all, here("processed-data", 
               "counts.csv"))
```


#######Something weird with week 17?

```{r, warning=F}
analysis_df %>% 
  filter(week == '17', 
         site %in% c("LL"), 
         tree %in% c(2378, 2370, 2382, 2379, 2010)
        ) %>% 
  mutate(tree_week = paste(tree, week, sep = "_")
  ) %>% 
  mutate(water_potential = water_potential) %>% 
  ggplot(aes(y = water_potential, 
             x = -1*lwa_g_cm2, 
             color =    as.factor(tree)
             #shape = as.factor(site)
             )
         ) +
    geom_point(aes(shape = time)) +
  geom_smooth(method = "lm") +
    facet_wrap(~site)

bad_df <- analysis_df %>% 
  filter(week == '17', 
         site %in% c("LL"), 
         tree %in% c(2378, 2370, 2382, 2379, 2010)
        ) 

analysis_df %>% 
  filter(week == '17', 
        # site %in% c("LL"), 
        # tree %in% c(2378, 2370, 2382, 2379, 2010)
        ) %>% 
  mutate(tree_week = paste(tree, week, sep = "_")
  ) %>% 
  mutate(water_potential = water_potential) %>% 
  ggplot(aes(y = lwc_mean, 
             x = -1*lwa_g_cm2, 
             color =    as.factor(tree)
             #shape = as.factor(site)
             )
         ) +
    geom_point(aes(shape = time)
               ) +
  geom_smooth(method = "lm")

analysis_df %>% 
  ggplot(aes(y = lma_g_cm2, x = lwc_mean )) +
  geom_point()
```

#---------------------

#Add in CWC: 

Rained on 9/12 -- dont use
```{r}
datver_cwc <- "20231128" #Just change this in the 1_combined_processing.Rmd
dataversion_cwc <- paste0("Data_", datver_cwc)

cwc_raw <- read_csv(here(dataversion_cwc,"cwc_alldates.csv"), show_col_types = F) %>% 
  clean_names() %>%
  pivot_longer(cols = c(5:17),
    names_to = "date_x", 
               values_to = "cwc") %>% 
  mutate(date = gsub("x", "", date_x, fixed = TRUE),
         date_shift = mdy(date), 
         week = week(date_shift), 
         tree = as.character(comment)
         ) %>% 
 # select() %>% 
  group_by(tree, week) %>% 
  mutate(cwc = mean(cwc)) %>% 
  ungroup() %>% 
  mutate(week_cwc_link = week) %>% 
  filter(cwc > -100) %>% 
  select(-species, - date_x, -comment, -source)
  
  
cwc_raw_mpa <- cwc_raw %>% 
  select(cwc, tree, date_shift, site) %>% 
  mutate(week_shift = week(date_shift)) %>% 
  distinct() %>% 
  mutate(week_cwc_link = case_when( #dates of overflight and water potential, give unique names (shifts flight number) so they line up with CWC:
    week_shift %in% c(20) ~ 21, #shift date = 2022-05-17, mpa dates = 2022-05-23 and 05-25
    date_shift %in% c("2022-04-29") ~ 17, #shift date = 2022-04-29, mpa date = 2022-04-25
    TRUE ~ as.numeric(week_shift)))

weeks_cwc <- cwc_raw_mpa %>% 
  select(date_shift, week_shift, week_cwc_link) %>% 
  distinct()

mpa_raw_cwc <- wp_mean_df %>% #analysis_bothspp_df %>% 
  ungroup() %>% 
  mutate(water_potential = mean_mpa) %>% 
  select(water_potential, tree, species, time, date_wp) %>% 
  mutate(week_cwc_link = week(date_wp)) %>% 
  distinct()  %>% 
  mutate(week_cwc_link = case_when( #dates of overflight and water potential, give unique names (shifts flight number) so they line up with CWC:
    #week_shift %in% c(20) ~ 21, #shift date = 2022-05-17, mpa dates = 2022-05-23 and 05-25
    date_wp %in% c("2022-03-30") ~ 14,#shift date = 2022-04-05, mpa PD date = 2022-03-30
    date_wp %in% c("2022-03-25") ~ 12,
    TRUE ~ as.numeric(week_cwc_link)))

#final df with mpas and cwc matched by closest week (overflight and mpa collection):
mpa_cwc_all <- merge(cwc_raw_mpa, mpa_raw_cwc, 
                     by  = c("tree", "week_cwc_link"), 
                     all = T)  %>% 
  mutate(week = week_cwc_link, 
         tree = as.numeric(tree)) %>% 
   mutate(time_season = case_when(
    week < 17 ~ "before leafout", 
    week >= 17 ~ "after leafout", 
    TRUE ~ as.character("none")
  )) %>% 
  arrange(tree, week, time) %>% 
#Dates below are so that the colors line up with the cwc outlines on the timeline; comment them out if we want to have the actual MPa dates (so that the points are offset from the outlines)
  mutate(date_new = case_when(
   # date_wp %in% c("2022-03-15") ~ "2022-03-08",#CWC also measured on March 15th, keep
    date_wp %in% c("2022-03-15") ~ "2022-03-16", #overflight on 3/16
    date_wp %in% c("2022-03-23", "2022-03-25") ~ "2022-03-22", #ok to combine these, CWC only measured on March 222, some MPas on March 23
    date_wp %in% c("2022-03-30","2022-04-06","2022-04-04") ~ "2022-04-05", #CWC also measured on April 06, just use 04-04
    date_wp %in% c("2022-04-13", "2022-04-11") ~ "2022-04-12", #CWC also measred on April 11, just keep 
    date_wp %in% c("2022-04-20") ~ "2022-04-20", #
    date_wp %in% c("2022-04-27", "2022-04-25") ~ "2022-04-29", #
    date_wp %in% c("2022-05-04") ~ "2022-05-03",
    date_wp %in% c("2022-05-09") ~ "2022-05-11",
    date_wp %in% c("2022-05-09") ~ "2022-05-17", #might not have MPas?
    date_wp %in% c("2022-05-23", "2022-05-25") ~ "2022-05-29",
    TRUE ~ as.character(date_wp)
  )) %>% 
  filter(tree > 1000) #remove shrubs

#shift data from 2022-05-03 not used. Is there something about this date that would make it better than 2022-04-29?

weeks_cwc <- mpa_cwc_all %>% 
  select(date_shift, date_wp, week_shift, week) %>% 
  distinct()

write.csv(mpa_cwc_all, here("processed-data", paste0("cwc_analysis",datver,".csv")))
```

For figuring out WTF is up with the dates: 

```{r}
mpa_cwc_dates <- merge(wp_df%>% mutate(date = date_wp), 
                       cwc_raw %>% mutate(date = date_shift), all = T
                       )

write.csv(mpa_cwc_dates, here("processed-data", paste0("mpa_cwc_dates",datver,".csv")))
```

#TIMELINE: 
```{r, fig.height=5, fig.width=6}
wc_qudo_df <- read_csv(here::here("processed-data", paste0("analysis",datver,".csv")), show_col_types = FALSE) %>% 
 # filter(species == "live oak") %>% 
  distinct() %>% 
 # group_by(week, time) %>% 
 # filter(n() >= 10) %>% #only run on weeks where we have a decent number of samples
#  ungroup() %>% 
#  filter(!week == 21) %>% #this week doesn't have very many samples,
  mutate(time_season = case_when(
    week  <= 17 ~ "before leafout", 
    week > 17 ~ "after leafout", 
    TRUE ~ as.character("none")
  )) %>% 
  select(-plot) %>% 
  distinct() %>% 
  filter(!(week == 19 & time == "pd" & water_potential < -2))%>% 
  filter(!(week == 11)) %>% 
  drop_na(water_potential) 

wc_qudo_df%>% 
  ggplot(aes(x = week, y = water_potential, color = water_potential))+
  geom_hline(yintercept = 0) + 
  geom_point(size = 5) +
  facet_wrap(~time) + 
  color_grad_rev +
  labs(color = "Water Potential 
       (MPa)")+
  theme(legend.position = "right", 
        axis.title.y = element_blank(), 
        axis.title.x = element_blank(), 
        axis.text.y = element_blank(),
        axis.ticks = element_blank(), 
       legend.title.align = -.5,
       legend.direction = "vertical",
       legend.key.size = unit(1, "cm")
        )

timeline_df <- wc_qudo_df %>% 
 # select(week, date_wp, time) %>% 
  mutate(date_wp = as.character(date_wp)) %>% 
  mutate(date_new = case_when(
    date_wp %in% c("2022-03-15") ~ "2022-03-08",
    date_wp %in% c("2022-03-23") ~ "2022-03-25",
    date_wp %in% c("2022-03-30","2022-04-06") ~ "2022-04-04",
    date_wp %in% c("2022-04-13") ~ "2022-04-11",
    date_wp %in% c("2022-04-27") ~ "2022-04-25",
    date_wp %in% c("2022-05-04") ~ "2022-05-09",
    date_wp %in% c("2022-05-23") ~ "2022-05-25",
    TRUE ~ as.character(date_wp)
  )) %>% 
  mutate(date_new = ymd(date_new)) %>% 
  select(date_new, week, water_potential, time) %>% 
  distinct()

mpa_cwc_dates


```


```{r, fig.height=5, fig.width=20}
# timeline_plot <- mpa_cwc_dates %>% 
#   #drop_na(water_potential) %>% 
#   filter(time == "md") %>% 
  ggplot() +
  geom_hline(yintercept = 0) +
  geom_point(aes(x = date, 
             y = 0, 
             color = mpa),
             data = mpa_cwc_dates %>% drop_na(mpa) %>% filter(time == "md"), 
             size = 10,
             pch=21, 
           size=10) +
    geom_point(aes(x = date,
             y = 0),
              data = mpa_cwc_dates %>% drop_na(cwc),
           pch=21,
           size=10,
           stroke=2,
           colour="black") +
  color_grad_rev+
  ylim(-.25, +.25)  + 
  theme(legend.position = "none", 
        axis.title.y = element_blank(), 
        axis.title.x = element_blank(), 
        axis.text.y = element_blank(),
        axis.ticks = element_blank(), 
        axis.text.x = element_text(angle = 60, vjust = 0.5, hjust=.5, size= 20)
        ) +
scale_x_date(date_labels = '%b. %d', 
             breaks = as.Date(mpa_cwc_dates$date))
timeline_plot

#ggsave(here("figures", "timeline_plot"), timeline_plot, device = "jpg", width = 12, height = 2, dpi = 300)
```


For Jean: 
```{r, eval = F}
jean_df <- cwc_analysis_df %>% 
  select(tree, species, time, date_shift, date_wp, cwc, lwc_mean, lma_g_cm2, lwa_g_cm2, water_potential) %>% 
  group_by(tree) %>% 
  fill(c(species), .direction = "downup")

write.csv(jean_df, here("processed-data", "jean_lwa.csv")) 
```

#---------

#Add in LFM: 

#####LFM + RWC + ALAS + WC + WP

LFM is split into different into different categories: stem vs. leaves (type column), and y1 vs. y0 (year column)


```{r}
lfm_df_all <- read_csv(here("processed-data", paste0("lfm_alldates_",datver,".csv"))) %>% 
  select(-'...1') %>% 
  mutate(year_alas = year) #rename year column so that it plays nicely with the main df (we want the alas year and the LFM years to line up, so year_alas and year_lfm area just the same)

lfm_rwc_alas_wp_wc_df <- merge(wp_wc_rwc_alas_df, lfm_df_all, by = c("tree", "week", "time", "type", "year_alas"), 
                               all = T) %>% 
  filter(tree > 100) %>% 
  distinct()  %>% 
  dplyr::group_by(tree, time, week) %>%
  fill(c(date_wp, mpa_mean, lwc_mean), .direction = "downup") %>%
  dplyr::ungroup() %>% 
  group_by(tree) %>% 
  fill(c(species, site, plot), .direction = "downup") %>% 
  ungroup() %>% 
  mutate(rwc_lfm_percent = ((lfm_wet_wt_g-lfm_dry_wt_g)/lfm_dry_wt_g)/mean_swc_per_dry_g * 100,
         rwc_lwc_percent = ((lwc_mean/mean_swc_per_dry_g)*100)) %>% 
  # mutate(rwc_lwc_percent = case_when(
  #   type %in% c("stem") ~ NA_real_, 
  #   TRUE ~ as.numeric(rwc_lwc_percent))) %>% 
  distinct() %>% 
  # filter(lfm_percent < 200, 
  #        lfm_percent > 0) %>% 
  rename(type_lfm = type) %>% 
 # filter(!(type_lfm == "stem")) %>% 
  distinct() %>% 
  mutate(ilfm = 1/(lfm_percent/100)) %>% 
  filter(ilfm < 2.5,
         ilfm > 0) #%>% 
  # select(tree, week, time, type_lfm, type_rwc, mpa_mean, lfm_percent, lfm_dry_wt_g, lfm_wet_wt_g, date_lfm, date_wp)
  #filter(!is.na(rwc_lwc_percent))
```

#*FINAL (NEW): write.csv

```{r}
lfm_rwc_alas_wp_wc_leaves_df <- lfm_rwc_alas_wp_wc_df %>% 
  filter(!type_lfm == "stem") %>% 
  select(tree, species, week, time, type_lfm, year, lfm_wet_wt_g, lfm_dry_wt_g, lfm_percent, date_lfm, mpa_mean, notes) %>% 
  distinct() %>% 
  mutate(tree = as.numeric(tree)) %>% 
  filter(tree > 100)

filtered_lfm_df <- lfm_rwc_alas_wp_wc_leaves_df %>%
  group_by(tree, species, week, date_lfm, type_lfm, time, year) %>%
  filter(n() > 1) %>%
  ungroup()

filtered_lfm_df %>% 
  write.csv(here("processed-data", paste0("lfm_dupes",datver,".csv")))

summary(lfm_rwc_alas_wp_wc_leaves_df)
  
write.csv(lfm_rwc_alas_wp_wc_leaves_df, here("processed-data", paste0("lfm_rwc_alas_wp_wc_df",datver,".csv")))
```

#LFM dataset (Pepperwood)

```{r, eval = F}
lfm_mpa_df <- lfm_rwc_alas_wp_wc_df %>% 
  group_by(tree, date_lfm, type_lfm, time) %>% 
  mutate(lfm_percent = mean(lfm_percent, na.rm = T)) %>% 
  mutate(ilfm = 1/(lfm_percent/100)) %>% 
  select(tree, week,  date_lfm, time, time, type_lfm, plot, site, species, mpa_mean, lfm_percent, ilfm) %>% 
  drop_na(species) %>% 
  filter(type_lfm == "leaf") %>% 
  distinct() %>% 
  mutate(date_lfm = ymd(date_lfm))

lfm_mpa_df %>% 
  ggplot(aes(y = lfm_percent, 
             x = date_lfm)) +
  geom_point() +
  scale_x_date(breaks = "month")

n_week <- lfm_mpa_df %>% 
  filter(time == "md") %>% 
  mutate(week = week(date_lfm)) %>% 
  group_by(week, species) %>% 
  mutate(n = n()) %>% 
  select(n, species, week) %>% 
  distinct()

unique(lfm_mpa_df$date_lfm)

write.csv(lfm_mpa_df, here("processed-data", paste0("sedgwick_22_lfm_mpa_df",datver,".csv")))
```

- water potentials for week 17 piney, cucu, and chamise are missing? (04.25.2022 and 04.27.2022) We have LFM and LWC for this week from those sites however.  

- missing LFM from week 18 (05.04.2022)

- Predawns from early may, middays from end of March are missing (?)
