==> scripts/2.1_var_decomp_mpa~lwc.Rmd <==
---
title: "Establishing relationships between LWC and MPa"
author: "Indra Boving"
date: "11/8/2022"
output: html_document
---

##Setup
```{r setup, include=FALSE}
#install.packages("MuMIn")
#devtools::install_github("an-bui/calecopal")

library(data.table)
library(janitor)
library(here)
library(tidyverse)
library(lubridate)
library(readxl)
library(gridExtra)
library(calecopal)
library(rstatix)
library(vars)
library(MuMIn)
library(lme4) #for mixed effects models
library(cowplot)
select = dplyr::select
#if (!require("BiocManager", quietly = TRUE))
   # install.packages("BiocManager")
#BiocManager::install("variancePartition")

library('variancePartition')

source(here::here("scripts", "scripts_functions", "plotPercentBars_IB.R"))

source(here::here("scripts", "scripts_functions", "figure_info.R"))
```

#read in data: 
```{r}
wc_all_raw <- read_csv(here("processed-data", paste0("wp_wc_rwc_",datver,".csv")), show_col_types = F) %>%   #mutate(lwc_mean = lwc_mean_combined) %>% 
 # here("processed-data", paste0("LWC_LFM_WP_dates",datver,".csv")))
  filter(lwc_mean > 0, 
         lwc_mean < 3)


# wc_all_raw2 <- read_csv(here("processed-data", paste0("LWC_LFM_WP_dates",datver,".csv")), show_col_types = F) %>%   #mutate(lwc_mean = lwc_mean_combined) %>% 
#  # here("processed-data", paste0("LWC_LFM_WP_dates",datver,".csv")))
#   filter(lwc_mean > 0, 
#          lwc_mean < 3)

wc_all_lma <- wc_all_raw %>% 
  filter( #week > 15, 
         species == "blue oak", 
         lwc_mean < 10) %>% 
  group_by(tree, week, time, lwc_mean, site) %>% 
  mutate(lma = mean(lma_g_cm2, na.rm = T)) 
  
```

Before leafout: 

```{r}
qudo_wc_all <- wc_all_raw %>% 
  filter( #week > 15, 
         species == "blue oak", 
         lwc_mean < 10) %>% 
  group_by(tree, week, time, lwc_mean, site) %>% 
  mutate(water_potential = mean(mpa_mean), 
         lwc = lwc_mean) 

qudo_wc_linear <- qudo_wc_all %>% 
  filter(week > 17)
```

```{r}
full_nice <- qudo_wc_all %>% 
  ggplot(aes(y = water_potential, 
             x = lwc, 
             color = week)) +
  geom_point() +
 # geom_smooth(method = "lm", 
            #  se = F, 
            #  color = "black") +
  color_grad +
  #color_many +
  labs(y = "MPa",
       x = "LWC", 
       color = "Week") +
  theme(legend.position = c(0.1, 0.78)) +
  scale_x_reverse()
full_nice
```

Making models of each variable of interest: 

Use al:as or lma to pick inflection point (currently just using week):

Variance decomposition, build 3-4 models:
LWC ~ null
~time
~mpa
~mpa + time

Look at marginal R^2, decompose add them together, overlap should be greater than sum (since they explain some same)


```{r}
m0 <- lmer(water_potential ~ 1 + (1|tree), data = qudo_wc_all)
m1 <- lmer(water_potential ~ week + (1|tree), data = qudo_wc_all)
m2 <- lmer(water_potential ~ lwc + (1|tree), data = qudo_wc_all)
m3 <- lmer(water_potential ~ week + lwc + (1|tree), data = qudo_wc_all)

#try using r.squaredGLMM:
mods_list <- list(m0, m1, m2, m3)

r2 <- lapply(mods_list, r.squaredGLMM)
r2

r2m <- c(r2[[1]][1], r2[[2]][1], r2[[3]][1], r2[[4]][1]) #vector of r2s
mods <- c("m0", "m1", "m2", "m3")

mod_all_df1_full <-  data.frame(mods, r2m)

mod_all_df_full <- mod_all_df1_full %>% 
  pivot_wider(names_from = mods, 
              values_from = r2m) %>% 
  mutate(
    #other_effects =  m3 - week_effect - lwc_effect,
    week_effect = m3 - m2, #week + mpa mod - mpa mod
         lwc_effect = m3 - m1, #week + mpa mod - week mod
         week_lwc_effect = m3, #week + mpa mod 
    other_effects =  m3 - week_effect - lwc_effect,
         ) #full minus the effects of mpa and week alone
mod_all_df_full

#for plotting (maybe?)
mod_all_long_full <- mod_all_df_full %>% 
  select(week_effect, lwc_effect, week_lwc_effect, other_effects) %>% 
  pivot_longer(cols = c(week_effect, lwc_effect, week_lwc_effect, other_effects), 
    names_to = "effect") %>% 
  distinct() %>% 
  mutate(analysis = case_when(
    effect %in% c("week_effect", "lwc_effect", "other_effects") ~ "Var. Decomp", 
    effect %in% c("week_lwc_effect") ~ "Full Mod.", 
    TRUE ~ as.character(effect)), 
    decomp = "Full Date Range")
mod_all_long_full

mod_all_long_full %>% 
  ggplot(aes(x = analysis, 
             y = value, 
         fill = effect)) +
  geom_col(stat = "identity") +
 # geom_text(aes(label = value)) +
  color_many +
  labs(y = "R squared", 
       x = "Predictor") +
  color_fill

```

Once leafout finishes, then things get linear: 

- currently just using week approximation, but will use LMA plateau once we have that together:

#Datasets: 
```{r}
qudo_wc_full <- wc_all_raw %>% 
   filter(#week > 15, 
         species == "blue oak", 
         lwc_mean < 10) %>% 
  group_by(tree, week, time, lwc_mean) %>% 
  mutate(water_potential = mean(mpa_mean), 
         lwc = lwc_mean) %>% 
  mutate(section = "full") %>% 
  select(section, species, week, water_potential, lwc, site)


qudo_wc_linear <- wc_all_raw %>% 
  filter(week > 18, 
         species == "blue oak", 
         lwc_mean < 10) %>% 
  group_by(tree, week, time, lwc_mean) %>% 
  mutate(water_potential = mean(mpa_mean), 
         lwc = lwc_mean) %>% 
  mutate(section = "linear")%>% 
  select(section, species, week, water_potential, lwc, site)

qudo_wc_all <- bind_rows(qudo_wc_linear, qudo_wc_full) %>% 
  distinct
```


```{r}
nice_all <- qudo_wc_all %>% 
   ggplot(aes(y = water_potential, 
             x = lwc, 
             color = week)) +
  geom_point() +
#  geom_smooth(method = "lm", se = F, color = "black") +
  facet_wrap(~section, scales = "free") +
  color_grad +
  labs(y = "Water Potential", 
       x = "LWC", 
       color = "Week") +
  theme(legend.position = "none") 
nice_all
```

```{r}
lin_nice <- qudo_wc_linear %>% 
  ggplot(aes(y = water_potential, 
             x = lwc, 
             color = week)) +
  geom_point() +
  geom_smooth(method = "lm", 
              se = F, 
              color = "black") +
  color_grad +
  labs(y = "", 
       x = "", 
       color = "Week") +
  theme(legend.position = "none") 
lin_nice
```

```{r}
together <- cowplot::plot_grid( full_nice, 
                                lin_nice,
                               ncol = 1, 
                               rel_heights = c(1, 1))
together
```

```{r}
m0 <- lmer(water_potential ~ 1 + (1|tree), data = qudo_wc_linear)
m1 <- lmer(water_potential ~ week + (1|tree), data = qudo_wc_linear)
m2 <- lmer(water_potential ~ lwc + (1|tree), data = qudo_wc_linear)
m3 <- lmer(water_potential ~ week + lwc + (1|tree), data = qudo_wc_linear)

#try using r.squaredGLMM:
mods_list <- list(m0, m1, m2, m3)

r2 <- lapply(mods_list, r.squaredGLMM)
r2

r2m <- c(r2[[1]][1], r2[[2]][1], r2[[3]][1], r2[[4]][1]) #vector of r2s
mods <- c("m0", "m1", "m2", "m3")

mod_all_df1 <-  data.frame(mods, r2m)

mod_all_df <- mod_all_df1 %>% 
  pivot_wider(names_from = mods, 
              values_from = r2m) %>% 
  mutate(
    #other_effects =  m3 - week_effect - lwc_effect,
    week_effect = m3 - m2, #week + mpa mod - mpa mod
         lwc_effect = m3 - m1, #week + mpa mod - week mod
         week_lwc_effect = m3, #week + mpa mod
    other_effects =  m3 - week_effect - lwc_effect
         ) #full minus the effects of mpa and week alone

mod_all_df

#for plotting (maybe?)
mod_all_long <- mod_all_df %>% 
  select(week_effect, lwc_effect, week_lwc_effect, other_effects) %>% 
  pivot_longer(cols = c(week_effect, lwc_effect, week_lwc_effect, other_effects), 
    names_to = "effect") %>% 
  distinct() %>% 
  mutate(analysis = case_when(
    effect %in% c("week_effect", "lwc_effect", "other_effects") ~ "Var. Decomp", 
    effect %in% c("week_lwc_effect") ~ "Full Mod.", 
    TRUE ~ as.character(effect)), 
    decomp = "After Leafout")
mod_all_long

mod_all_long %>% 
  ggplot(aes(x = analysis, 
             y = value, 
         fill = effect)) +
  geom_col(stat = "identity") +
 # geom_text(aes(label = value)) +
  color_many +
  labs(y = "R squared", 
       x = "Predictor") +
  color_fill
```

```{r}
mod_both_df_long <- bind_rows(mod_all_long_full, mod_all_long) %>% 
  filter(!effect %in% c("week_lwc_effect"))

mod_both_df_long %>% 
  ggplot(aes(x = decomp, 
             y = value, 
         fill = effect)) +
  geom_col(stat = "identity") +
 # geom_text(aes(label = value)) +
  color_many +
  labs(y = "R squared", 
       x = "Data", 
       fill = "Effect") +
  scale_fill_manual(labels = c("LWC", "LWC & Week", "Week"), 
                    values = met.brewer("Cross", direction = -1))
```

#LEEs example: (use this)

#####Whole df: 
```{r}
#alternative method with random effects

mvar <- lmer(water_potential ~ lwc + (1|site/tree) + (1|week), data=qudo_wc_full)

r2m <- r.squaredGLMM(mvar)[1] # plust the marginal R2 of lwc
#r2m 

var(qudo_wc_full$water_potential, na.rm=T) # 1.18
vest <- data.frame(VarCorr(mvar))
#sum(vest$vcov) + 0.09 #1.001

wc_all_vest_df <- vest %>% 
  add_row(vcov = r2m, 
          grp = "marginal_r2") %>% 
  mutate(dataset= "All Dates")

#wc_all_vest_df
sum(wc_all_vest_df$vcov)
```

#####Linear: 
```{r}
#alternative method with random effects

mvar_lin <- lmer(water_potential ~ lwc + (1|site/tree) + (1|week), data=qudo_wc_linear)

r2m <- r.squaredGLMM(mvar_lin)[1] # plust the marginal R2 of lwc
 
r2m #0.107

var(qudo_wc_linear$water_potential, na.rm=T) # 1.0239
vest_lin <- data.frame(VarCorr(mvar_lin))
sum(vest$vcov) + 0.09 #1.001

wc_linear_vest_df <- vest_lin %>% 
  add_row(vcov = r2m, 
          grp = "marginal_r2") %>% 
  mutate(dataset= "After Leaf Expansion")

#wc_linear_vest_df
#sum(wc_linear_vest_df$vcov) #Close enough to 1?
```

#****Plot

```{r}
#Colors: 
color_fill_vdc <- scale_fill_manual(values = c("Residuals" = "#D9D9D9",
                                               "Site" = "#252525",
                                                "Tree" =  "#525252",
                                               "LWC" = "#C969A1" ,
                                                "Week" = "#62929a"))

#greys <- list(c("#FFFFFF" "#F0F0F0" "#D9D9D9" "#BDBDBD" "#969696" "#737373" "#525252" "#252525" "#000000"))

#Combine into 1 df: 
vdc2 <- bind_rows(wc_linear_vest_df, wc_all_vest_df) %>% 
  group_by(dataset) %>% 
  mutate(total = sum(vcov)) %>% 
  ungroup() %>% 
  mutate(percent_var = vcov/total) %>% 
  mutate(grp_new = case_when(
    grp %in% c("marginal_r2") ~ "LWC", 
    grp %in% c("Residual") ~ "Residuals",
    grp %in% c("site") ~ "Site",
      grp %in% c("tree:site") ~ "Tree", 
      grp %in% c("week") ~ "Week", 
      TRUE ~ as.character(grp)
  ))

vdc2_plot <- vdc2 %>% ggplot(aes(x = dataset, 
                    y = percent_var,
                    fill = grp_new)) +
  geom_col(stat = "identity") +
  color_fill_vdc +
  labs(y = "", 
       x = "", 
       fill = "Variable") +
  theme(
    legend.position = "top",
  strip.background = element_blank(),
 # strip.text.y = element_blank(), 
  axis.text.x = element_text(size = 12), 
   axis.text.y = element_text(size = 12), 
 # axis.text.y = element_blank(), 
  #axis.ticks.y = element_blank(), 
  axis.title = element_text(size = 20), 
 # legend.title = element_text(size = 14), 
  legend.title = element_blank(), 
 legend.text= element_text(size = 12), 
  ) +
    scale_x_discrete(limits=rev)
  
vdc2_plot

ggsave(here("figures", "vdc2_plot"), vdc2_plot, device = "jpg", width = 5, height = 7, dpi = 300)
```

#****LMA over time: 

```{r}
lma_plot <- wc_all_lma  %>% 
  filter(lma_g_cm2 < .025) %>% 
  group_by(week, species) %>% 
  mutate(mean_lma = mean(lma_g_cm2, na.rm = T)) %>% 
  ggplot(aes(y = mean_lma, x = week )) +
  geom_point() +
  labs(x = "Week", 
       color = "week", 
       y = "Leaf Mass/Area") +
  theme(legend.position="top",
      strip.background = element_blank(),
      strip.text.x = element_text(size = 12),
      plot.title = element_text(size=13),
      axis.title = element_text(size = 16),
     axis.text = element_text(size = 12),
     legend.key=element_blank(), 
     legend.text = element_text(size = 13), 
     legend.title = element_text(size = 16),
     legend.margin=margin(0,0,0,0),
      legend.box.margin=margin(-5,-8,-8,-8)
    ) +
 # geom_hline(yintercept = 0, linetype="dotted") + 
    color_many_2 +
  guides(color = guide_legend(nrow = 2))  + 
  scale_x_continuous(breaks = c(10, 13, 12, 14,15, 17, 19, 21, 29, 33, 37)
                      # seq(10, 30, by = 1)
                     ) 
```

```{r}
vardecomp_lma_plot <- cowplot::plot_grid(vdc2_plot, lma_plot, 
                                   nrow = 1, rel_heights = c(1, 1), 
                                   rel_widths = c(1, 1)
                                   ) 

vardecomp_lma_plot

ggsave(here("figures", "vardecomp_lma_plot"), vardecomp_lma_plot, device = "jpg", width = 12, height = 5, dpi = 300)
```
