==> scripts/3.00_BLUES_LWA.Rmd <==
---
title: "Establishing relationships between lwa_g_cm2_new and MPa, Part 2"
author: "Indra Boving"
date: "11/9022"
output:
  html_document: default
  word_document: default
---

#*BOOTSTRAPPING*

#Setup

```{r setup, include=FALSE}
knitr::opts_chunk$set(warning = FALSE, message = FALSE) 
library(tidyverse)
library(boot)
source(here::here("scripts", "scripts_functions", "figure_info.R"))
library(lme4)
library(MuMIn) #for r.squaredGLMM fxn
```

#DATA: 

```{r}
data_qudo_raw <- read_csv(here::here("processed-data", paste0("analysis",datver,".csv")), show_col_types = FALSE) %>% 
 # filter(species == "live oak") %>% 
  distinct() %>% 
 # group_by(week, time) %>% 
 # filter(n() >= 10) %>% #only run on weeks where we have a decent number of samples
#  ungroup() %>% 
#  filter(!week == 21) %>% #this week doesn't have very many samples,
  mutate(time_season = case_when(
    week  <= 17 ~ "before leafout", 
    week > 17 ~ "after leafout", 
    TRUE ~ as.character("none")
  )) %>% 
  select(-plot) %>% 
  distinct()

data_og <- data_qudo_raw %>% 
  select(tree, week, water_potential, lwa_g_cm2_new, time_season, time, species, site) %>% 
  distinct()

daat_wk21 <- data_og %>% 
  filter(week == 21)
unique(data_og$tree)
```


#TIME:
```{r}
data_all <- data_og %>% 
  filter(time == "md") 

# Filter for trees measured "before leafout"
before_leafout_data <- data_all %>%
  filter(time_season == "before leafout")

# Filter for trees measured "after leafout"
after_leafout_data <- data_all %>%
  filter(time_season == "after leafout")

# Find the common set of trees that appear in both datasets
common_trees <- intersect(before_leafout_data$tree, after_leafout_data$tree)

# Filter the original dataset to include only the common trees
common_trees_data <- data_all %>%
  filter(tree %in% common_trees)

# Assuming you have the 'common_trees_data' dataframe

# Get the unique tree IDs in 'common_trees_data'
unique_tree_ids <- unique(common_trees_data$tree)

# Count the number of unique tree IDs
num_unique_tree_ids <- length(unique_tree_ids)

# Print the number of unique tree IDs
cat("Number of unique tree IDs in common_trees_data:", num_unique_tree_ids, "\n")

# Filter the original dataset to include only the common tree IDs
filtered_data <- data_all[data_all$tree %in% common_trees, ] 
```

Pull 14 trees, get mean slope for those 40, do 1000x. 


###Before leafout: 
```{r}
data <- filtered_data  %>% 
  filter(#time_season == "before leafout", 
         week %in% c(10:17)) %>% 
  distinct() %>% 
  drop_na(water_potential, lwa_g_cm2_new) 

num_unique_trees <- length(unique(data$tree))
cat("Number of unique trees in the dataset:", num_unique_trees, "\n")

# Assuming your dataset is called 'data'
# Make sure 'data' is properly loaded before proceeding

# Define the number of bootstrap iterations
n_iterations <- 10

# Number of random trees to sample
n_sample_trees <- 40  # Adjust as needed

# Create an empty data frame to store bootstrapped slopes
bootstrapped_slopes_df <- data.frame()

# Function to calculate the slope for a specific tree within each iteration
calculate_slope_for_tree <- function(tree_data) {
  model <- lm(water_potential ~ lwa_g_cm2_new, data = tree_data)
  return(coef(model)[2])  # Extract the slope coefficient
}

#Do the bootstrapping: 

 for (i in 1:n_iterations) {
   
# Randomly sample 'n_sample_trees' trees from your dataset
sampled_tree_ids <- sample(unique(data$tree), size = n_sample_trees, replace = TRUE)

# Iterate through each sampled tree and perform bootstrapping separately
      for (tree_id in sampled_tree_ids) {
        
        tree_data <- data %>%
          filter(tree == tree_id)
        
        # Initialize a container to store slopes for each iteration
        slopes <- vector("double", length = 1)
      
          # Calculate the slope for the bootstrapped sample
          slopes[] <- calculate_slope_for_tree(tree_data)
        
        # Store the bootstrapped slopes for this tree in the data frame
        bootstrapped_slopes_df <- bind_rows(
          bootstrapped_slopes_df,
          data.frame(
            tree = tree_id,
            slope = slopes, 
            round = i
          ) 
        ) 
      }
 }

        ci_values <- bootstrapped_slopes_df %>%
        group_by(round) %>% 
        drop_na(slope) %>%
       mutate(
          mean_slope_round = mean(slope, na.rm = T)
        ) %>% 
          ungroup() %>% 
          mutate(mean = mean(mean_slope_round, na.rm = T),
          lower_bound = quantile(mean_slope_round, 0.025, na.rm = TRUE),
          upper_bound = quantile(mean_slope_round, 0.975, na.rm = TRUE))
          
        
        # Merge the CI values back into the original dataframe
      tbootstrapped_slopes_df_before <- ci_values 
      
# Print the updated dataframe with CI values
print(tbootstrapped_slopes_df_before)

summary(tbootstrapped_slopes_df_before)

mean_slope_before = tbootstrapped_slopes_df_before %>% 
  mutate(mean_slope = mean(mean, na.rm = T),
         pre_post = "before leafout") %>% 
  select(mean_slope, lower_bound, upper_bound, pre_post) %>% 
  distinct()
mean_slope_before
```

###After leafout: 
```{r}
data <- common_trees_data %>% 
   filter(week %in% c(18:40)) %>% 
 # filter(time_season == "after leafout") %>% 
  select(tree, week, water_potential, lwa_g_cm2_new) %>% 
  distinct() %>% 
  drop_na(water_potential, lwa_g_cm2_new)

# Assuming your dataset is called 'data'
# Make sure 'data' is properly loaded before proceeding

# Define the number of bootstrap iterations
n_iterations <- 10

# Number of random trees to sample
n_sample_trees <- 40  # Adjust as needed

# Create an empty data frame to store bootstrapped slopes
bootstrapped_slopes_df <- data.frame()

# Function to calculate the slope for a specific tree within each iteration
calculate_slope_for_tree <- function(tree_data) {
  model <- lm(water_potential ~ lwa_g_cm2_new, data = tree_data)
  return(coef(model)[2])  # Extract the slope coefficient
}

#Do the bootstrapping: 

 for (i in 1:n_iterations) {
   
# Randomly sample 'n_sample_trees' trees from your dataset
sampled_tree_ids <- sample(unique(data$tree), size = n_sample_trees, replace = TRUE)

# Iterate through each sampled tree and perform bootstrapping separately
      for (tree_id in sampled_tree_ids) {
        
        tree_data <- data %>%
          filter(tree == tree_id)
        
        # Initialize a container to store slopes for each iteration
        slopes <- vector("double", length = 1)
      
          # Calculate the slope for the bootstrapped sample
          slopes[] <- calculate_slope_for_tree(tree_data)
        
        # Store the bootstrapped slopes for this tree in the data frame
        bootstrapped_slopes_df <- bind_rows(
          bootstrapped_slopes_df,
          data.frame(
            tree = tree_id,
            slope = slopes, 
            round = i
          ) 
        ) 
      }
 }

        ci_values <- bootstrapped_slopes_df %>%
        group_by(round) %>% 
        drop_na(slope) %>%
       mutate(
          mean_slope_round = mean(slope, na.rm = T)
        ) %>% 
          ungroup() %>% 
          mutate(mean = mean(mean_slope_round, na.rm = T),
          lower_bound = quantile(mean_slope_round, 0.025, na.rm = TRUE),
          upper_bound = quantile(mean_slope_round, 0.975, na.rm = TRUE))
          
        
        # Merge the CI values back into the original dataframe
      tbootstrapped_slopes_df_after <- ci_values 
      
# Print the updated dataframe with CI values
print(tbootstrapped_slopes_df_after)

summary(tbootstrapped_slopes_df_after)

mean_slope_after = tbootstrapped_slopes_df_after %>% 
  mutate(mean_slope = mean(mean, na.rm = T),
         pre_post = "after leafout") %>% 
  select(mean_slope, lower_bound, upper_bound, pre_post) %>% 
  distinct()
mean_slope_after
```

```{r}
#Combining before and after: 
time_boot_df <- bind_rows(mean_slope_before, mean_slope_after) %>% 
  select(pre_post, mean_slope, lower_bound, upper_bound) %>% 
  distinct() %>% 
  mutate(analysis = "time, ind. tree")
```

```{r}
time_boot_df %>% 
  group_by(pre_post) %>% 
  mutate(tree_mean = mean(mean_slope, na.rm = T)) %>% 
  ggplot() +
  geom_point(aes(y = tree_mean, 
             x = pre_post)) +
  geom_point(aes(y = upper_bound, 
                 x= pre_post), color = "blue")  +
  geom_point(aes(y = lower_bound, 
                 x= pre_post), color = "pink")  +
  theme_minimal()
```

#SPACE: 

```{r}
data <- data_og %>% 
  filter(time == "md") %>% 
  select(tree, week, water_potential, lwa_g_cm2_new) %>% 
  distinct() %>% 
  drop_na(lwa_g_cm2_new, water_potential)

head(data)

data %>% 
ggplot(aes(y = water_potential, 
           x = lwa_g_cm2_new, 
           color = as.factor(week))) +
  geom_point(alpha = .5) +
  geom_smooth(method = "lm", se = F)

# Define the number of bootstrap iterations
n_iterations <- 10

# Create an empty dataframe to store results
final_df_space <- data.frame(week = integer(), mean_slope = numeric(), lower_bound = numeric(), upper_bound = numeric())

# Get unique weeks in your dataset
unique_weeks <- unique(data$week)

# Iterate through each week
for (x in unique_weeks) {
  # Step 1: Filter data for the current week
  week_data <- data %>%
    filter(week == x)
  
  # Step 2: Initialize a container to store bootstrapped slopes
  slopes <- vector("double", length = n_iterations)
  
  # Step 3: Perform bootstrapping to estimate slopes
  for (i in 1:n_iterations) {
    # Randomly sample rows with replacement
    boot_indices <- sample(nrow(week_data), replace = TRUE)
    boot_sample <- week_data[boot_indices, ]
    
    # Calculate the slope for the bootstrapped sample
    model <- lm(water_potential ~ lwa_g_cm2_new, data = boot_sample)
    slopes[i] <- coef(model)[2]
  }
  
  # Step 4: Calculate the mean slope and 95% CI
  mean_slope <- mean(slopes)
  lower_bound <- quantile(slopes, 0.025)
  upper_bound <- quantile(slopes, 0.975)
  
  # Step 5: Store the results in the final dataframe
  final_df_space <- bind_rows(final_df_space, data.frame(week = x, mean_slope = mean_slope, lower_bound = lower_bound, upper_bound = upper_bound))
}

# Print the final dataframe
print(final_df_space)

final_df_space %>% 
  ggplot(aes(x = week)) +
   geom_segment(aes(y= lower_bound, 
                   yend = upper_bound, 
                   x = week, 
                   xend = week, 
                   color = as.factor(week)))+
  geom_point(aes(y = mean_slope), color = "cyan") +
  geom_point(aes(y = lower_bound), color = "pink") +
  geom_point(aes(y = upper_bound), color = "gold") +
  geom_hline(yintercept =0) +
  labs(y = "Slope",
       x = "Week")
  
```
#HYDRATION: 

```{r}
data <- data_og %>% 
  select(tree, week, water_potential, lwa_g_cm2_new, time) %>%
  drop_na(lwa_g_cm2_new, water_potential) %>% 
  group_by(tree, week) %>%
  filter(all(c("pd", "md") %in% time)) %>%
  ungroup() %>% 
  distinct() 

df_all_groups <- data %>% 
  mutate(tree_week = paste(tree, week, sep = "_")
  ) %>% 
  mutate(water_potential = water_potential)

df_grey <- df_all_groups

df_1 <- df_all_groups %>% 
  filter(week %in% c('12'))

df_2 <- df_all_groups %>% 
  filter(week %in% c('21'))

# df_3 <- df_all_groups %>% 
#   filter(week %in% c('19'))

df_4 <- df_all_groups %>% 
  filter(week %in% c('37'))

nice_hydration_layered <- plotly::ggplotly(ggplot() +
    geom_point(aes(y = water_potential, x = lwa_g_cm2_new, shape = time), data = df_grey, color = "grey") +
    geom_line(aes(y = water_potential, x = lwa_g_cm2_new, group = tree_week), color = "grey", data = df_grey) +
    geom_point(aes(y = water_potential, x = lwa_g_cm2_new, shape = time, color = "df_1"), 
               data = df_1 ) +
    geom_line(aes(y = water_potential, x = lwa_g_cm2_new, group = tree_week, color = "df_1"), 
              data = df_1) +
    geom_point(aes(y = water_potential, x = lwa_g_cm2_new, shape = time, color = "df_2"), 
               data = df_2) +
    geom_line(aes(y = water_potential, x = lwa_g_cm2_new, group = tree_week, color = "df_2"), 
              data = df_2) +
    # geom_point(aes(y = water_potential, x = lwa_g_cm2_new, shape = time, color = "df_3"),
    #            data = df_3) +
    # geom_line(aes(y = water_potential, x = lwa_g_cm2_new, group = tree_week, color = "df_3"),
    #           data = df_3) +
    geom_point(aes(y = water_potential, x = lwa_g_cm2_new, shape = time, color = "df_4"),
               data = df_4) +
    geom_line(aes(y = water_potential, x = lwa_g_cm2_new, group = tree_week, color = "df_4"),
              data = df_4
              ) +
    scale_colour_manual(name = 'Week', 
         values =c("df_1" ="#122451", 
                   "df_2"="#62929a",
                  # "df_3"="#EC7E28",
                   "df_4"="#c969a8"
                   ), 
         labels = c('12',
                    '18',
                    #'19',
                    '33')) +
    labs(y = "Water Potential (MPa)", 
       x = "Leaf Water per Area (g)", 
       color= "Week", 
       shape = "Time"
       #caption = "triangles = predawn, circles = midday"
       ) +
  theme(
    #legend.position = c(.9, .35),
   # legend.position = "none",
     legend.position = "right",
  #strip.background = element_blank(),
  #strip.text.y = element_blank(), 
  #strip.text.x = element_text(size = 16), 
  axis.text.y = element_text(size = 12), 
  axis.text.x = element_text(size = 12), 
  #axis.ticks.y = element_blank(), 
  axis.title.y = element_text(size = 20), 
  axis.title.x = element_text(size = 20), 
 # axis.title.x = element_blank(),
  legend.title = element_text(size = 14), 
  legend.text = element_text(size = 14), 
# legend.position = "none"
  ) +
 #  xlim(.5,2) +
  scale_x_reverse())

nice_hydration_layered
```


Pull X trees, get mean slope for those X, do 1000x. 
```{r}
data %>% 
  #group_by(week) %>% 
  #summarise() %>% 
  count(as.factor(week)) %>% 
  mutate(trees = n/2)
```


```{r}
data <- data %>% 
  filter(!(week == 21)) #this week only has 1 pair of md/pd (missing scans, need to check on that) so remove it!

# Assuming your dataset is called 'data'
# Make sure 'data' is properly loaded before proceeding

# Define the number of bootstrap iterations
n_iterations <- 10

# Number of random trees to sample
n_sample_trees <- 6  # Adjust as needed

# Create an empty data frame to store bootstrapped slopes
bootstrapped_slopes_df <- data.frame()

# Function to calculate the slope for a specific tree within each iteration
calculate_slope_for_tree <- function(tree_data) {
  model <- lm(water_potential ~ lwa_g_cm2_new, data = tree_data)
  return(coef(model)[2])  # Extract the slope coefficient
}

#Do the bootstrapping: 

# Iterate through each unique week
unique_weeks <- unique(data$week)

for (x in unique_weeks) {
  # Filter data for the current week
  week_data <- data %>%
    filter(week == x)

 for (i in 1:n_iterations) {
   
# Randomly sample 'n_sample_trees' trees from your dataset
sampled_tree_ids <- sample(unique(week_data$tree), size = n_sample_trees, replace = TRUE)

# Iterate through each sampled tree and perform bootstrapping separately
      for (tree_id in sampled_tree_ids) {
        
        tree_data <- week_data %>%
          filter(tree == tree_id)
        
        # Initialize a container to store slopes for each iteration
        slopes <- vector("double", length = 1)
      
          # Calculate the slope for the bootstrapped sample
          slopes[] <- calculate_slope_for_tree(tree_data)
        
        # Store the bootstrapped slopes for this tree in the data frame
        bootstrapped_slopes_df <- bind_rows(
          bootstrapped_slopes_df,
          data.frame(
            tree = tree_id,
            slope = slopes, 
            round = i, 
            week = x
          ) 
        ) 
      }
  }

}

# round1 <- bootstrapped_slopes_df %>% 
#   filter(round == 1)
 summary_df_hyd <- bootstrapped_slopes_df %>%
        #filter(slope > -100, slope < 100) %>% 
        drop_na(slope) %>%
        group_by(week, round) %>% #get the mean of each round and week
       mutate(slope_round = mean(slope, na.rm = T)) %>% 
          ungroup() %>% 
        group_by(week) %>% 
          mutate(mean_slope = mean(slope_round, na.rm = T),
          lower_bound = quantile(slope_round, 0.025, na.rm = TRUE),
          upper_bound = quantile(slope_round, 0.975, na.rm = TRUE),
          week = week) %>% 
  select(week, mean_slope, lower_bound, upper_bound) %>% 
  distinct()

summary_df_hyd %>% 
  ggplot(aes(x = week)) +
   geom_segment(aes(y= lower_bound, 
                   yend = upper_bound, 
                   x = week, 
                   xend = week, 
                   color = as.factor(week)))+
  geom_point(aes(y = mean_slope), color = "cyan") +
  geom_point(aes(y = lower_bound), color = "pink") +
  geom_point(aes(y = upper_bound), color = "gold") +
  labs(y = "Slope",
       x = "Week")
```
#COMBINED DF: 

Should be week, mean_slope, lower_bound, upper_bound

```{r}
final_df_hyd <- summary_df_hyd %>% 
  mutate(analysis = "hydration, ind. trees", 
         time_season = "NA")

final_df_space <- final_df_space %>%
  mutate(analysis = "space",
         time_season = "NA")

time_boot_df <- time_boot_df %>% 
  mutate(week = 0,
         time_season = pre_post)

final_df_all <- bind_rows(final_df_hyd, final_df_space, time_boot_df) %>% 
  mutate(slope = mean_slope, 
         lwr = lower_bound, 
         upr = upper_bound) 

final_df_all %>% 
  write_csv(here::here("processed-data", "bootstrapped_df_qudo_lwa_g_cm2_new.csv"))
```

#-------

#*ANALYSIS*: 

```{r setup, include=FALSE}
knitr::opts_chunk$set(warning = FALSE, message = FALSE) 
#install.packages("")
#devtools::install_github("an-bui/calecopal")

library(data.table)
library(janitor)
library(here)
library(tidyverse)
library(lubridate)
library(readxl)
library(gridExtra)
library(calecopal)
library(rstatix)
library(vars)
library(MuMIn)
library(lme4) #for mixed effects models
library(arm)
library(nlme)
select = dplyr::select

source(here::here("scripts", "scripts_functions", "figure_info.R"))

#datver <- "20230724"
dataversion <- paste0("Data_", datver)
```

#Data wrangling: 

```{r, echo=F}
wc_qudo_df <- data_qudo_raw %>% 
  select(tree, week, water_potential, lwa_g_cm2_new, time_season, time, species, site) %>% 
  distinct()

wc_qudo_df %>% 
  group_by(week, time) %>% 
  summarise(n = n())

wc_lm_df <- wc_qudo_df %>% 
  filter(week > 17)
```

From what weeks do we have the most trees?
```{r, echo=F}
wc_qudo_df %>% 
  select(week, time, tree, lwa_g_cm2_new) %>% 
  drop_na(lwa_g_cm2_new) %>% 
  unique() %>% 
  group_by(week, time) %>% 
  count() %>% 
  ggplot(aes(x = week, y = n, fill = time)) +
  geom_col(stat = "identity") + 
  scale_x_continuous(breaks = seq(9, 40, by = 1))+
  labs(y = "lwa_g_cm2_new counts") +
  color_fill

wc_qudo_df %>% 
    select(week, time, tree, water_potential) %>% 
  drop_na(water_potential) %>% 
 # unique() %>% 
  group_by(week, time) %>% 
  count() %>% 
  ggplot(aes(x = week, y = n, fill = time)) +
  geom_col(stat = "identity") +
  scale_x_continuous(breaks = seq(9, 40, by = 1))+
  labs(y = "MPa counts") +
  color_fill
```
A: week 17, 21, 29, 10. 

IMPORTANT: NO MIDDAYS FOR WEEK 14!

#Timeline

```{r, fig.height=.75, fig.width=6}
wc_qudo_df %>% 
  filter(time == "md") %>% 
  drop_na(water_potential) %>% 
  #ggplot(aes(x = week, y = 0, color = mean(water_potential, na.rm = T)))+
  ggplot(aes(x = week, y = 0, color = water_potential))+
  geom_hline(yintercept = 0) +
  geom_point(size = 5) +
  geom_point(data= wc_qudo_df %>% 
               filter(week %in% c(10, 12, 14,15, 17, 18, 19, 21, 37)), 
           pch=21, 
           size=5, 
           colour="black") +
  color_grad_rev +
  ylim(-.25, +.25)  + 
  scale_x_continuous(breaks = c(10, 12, 14,15, 16,17, 18, 19, 21, 29, 33, 37)
                      # seq(10, 30, by = 1)
                     ) +
  theme(legend.position = "none", 
        axis.title.y = element_blank(), 
        axis.title.x = element_blank(), 
        axis.text.y = element_blank(),
        axis.ticks = element_blank(), 
        )
```


#Bootsrapped dataset: 
```{r}
final_df_all
```

#HYDRATION

MPa ~ lwa_g_cm2_new + (lwa_g_cm2_new|tree) + (1|wk)

look at it: 
```{r}
data <- data_og %>% 
  mutate(tree_week = paste(tree, week, sep = "_")
  ) %>% 
  mutate(water_potential = water_potential)
  

plotly::ggplotly(ggplot() +
    geom_point(aes(y = water_potential, 
                   x = lwa_g_cm2_new, 
                   shape = time, 
                   color = site, 
                   label = tree), 
               data = data) +
    geom_line(aes(y = water_potential, x = lwa_g_cm2_new, group = tree_week, color = site
                  ),
              data = data
              ) +
    labs(y = "Water Potential (MPa)", 
       x = "Leaf Water Content (g)", 
       color= "Week", 
       shape = "Time"
       ) +
  theme(
    #legend.position = c(.9, .35),
   # legend.position = "none",
  #strip.background = element_blank(),
  #strip.text.y = element_blank(), 
  #strip.text.x = element_text(size = 16), 
  # axis.text.y = element_text(size = 12), 
   axis.text.x = element_text(size = 12), 
   axis.text.y = element_text(size = 12),
    # axis.text.x  = element_blank(),
  #axis.ticks.y = element_blank(), 
  #axis.title.y = element_blank()
  axis.title.x = element_blank(),
  axis.title.y = element_blank(),
  legend.title = element_text(size = 14), 
  legend.text = element_text(size = 14)
# legend.position = "none"
  ) +
  facet_wrap(~week, scales = "free"))

```

###DATA
```{r, warning = F}
wk_all <- wc_qudo_df %>% 
  mutate(site = as.factor(site))

#Any instances where there are duplicates? 
wk_all_dupes <- wk_all %>% 
  #select(tree, week, time, water_potential) %>% 
  group_by(tree, week, time) %>% 
  filter(n() > 1)

#NO, yay!
```


```{r, warning = F}
#I think everything below here is unesessary? 
# wk_all_wide <- wk_all %>% 
#   #drop_na(date_wp) %>% 
#   pivot_wider(names_from=time, 
#               values_from = c(lwa_g_cm2_new, water_potential)) %>% 
#     dplyr::group_by(tree, week) %>%
#   fill(c(lwa_g_cm2_new_md, lwa_g_cm2_new_pd, water_potential_pd, water_potential_md), .direction = "downup") %>%
#   dplyr::ungroup()%>% 
#  # select(-date_wp) %>% 
#   distinct()
# 
# wk_all_long_lwa_g_cm2_new <- wk_all_wide %>% 
#   drop_na(lwa_g_cm2_new_pd, lwa_g_cm2_new_md, water_potential_md, water_potential_pd) %>% 
#   select(-water_potential_md, -water_potential_pd) %>% 
#   mutate(pd = lwa_g_cm2_new_pd, 
#          md = lwa_g_cm2_new_md) %>% 
#   pivot_longer(cols = c(pd, md), 
#                names_to = "time", 
#                values_to = "lwa_g_cm2_new") %>% 
#   select(-lwa_g_cm2_new_pd, -lwa_g_cm2_new_md)
# 
# wk_all_long_mpa <- wk_all_wide %>% 
#   drop_na(lwa_g_cm2_new_pd, lwa_g_cm2_new_md, water_potential_md, water_potential_pd) %>% 
#   select(-lwa_g_cm2_new_md, -lwa_g_cm2_new_pd) %>% 
#   mutate(pd = water_potential_pd, 
#          md = water_potential_md) %>% 
#   pivot_longer(cols = c(pd, md), 
#                names_to = "time", 
#                values_to = "water_potential") %>% 
#   select(-water_potential_pd, -water_potential_md)
# 
# wk_all_long <- merge(wk_all_long_mpa, wk_all_long_lwa_g_cm2_new) %>% 
#   distinct() %>% 
#  # select(-lwa_g_cm2_new_NA, -water_potential_NA) %>% 
#   group_by(tree, week, time, site
#            ) %>% 
#   summarise(lwa_g_cm2_new = mean(lwa_g_cm2_new, na.rm = T), 
#             water_potential = mean(water_potential, na.rm = T))
```

Slopes of the formula: 

```{r, warning=T}
m_wk_rand <- lmer(water_potential ~ lwa_g_cm2_new + (lwa_g_cm2_new|tree) + (lwa_g_cm2_new|week), data = wk_all )
```

Should we include site? 

```{r, include = F}
wk_all_sites <- wk_all %>% 
  drop_na(site)

m_wk_rand_nosite <- lmer(water_potential ~ lwa_g_cm2_new + (lwa_g_cm2_new|tree) + (lwa_g_cm2_new|week), data = wk_all_sites, REML = T)
#m_wk_rand_nosite

m_wk_rand_site <- lmer(water_potential ~ lwa_g_cm2_new + (lwa_g_cm2_new|tree) + (lwa_g_cm2_new|week) + (1|site), data = wk_all_sites, REML = T)
#m_wk_rand_site

anova(m_wk_rand_nosite, m_wk_rand_site) 

summary(m_wk_rand_nosite)

#rename so it goes with code below: 
m_wk_rand <- lmer(water_potential ~ lwa_g_cm2_new + (lwa_g_cm2_new|tree) + (lwa_g_cm2_new|week) + (1|site), data = wk_all_sites, REML = T)
#m_wk_rand_site

hyd_mod_df <- broom::tidy(m_wk_rand)
```

####Slopes, bootstrapped:

```{r}
hyd_slopes_bs_inds <- final_df_all %>% 
   filter(!week %in% c(9)) %>% 
  filter(analysis == "hydration, ind. trees") %>% 
  ggplot(aes(color= week)
    ) + 
  geom_point(aes(y = slope ,x = week), size =3) +
    geom_point(aes(y = upr, x = week), size = 1) +
  geom_point(aes(y = lwr, x = week), size = 1) +
  #geom_line(aes(group = week, y = week, x = upr)) +
  geom_segment(aes(y= lwr, 
                   yend = upr, 
                   x = week, 
                   xend = week, 
                   color = week))+
  labs(#x = "Week", 
       color = "week", 
       y = "Slope") +
   # facet_wrap(~y_var_name, scales = "free") + 
theme(
  strip.background = element_blank(),
  strip.text.y = element_blank(), 
  strip.text.x = element_text(size = 16), 
 # axis.text.y = element_blank(), 
  #axis.ticks.y = element_blank(), 
 axis.title.x = element_blank(),
  axis.title.y = element_text(size = 20), 
  legend.title = element_text(size = 18), 
 legend.position = "none"
  ) +
  geom_hline(yintercept = 0, linetype="dotted") + 
  # scale_x_continuous(breaks = c(10,17, 19, 21, 29, 37)
  #                     # seq(10, 30, by = 1)
  #                    ) +
  theme(legend.position = "none") +
  #xlim(0, 40) +
  color_grad 
hyd_slopes_bs_inds

ggsave(here("figures", "qudo figures", "LWA", "nice_hydration_slopes_bs_inds"), hyd_slopes_bs_inds, device = "jpg", width = 5, height = 2.5, dpi = 300)
```


#SPACE:

1. Center and scale by week
2. Plot all midday grand slopes

###Midday: 

```{r, warning = F}
wc_wk_centered_df <- wk_all %>% 
  filter(!week %in% c(9)) %>% 
  #filter(lwa_g_cm2_new > .44) %>% 
  mutate(water_potential = water_potential) %>% 
  filter(time == "md") %>% 
  group_by(week) %>% 
  mutate(mean_lwa_g_cm2_new = mean(lwa_g_cm2_new, na.rm = T)) %>% 
  ungroup() %>% 
  mutate(lwa_g_cm2_new_centered = lwa_g_cm2_new - mean_lwa_g_cm2_new)
  # mutate(lwa_g_cm2_new_centered = scale(lwa_g_cm2_new,  scale=FALSE)) %>% 
  # filter(lwa_g_cm2_new_centered < 4)
```


```{r, warning = F, echo=FALSE}
#scaled
space_nice <- wc_wk_centered_df %>% 
  #drop_na(site) %>% 
  ggplot(aes(y = water_potential, 
            # x = lwa_g_cm2_new_centered, 
            x = lwa_g_cm2_new,
             #color = tree_factor
             color = as.factor(week)
             )) +
  geom_point(alpha = .5, 
            # aes (shape = site)
             ) +
  theme(legend.position="none",
       # legend.position= c(.7, .2),
      strip.background = element_blank(),
      strip.text.x = element_text(size = 12),
      plot.title = element_text(size=13),
     # axis.title = element_text(size = 16),
     axis.text = element_text(size = 12),
     legend.key=element_blank(), 
      axis.title.x = element_blank(),
     axis.title.y = element_text(size = 20), 
     legend.text = element_text(size = 13), 
     legend.title = element_text(size = 16),
     legend.margin=margin(0,0,0,0),
      legend.box.margin=margin(-5,-8,-8,-8)
    ) +
  geom_smooth(method = "lm", se = F) + 
 # geom_vline(xintercept = 0, linetype = "dotted") +
  labs(y= "Middays (MPa)", 
       x = "Leaf Water per Area", 
       color = "Week", 
       shape = "Site") +
  color_many_2 +
  #color_week_all +
  guides(color = guide_legend(nrow = 2))
space_nice

ggsave(here("figures", "qudo figures", "LWA", "nice_space"), space_nice, device = "jpg", width = 6, height = 4, dpi = 300)
```

```{r, warning = F, echo=FALSE}
#scaled
space_nice_centered <- wc_wk_centered_df %>% 
  #drop_na(site) %>% 
  ggplot(aes(y = water_potential, 
             x = lwa_g_cm2_new_centered, 
            #x = lwa_g_cm2_new,
             #color = tree_factor
             color = as.factor(week)
             )) +
  geom_point(alpha = .5, 
            # aes (shape = site)
             ) +
  theme(legend.position="top",
        #legend.position= c(.7, .2),
      strip.background = element_blank(),
      strip.text.x = element_text(size = 12),
      plot.title = element_text(size=13),
      axis.title = element_text(size = 16),
     axis.text = element_text(size = 12),
     legend.key=element_blank(), 
     legend.text = element_text(size = 13), 
     legend.title = element_text(size = 16),
     legend.margin=margin(0,0,0,0),
      legend.box.margin=margin(-5,-8,-8,-8)
    ) +
  geom_smooth(method = "lm", se = F) + 
 # geom_vline(xintercept = 0, linetype = "dotted") +
  labs(y= "Middays (MPa)", 
       x = "Centered Leaf Water per Area", 
       color = "Week", 
       shape = "Site") +
  color_many_2 +
  #color_week_all +
  guides(color = guide_legend(nrow = 1))

space_nice_centered

ggsave(here("figures", "qudo figures", "LWA", "nice_space_centered"), space_nice_centered, device = "jpg", width = 6, height = 4, dpi = 300)
```

###Models: 
Do we need to include site here? 

Centered lwa_g_cm2_new:

```{r}
space_df <- wc_wk_centered_df %>% 
  drop_na()

space_mod <- lmer(water_potential ~ lwa_g_cm2_new_centered + (lwa_g_cm2_new_centered|week) + (1|tree), data = space_df, REML = T)
#summary(space_mod)

space_mod_site <- lmer(water_potential ~ lwa_g_cm2_new_centered  + (lwa_g_cm2_new_centered|week) + (1|tree) + (1|site), data = space_df,  REML = T)
#summary(space_mod_site)

anova(space_mod, space_mod_site)
```

YES

lwa_g_cm2_new uncentered:

```{r}
space_df_nosite <- wc_wk_centered_df %>% 
  drop_na(site)

space_mod <- lmer(water_potential ~ lwa_g_cm2_new + (lwa_g_cm2_new|week) + (1|tree), data = space_df_nosite, REML = T)
#summary(space_mod)

space_mod_site <- lmer(water_potential ~ lwa_g_cm2_new  + (lwa_g_cm2_new|week) + (1|tree) + (1|site), data = space_df_nosite,  REML = T)
#summary(space_mod_site)

anova(space_mod, space_mod_site)

space_mod <- lmer(water_potential ~ lwa_g_cm2_new_centered + (lwa_g_cm2_new_centered|week) + (1|tree) + (1|site), data =  wc_wk_centered_df, REML = T)

space_mod_df <- broom.mixed::tidy(space_mod)
```
YES

###Slopes, bootsrapped: 

```{r}
space_slopes <- final_df_all %>% 
   filter(!week %in% c(9)) %>% 
  filter(analysis == "space") %>% 
   mutate(week = as.factor(week)) %>% 
  ggplot(aes(color= week)
    ) + 
  geom_point(aes(y = slope ,x = week), size =2) +
    geom_point(aes(y = upr, x = week), size = 1) +
  geom_point(aes(y = lwr, x = week), size = 1) +
  #geom_line(aes(group = week, y = week, x = upr)) +
  geom_segment(aes(y= lwr, 
                   yend = upr, 
                   x = week, 
                   xend = week, 
                   color = week))+
  labs(x = "Week", 
       color = "Week", 
       y = "Slope - space") +
  theme(legend.position="none",
      strip.background = element_blank(),
      strip.text.x = element_text(size = 12),
      plot.title = element_text(size=13),
      axis.title.y = element_text(size = 16),
      axis.title.x = element_blank(),
     axis.text = element_text(size = 12),
     legend.key=element_blank(), 
     legend.text = element_text(size = 13), 
     legend.title = element_text(size = 16),
     legend.margin=margin(0,0,0,0),
      legend.box.margin=margin(-5,-8,-8,-8)
    ) +
  geom_hline(yintercept = 0, linetype="dotted") + 
    color_many_2 +
  guides(color = guide_legend(nrow = 2))
space_slopes
```


#TIME

```{r}
wc_qudo_df %>% 
  #filter(!week %in% c(9, 18, 19)) %>% 
  drop_na(week) %>% 
  ggplot(aes(y = lwa_g_cm2_new, 
             x = as.factor(week), 
             #x = date,
            # color = as.factor(week), 
             fill = time)) +
  geom_boxplot() +
 # facet_wrap(~time) +
  color_many +
  color_fill +
  labs(y = "lwa_g_cm2_new", 
       x = "Week", 
       fill = "Time")
```

```{r, echo=FALSE, warning = F}
time <- wc_qudo_df %>% 
  drop_na(week) %>% 
 # filter(time == "md") %>% 
  ggplot(aes(y = water_potential, 
             x = lwa_g_cm2_new, 
             color = time_season, 
             size = week,
             shape = as.factor(tree),
             show.legend = FALSE)) +
  geom_point(size = 1,
             alpha = .3)+
  geom_smooth(method = "lm", se = F) +
 # theme(legend.position = "none") +
  labs(x = "lwa_g_cm2_new", 
       y = "Midday MPa") +
  #color_two_grey +
  theme(
      legend.position="none",
      strip.background = element_blank(),
      strip.text.x = element_text(size = 12),
      plot.title = element_text(size=13),
      axis.title = element_text(size = 16),
     axis.text = element_text(size = 12),
     legend.key=element_blank(), 
     legend.text = element_text(size = 13), 
     legend.title = element_text(size = 16),
     legend.margin=margin(0,0,0,0),
      legend.box.margin=margin(-5,-8,-8,-8)
    ) +
  geom_smooth(method = "lm", se = F) + 
 # geom_vline(xintercept = 0, linetype = "dotted") +
  labs(y= "Middays (MPa)", 
       x = "Leaf Water per Area (g)", 
       color = "Week", 
       ) +
  guides(color = guide_legend(nrow = 2)) +
  facet_wrap(~time, nrow = 2) +
  color_two_grey
time 

ggsave(here("figures", "qudo figures", "LWA", "nice_time"), time, device = "jpg", width = 4, height = 2, dpi = 300)
```



```{r, echo=FALSE}
wc_qudo_df %>% 
  filter(time == "pd") %>% 
  ggplot(aes(y = water_potential, 
             x = lwa_g_cm2_new, 
             color = as.factor(tree), 
             show.legend = FALSE)) +
  geom_point(aes(size = week), 
             alpha = .2)+
  geom_smooth(method = "lm", se = F) +
 # theme(legend.position = "none") +
  labs(x = "lwa_g_cm2_new", 
       y = "Predawn MPa") +
  color_very_many +
  theme(legend.position = "none")
```

###Models: 
####All time

```{r}
wc_qudo_md_df <- wc_qudo_df %>% 
  filter(time == "md")
  
time_mod <- lmer(water_potential ~ lwa_g_cm2_new + (1|week) + (lwa_g_cm2_new|tree), data = wc_qudo_md_df)
#time_mod <- lmer(water_potential ~ lwa_g_cm2_new + (lwa_g_cm2_new|tree), data = wc_qudo_md_df)
 
#with separate periods: 
time_season_mod <- lmer(water_potential ~ lwa_g_cm2_new + time_season + (1|week) + (lwa_g_cm2_new|tree), data = wc_qudo_md_df)

time_season_mod_int <- lmer(water_potential ~ lwa_g_cm2_new*time_season + (1|week) + (lwa_g_cm2_new|tree), data = wc_qudo_md_df)

time_season_mod_df <- broom::tidy(time_season_mod)
```


####Just linear: 
```{r}
lin_wc_qudo_md_df <- wc_qudo_df %>% 
  filter(time == "md") %>% 
  filter(time_season == "after leafout")
  
  
lin_time_mod <- lmer(water_potential ~ lwa_g_cm2_new + (1|week) + (lwa_g_cm2_new|tree), data = lin_wc_qudo_md_df)

summary(lin_time_mod)

MuMIn::r.squaredGLMM(lin_time_mod)
```


####Before leafout: 

```{r}
bef_wc_qudo_md_df <- wc_qudo_df %>% 
  filter(time == "md") %>% 
  #filter(time_season == "before leafout") %>% 
  filter(week <= 17) %>% 
  drop_na(site)
  
  
bef_time_mod <- lmer(water_potential ~ lwa_g_cm2_new + (1|week) + (lwa_g_cm2_new|tree), data = bef_wc_qudo_md_df, REML = T)
summary(bef_time_mod)
MuMIn::r.squaredGLMM(bef_time_mod)
```


##Slopes, bootstrapped:
 
```{r}
time_summary_bs_tree <- final_df_all %>% 
   filter(!week %in% c(9)) %>% 
  filter(analysis == "time, ind. tree") %>% 
   mutate(week = as.factor(week)) %>% 
ggplot() + 
  geom_point(aes(y = slope, x = time_season), size =3) +
    geom_point(aes(y = upr, x = time_season), size = 1) +
  geom_point(aes(y = lwr, x = time_season), size = 1) +
  #geom_line(aes(group = tree, y = tree, x = upr)) +
  geom_segment(aes(y= lwr, 
                   yend = upr, 
                   x = time_season, 
                   xend = time_season)) +
  labs(y = "Bootstrapped slope and CI", 
       x = "Time")
time_summary_bs_tree


ggsave(here("figures", "qudo figures", "LWA", "time_summary_slopes_boot_tree"), time_summary_bs_tree, device = "jpg", width = 2.5, height = 3, dpi = 300)
```

#MODELS:

glanced mods

```{r}
glance_df <- bind_rows(space_glance <- broom.mixed::glance(space_mod) 
                       %>% mutate(mod = "space"),
                       hyd_glance <-  broom.mixed::glance(m_wk_rand) 
                       %>% mutate(mod = "hyd"),
                       time_glance <- broom.mixed::glance(time_season_mod_int)
                       %>% mutate(mod = "time"),
                       time_glance <- broom.mixed::glance(time_season_mod)
                       %>% mutate(mod = "interaction"),
                       as_data_frame(r.squaredGLMM(space_mod)) 
                       %>% mutate(mod = "space"),
                       as_data_frame(r.squaredGLMM(m_wk_rand)) 
                       %>% mutate(mod = "hyd"),
                       as_data_frame(r.squaredGLMM(time_season_mod)) 
                       %>% mutate(mod = "time"),
                       as_data_frame(r.squaredGLMM(time_season_mod_int)) 
                       %>% mutate(mod = "interaction")
                       ) %>% 
  group_by(mod) %>% 
  fill(c(R2m, R2c), .direction = "downup") %>% 
  distinct() %>% 
  drop_na(df.residual)
glance_df
```

Tidy mods
```{r}
lwa_g_cm2_new_mods <- bind_rows(space_mod_df <- space_mod_df %>% 
                        mutate(mod = "space"), 
                      hyd_mod_df <- hyd_mod_df %>% 
                        mutate(mod = "hyd"),
                      time_season_mod_df <- 
                        hyd_mod_df %>% mutate(mod= "time")) %>% 
  bind_rows(glance_df) %>% 
    mutate(var = "lwa_g_cm2_new")

lwa_g_cm2_new_mods

write_csv(lwa_g_cm2_new_mods, here::here("processed-data", "lwa_g_cm2_new_qudo_mods.csv"))
```

#-------

#*FIGURES*

```{r}
library(cowplot)
library(grid)
library(gridExtra)
```

#Data: 

#Scatterplots:

#####Hydration

```{r}
wk_figs <-data_og %>% 
  group_by(tree, week) %>% 
  filter(n() == 2) %>% 
  filter(week %in% c(10, 12, 14, 15, 17, 19, 21, 29, 33, 37), 
         #lwa_g_cm2_new < 6, 
        # lwa_g_cm2_new > 0
         ) %>% 
  mutate(week_nice = case_when(
    week == 10 ~ "Week 10", 
    week == 12 ~ "Week 12", 
    week == 14 ~ "Week 14", 
    week == 15 ~ "Week 15", 
    week == 17 ~ "Week 17", 
    week == 19 ~ "Week 19", 
    week == 21 ~ "Week 21", 
    week == 29 ~ "Week 29",
    week == 33 ~ "Week 33",
    week == 37 ~ "Week 37"
  )) 

wk_figs_labels <-
  data.frame(
    label = c('10', '12', '14', '15', '17', '19', '21', '29', '33', '33'),
    week_nice = c(
      "Week 10",
      "Week 12",
       "Week 14",
      "Week 15",
      "Week 17",
      "Week 19",
      "Week 21",
      "Week 29",
      "Week 33",
      "Week 37"
    )
  )

df_all_groups <-  wk_figs %>% 
  mutate(tree_week = paste(tree, week, sep = "_")
  ) %>% 
  mutate(water_potential = water_potential)
  
df_grey <- df_all_groups

df_1 <- df_all_groups %>% 
  filter(week %in% c('10'))

df_2 <- df_all_groups %>% 
  filter(week %in% c('14'))

df_3 <- df_all_groups %>% 
  filter(week %in% c('17'))

df_4 <- df_all_groups %>% 
  filter(week %in% c('37'))

nice_hydration_layered_1 <- ggplot() +
    geom_point(aes(y = water_potential, x = lwa_g_cm2_new, shape = time), data = df_grey, color = "grey") +
    geom_line(aes(y = water_potential, x = lwa_g_cm2_new, group = tree_week), color = "grey", data = df_grey) +
    geom_point(aes(y = water_potential, x = lwa_g_cm2_new, shape = time, color = "df_1"), 
               data = df_1 ) +
    geom_line(aes(y = water_potential, x = lwa_g_cm2_new, group = tree_week, color = "df_1"), 
              data = df_1) +
    geom_point(aes(y = water_potential, x = lwa_g_cm2_new, shape = time, color = "df_2"), 
               data = df_2) +
    geom_line(aes(y = water_potential, x = lwa_g_cm2_new, group = tree_week, color = "df_2"), 
              data = df_2) +
    geom_point(aes(y = water_potential, x = lwa_g_cm2_new, shape = time, color = "df_4"), 
               data = df_4) +
    geom_line(aes(y = water_potential, x = lwa_g_cm2_new, group = tree_week, color = "df_4"), 
              data = df_4
              ) +
     geom_point(aes(y = water_potential, x = lwa_g_cm2_new, shape = time, color = "df_3"), 
               data = df_3) +
    geom_line(aes(y = water_potential, x = lwa_g_cm2_new, group = tree_week, color = "df_3"),
              data = df_3
              ) +
    scale_colour_manual(name = 'Week', 
        values =c("df_1" ="#48578E", 
                   "df_2"="#528791",
                   "df_3"="#8E9D30",
                   "df_4"="#CC5265"
                   ), 
         labels = c('10',
                    '14',
                    '17',
                    '37')) +
    labs(y = "Water Potential (MPa)", 
         #y = " ",
       x = "Leaf Water per Area (g)", 
       color= "Week", 
       shape = "Time"
       #caption = "triangles = predawn, circles = midday"
       ) +
  theme(
    #legend.position = c(.9, .35),
    legend.position = "none",
  #strip.background = element_blank(),
  #strip.text.y = element_blank(), 
  #strip.text.x = element_text(size = 16), 
  # axis.text.y = element_text(size = 12), 
   axis.text.x = element_text(size = 12), 
   axis.text.y = element_text(size = 12),
    # axis.text.x  = element_blank(),
  #axis.ticks.y = element_blank(), 
  #axis.title.y = element_blank()
  axis.title.x = element_blank(),
  axis.title.y = element_blank(),
  legend.title = element_text(size = 14), 
  legend.text = element_text(size = 14), 
  plot.margin = unit(c(5.5, 5.5, 0, 5.5), units = "pt")
# legend.position = "none"
  ) +
   #ylim(-5,0)+
   #xlim(.5,2) +
  #scale_x_reverse() +
scale_y_continuous(labels = scales::number_format(accuracy = 0.1))
nice_hydration_layered_1
```
#####Space

```{r}
#scaled
space_nice_1 <- wc_wk_centered_df %>% 
  drop_na(site) %>% 
  ggplot(aes(y = water_potential, 
            x = lwa_g_cm2_new,
             color = as.factor(week)
             )) +
  geom_point(alpha = .5, 
            # aes (shape = site)
             ) +
  theme(legend.position="none",
       # legend.position= c(.7, .2),
      strip.background = element_blank(),
      strip.text.x = element_text(size = 12),
      plot.title = element_text(size=13),
     # axis.title = element_text(size = 16),
     axis.text.y = element_text(size = 12),
     axis.text.x = element_text(size = 12),
    # axis.text.x  = element_blank(),
     legend.key=element_blank(), 
      axis.title.x = element_blank(),
     axis.title.y = element_blank(),
     legend.text = element_text(size = 13), 
     legend.title = element_text(size = 16),
     legend.margin=margin(0,0,0,0),
      legend.box.margin=margin(-5,-8,-8,-8),
    plot.margin = unit(c(5.5, 5.5, 0, 5.5), units = "pt")
    ) +
  geom_smooth(method = "lm", se = F,size = .5) + 
 # geom_vline(xintercept = 0, linetype = "dotted") +
  labs(#y= "Middays (MPa)", 
       y = " ",
       x = "Leaf Water per Area", 
       color = "Week", 
       shape = "Site") +
  #color_many_2 +
  color_week_assigned +
 # xlim(.5,2) +
 # ylim(-5,0)+
  guides(color = guide_legend(nrow = 2)) +
  #scale_x_reverse() +
scale_y_continuous(labels = scales::number_format(accuracy = 0.1))
space_nice_1


```
#####Time:

```{r}
time_1 <- wc_qudo_df %>% 
  mutate(time = case_when(
    time == "md" ~ "Midday", 
    time == "pd" ~ "Predawn"
  )) %>% 
  filter(time == "Midday") %>% 
  drop_na(week) %>% 
 # filter(time == "md") %>% 
  ggplot(aes(y = water_potential, 
             x = lwa_g_cm2_new, 
             color = time_season, 
             size = week,
             shape = as.factor(tree),
             show.legend = FALSE)) +
  geom_point(size = 1,
             alpha = .1)+
  geom_smooth(method = "lm", se = F, size = .5, alpha = .4) +
 # theme(legend.position = "none") +
  labs(x = "lwa_g_cm2_new", 
       y = "Midday MPa") +
  #color_two_grey +
  theme(
      legend.position="none",
      strip.background = element_blank(),
     # strip.text.x = element_text(size = 12),
      strip.text.x = element_text(size = 12),
      plot.title = element_text(size=13),
      #axis.title = element_text(size = 15),
     axis.title = element_blank(),
     axis.text = element_text(size = 12),
     axis.text.x = element_text(size = 12),
     legend.key=element_blank(),
     legend.text = element_text(size = 13),
     legend.title = element_text(size = 16),
     #legend.title = element_blank(),
     legend.margin=margin(0,0,0,0),
      legend.box.margin=margin(-5,-8,-8,-8),
     plot.margin = unit(c(5.5, 5.5, 0, 5.5), units = "pt")
    ) +
  labs(#y= "Water Potential (MPa)", 
       y = "Water Potential (MPa)",
       x = "Leaf Water per Area (g)", 
       color = "Week", 
       ) +
  guides(color = guide_legend(nrow = 2)) +
 # facet_wrap(~time, nrow = 2) +
  # xlim(.5,2) +
 # ylim(-5,0)+
  #scale_x_reverse()+
  color_two_grey +
scale_y_continuous(labels = scales::number_format(accuracy = 0.1))

time_1 
```


#Slopes: 

####Hydration

Bootstrapped: 

```{r}
hyd_slopes <- final_df_all %>% 
   #filter(!week %in% c(9, 19)) %>% 
  filter(analysis == "hydration, ind. trees") %>% 
  ggplot(aes(color= as.factor(week))
    ) + 
  geom_point(aes(y = slope ,x = as.factor(week)), size =2) +
  geom_point(aes(y = upr, x = as.factor(week)), size = 1) +
  geom_point(aes(y = lwr, x = as.factor(week)), size = 1) +
  #geom_line(aes(group = week, y = week, x = upr)) +
  geom_segment(aes(y= lwr, 
                   yend = upr, 
                   x = as.factor(week), 
                   xend = as.factor(week))) +
  labs(x = "Week", 
       color = "week", 
       y = "Slope") +
  theme(legend.position="none",
      strip.background = element_blank(),
      strip.text.x = element_text(size = 12),
      plot.title = element_text(size=13),
      axis.title = element_blank(),
     axis.text = element_text(size = 12),
     legend.key=element_blank(), 
     legend.text = element_text(size = 13), 
     legend.title = element_blank(),
     legend.margin=margin(0,0,0,0),
      legend.box.margin=margin(-5,-8,-8,-8),
     plot.margin = unit(c(5.5, 5.5, 0, 5.5), units = "pt")
    ) +
  geom_hline(yintercept = 0, linetype="dotted") + 
   #scale_x_continuous(breaks = c(10, 12, 14,15, 17, 19, 21, 29, 33, 37)) +
  theme(legend.position = "none") +
  color_week_assigned
hyd_slopes
```

####Space

Bootstrapped: 
```{r}
space_slope_bs <-  final_df_all %>% 
   #filter(!week %in% c(9, 18)) %>% 
  filter(analysis == "space") %>% 
 # mutate(week = as.numeric(week)) %>% 
   ggplot(aes(color= as.factor(week))
    ) + 
  geom_point(aes(y = slope ,x = as.factor(week)), size =2) +
    geom_point(aes(y = upr, x = as.factor(week)), size = 1) +
  geom_point(aes(y = lwr, x = as.factor(week)), size = 1) +
  #geom_line(aes(group = week, y = week, x = upr)) +
  geom_segment(aes(y= lwr, 
                   yend = upr, 
                   x = as.factor(week), 
                   xend = as.factor(week), 
                   color = as.factor(week)))+
  labs(x = "Week", 
       color = "Week", 
       y = " ") +
  theme(legend.position="none",
      strip.background = element_blank(),
      strip.text.x = element_text(size = 12),
      plot.title = element_text(size=13),
      axis.title.y = element_text(size = 15),
      axis.title.x = element_blank(),
     axis.text = element_text(size = 12),
     legend.key=element_blank(), 
     legend.text = element_text(size = 13), 
    # legend.title = element_text(size = 16),
    legend.title = element_blank(),
     legend.margin=margin(0,0,0,0),
      legend.box.margin=margin(-5,-8,-8,-8),
    plot.margin = unit(c(5.5, 5.5, 0, 5.5), units = "pt")
    ) +
  #scale_x_continuous(breaks = c(10, 12, 14,15, 17, 19, 21, 29, 33, 37)
                  #   breaks = c(37, 33, 29, 21, 19, 17, 15, 14, 12, 10)
                      # seq(10, 30, by = 1)
  #                   ) +
  geom_hline(yintercept = 0, linetype="dotted") + 
  color_week_assigned+
 # color_many_2 +
  guides(color = guide_legend(nrow = 2)) #+
  #ylim(-8, 35)
 # scale_x_reverse()
space_slope_bs
```


####Time

BS:

```{r}
time_summary <- final_df_all %>% 
   filter(!week %in% c(9)) %>% 
  filter(analysis == "time, ind. tree") %>% 
   mutate(week = as.factor(week), 
          time_season = case_when(
            time_season %in% c("before leafout") ~ "During", 
             time_season %in% c("after leafout") ~ "After", 
            
          )) %>% 
ggplot(aes(color = time_season)) + 
  geom_point(aes(y = slope, x = time_season), size =3) +
  geom_point(aes(y = upr, x = time_season), size = 1) +
  geom_point(aes(y = lwr, x = time_season), size = 1) +
  geom_segment(aes(y= lwr, 
                   yend = upr, 
                   x = time_season, 
                   xend = time_season)) +
  labs(y = "",
       x = "Time") +
  color_two_grey +
  geom_hline(yintercept = 0, linetype="dotted") + 
  theme(legend.position="none",
      strip.background = element_blank(),
      strip.text.x = element_text(size = 12),
      plot.title = element_text(size=13),
      axis.title.y = element_text(size = 15),
      #axis.title.x = element_text(size = 13),
      axis.title.x = element_blank(),
     axis.text = element_text(size = 12),
     legend.key=element_blank(), 
     legend.text = element_text(size = 13), 
     legend.title = element_text(size = 16),
     legend.margin=margin(0,0,0,0),
      legend.box.margin=margin(-5,-8,-8,-8),
     plot.margin = unit(c(5.5, 5.5, 0, 5.5), units = "pt")
    ) +
  scale_x_discrete(limits = rev)

time_summary
```

##PV curves

```{r, fig.width=2, fig.height=2}
mean_slope_all <- read_csv(here::here("processed-data", "pv_lwa_slopes.csv"))

pv_boot_df <- mean_slope_all %>% 
  select(pre_post, mean_slope, lower_bound, upper_bound) %>% 
  distinct() %>% 
  mutate(analysis = "pv curves") %>% 
  mutate(slope = mean_slope, 
         upr = as.numeric(upper_bound), 
         lwr = as.numeric(lower_bound))


pv_slope <-  pv_boot_df %>% 
   #filter(!pre_post %in% c(9)) %>% 
  #filter(analysis == "space") %>% 
 # mutate(pre_post = as.numeric(pre_post)) %>% 
   ggplot(aes(color= as.factor(pre_post))
    ) + 
  geom_point(aes(y = slope ,x = as.factor(pre_post)), size =2) +
    geom_point(aes(y = upr, x = as.factor(pre_post)), size = 1) +
  geom_point(aes(y = lwr, x = as.factor(pre_post)), size = 1) +
  #geom_line(aes(group = pre_post, y = pre_post, x = upr)) +
  geom_segment(aes(y= lwr, 
                   yend = upr, 
                   x = as.factor(pre_post), 
                   xend = as.factor(pre_post), 
                   color = as.factor(pre_post)))+
  labs(x = "PV", 
       color = "pre_post", 
      # y = "Slope"
       ) +
  theme(legend.position="none",
      strip.background = element_blank(),
      strip.text.x = element_text(size = 12),
      axis.text.x = element_blank(),
      plot.title = element_text(size=13),
     # axis.title.y = element_text(size = 15),
      axis.title.y = element_blank(),
    #  axis.title.x = element_text(size = 15),
     # axis.title.x = element_blank(),
     axis.text = element_text(size = 12),
     legend.key=element_blank(), 
     legend.text = element_text(size = 13), 
    # legend.title = element_text(size = 16),
    legend.title = element_blank(),
     legend.margin=margin(0,0,0,0),
      legend.box.margin=margin(-5,-8,-8,-8),
    plot.margin = unit(c(5.5, 5.5, 0, 5.5), units = "pt")
    ) +
  geom_hline(yintercept = 0, linetype="dotted") +
 # color_pre_post_assigned+
 # color_many_2 +
  guides(color = guide_legend(nrow = 2)) +
 scale_color_manual(values = c("#525252")) 
 # scale_x_reverse()
pv_slope

ggsave(here("figures", "qudo figures", "LWA", "pv_slope_plot_qudo"), pv_slope, device = "jpg", width = 1, height = 3, dpi = 300)
```


```{r, fig.width=2, fig.height=2}
empty_plot <-  pv_boot_df %>% 
   filter(pre_post = NA) %>% #we want an empty plot
  #filter(analysis == "space") %>% 
 # mutate(pre_post = as.numeric(pre_post)) %>% 
   ggplot(aes(color= as.factor(pre_post))
    ) + 
  geom_point(aes(y = slope ,x = as.factor(pre_post)), size =2) +
    geom_point(aes(y = upr, x = as.factor(pre_post)), size = 1) +
  geom_point(aes(y = lwr, x = as.factor(pre_post)), size = 1) +
  #geom_line(aes(group = pre_post, y = pre_post, x = upr)) +
  geom_segment(aes(y= lwr, 
                   yend = upr, 
                   x = as.factor(pre_post), 
                   xend = as.factor(pre_post), 
                   color = as.factor(pre_post)))+
  labs(x = "PV", 
       color = "pre_post", 
      # y = "Slope"
       ) +
  theme(legend.position="none",
      strip.background = element_blank(),
      strip.text.x = element_text(size = 12),
      axis.text.x = element_blank(),
      plot.title = element_text(size=13),
     # axis.title.y = element_text(size = 15),
      axis.title.y = element_blank(),
      axis.title.x = element_text(size = 15),
     # axis.title.x = element_blank(),
     axis.text = element_text(size = 12),
     legend.key=element_blank(), 
     legend.text = element_text(size = 13), 
    # legend.title = element_text(size = 16),
    legend.title = element_blank(),
     legend.margin=margin(0,0,0,0),
      legend.box.margin=margin(-5,-8,-8,-8)
    ) +
  geom_hline(yintercept = 0, linetype="dotted") +
 # color_pre_post_assigned+
 # color_many_2 +
  guides(color = guide_legend(nrow = 2)) +
 scale_color_manual(values = c("darkgrey", "darkgrey")) 
 # scale_x_reverse()
empty_plot
```

######Legend:

```{r}
legend_plot <- wk_figs %>% 
  drop_na(site) %>% 
  mutate(time = case_when(
    time %in% c("md") ~ "Midday",
    time %in% c("pd") ~ "Predawn"
  )) %>% 
  ggplot(aes(y = water_potential,
             x = lwa_g_cm2_new,
             color = as.factor(week), 
             shape = as.factor(time),
             )) +
  geom_point(alpha = .5) +
  theme(legend.position="top",
       # legend.position= c(.7, .2),
      strip.background = element_blank(),
      strip.text.x = element_text(size = 12),
      plot.title = element_text(size=13),
     # axis.title = element_text(size = 16),
     axis.text = element_text(size = 12),
     legend.key=element_blank(), 
     axis.title.x = element_blank(),
     axis.title.y = element_blank(),
     legend.text = element_text(size = 13), 
     legend.title = element_text(size = 16),
     legend.margin=margin(0,0,0,0),
    legend.box.margin=margin(-5,-8,-8,-8)
    ) +
  geom_smooth(method = "lm", se = F, size = .5) + 
  labs(color = "Week", 
       shape = "Time") +
  color_week_assigned +
 # ylim(0,2) +
  guides(color = guide_legend(nrow = 1))
legend_plot

legend_scatterplots <- cowplot::get_legend(legend_plot)
legend_scatterplots
```

#COMBINE

```{r}
library(grid)
#labels of slopes: 
y.grob1 <- textGrob("Slope", 
                   gp=gpar(fontface="bold", #col="blue", 
                           fontsize=15), rot=90)

x.grob1 <- textGrob("Time", 
                   gp=gpar(fontface="bold", #col="blue", 
                           fontsize=15))

combined_slopes_wide <- cowplot::plot_grid(space_slope_bs, NULL, time_summary, NULL, hyd_slopes, NULL, pv_slope,
                                   nrow = 1, 
                                  rel_widths = c(1, -.021, 1, -.02, 1, -.02, .5), 
                                  labels = c("D", "", "E", "", "F", "", "G"),
                                  label_x = .2,
                                  label_y = .98
                                  # rel_heights = c(.15, 1, .3), 
                                   #rel_widths = c(1, 1, 1), 
                                  # rel_widths = c(.7, .8, .7, .3)
                                   ) 

combined_slopes_wide

combined_slopes_wide_labels <- grid.arrange(arrangeGrob(combined_slopes_wide, 
                            left = y.grob1, 
                            bottom = x.grob1),
                            heights = c(1, .05))

#labels of scatterplots: 
y.grob2 <- textGrob("Water Potential (MPa)", 
                   gp=gpar(fontface="bold", #col="blue", 
                           fontsize=15), rot=90)

x.grob2 <- textGrob("Leaf Water per Area (g/cm2)", 
                   gp=gpar(fontface="bold", #col="blue",
                           fontsize=15))


combined_plots_wide_nolegend <- cowplot::plot_grid(space_nice_1, NULL, 
                                      time_1, NULL, 
                                   nice_hydration_layered_1,
                                   labels=c("A","", "B", "", "C"),
                                   label_x = .112,
                                   label_y = .98,
                                   nrow = 1,
                                   rel_widths = c(1, -.015, 1, -.015, 1)
                                   #rel_widths = c(1, 1, 1)
                                   ) 
combined_plots_wide_nolegend

combined_plots_wide <- cowplot::plot_grid(legend_scatterplots,
                                      combined_plots_wide_nolegend,
                                          nrow = 2, 
                                          rel_heights = c(.1, 1))

combined_plots_wide_labels <- grid.arrange(arrangeGrob(combined_plots_wide, 
                                left = y.grob2, 
                                bottom = x.grob2,
                                heights = c(1, 0)))

combined_plots_all <- cowplot::plot_grid(
  combined_plots_wide_labels, 
  combined_slopes_wide_labels, 
          nrow = 2, 
          rel_heights = c(.75, .75))

combined_plots_all

ggsave(here("figures", "qudo figures", "LWA", "combined_plots_all_qudo"), combined_plots_all, device = "jpg", width = 10, height = 9, dpi = 300)
```
#-----------
#*FIGURES NEW*

```{r}
library(cowplot)
library(grid)
library(gridExtra)
```

#Data: 

#Scatterplots:

#####Hydration

```{r}
wk_figs <-data_og %>% 
  group_by(tree, week) %>% 
  filter(n() == 2) %>% 
  filter(week %in% c(10, 12, 14, 15, 17, 19, 21, 29, 33, 37), 
         #lwa_g_cm2_new < 6, 
        # lwa_g_cm2_new > 0
         ) %>% 
  mutate(week_nice = case_when(
    week == 10 ~ "Week 10", 
    week == 12 ~ "Week 12", 
    week == 14 ~ "Week 14", 
    week == 15 ~ "Week 15", 
    week == 17 ~ "Week 17", 
    week == 19 ~ "Week 19", 
    week == 21 ~ "Week 21", 
    week == 29 ~ "Week 29",
    week == 33 ~ "Week 33",
    week == 37 ~ "Week 37"
  )) 

wk_figs_labels <-
  data.frame(
    label = c('10', '12', '14', '15', '17', '19', '21', '29', '33', '33'),
    week_nice = c(
      "Week 10",
      "Week 12",
       "Week 14",
      "Week 15",
      "Week 17",
      "Week 19",
      "Week 21",
      "Week 29",
      "Week 33",
      "Week 37"
    )
  )

df_all_groups <-  wk_figs %>% 
  mutate(tree_week = paste(tree, week, sep = "_")
  ) %>% 
  mutate(water_potential = water_potential)
  
df_grey <- df_all_groups

df_1 <- df_all_groups %>% 
  filter(week %in% c('10'))

df_2 <- df_all_groups %>% 
  filter(week %in% c('14'))

df_3 <- df_all_groups %>% 
  filter(week %in% c('17'))

df_4 <- df_all_groups %>% 
  filter(week %in% c('37'))

nice_hydration_layered_1 <- ggplot() +
    geom_point(aes(y = water_potential, x = lwa_g_cm2_new, shape = time), data = df_grey, color = "grey") +
    geom_line(aes(y = water_potential, x = lwa_g_cm2_new, group = tree_week), color = "grey", data = df_grey) +
    geom_point(aes(y = water_potential, x = lwa_g_cm2_new, shape = time, color = "df_1"), 
               data = df_1 ) +
    geom_line(aes(y = water_potential, x = lwa_g_cm2_new, group = tree_week, color = "df_1"), 
              data = df_1) +
    geom_point(aes(y = water_potential, x = lwa_g_cm2_new, shape = time, color = "df_2"), 
               data = df_2) +
    geom_line(aes(y = water_potential, x = lwa_g_cm2_new, group = tree_week, color = "df_2"), 
              data = df_2) +
    geom_point(aes(y = water_potential, x = lwa_g_cm2_new, shape = time, color = "df_4"), 
               data = df_4) +
    geom_line(aes(y = water_potential, x = lwa_g_cm2_new, group = tree_week, color = "df_4"), 
              data = df_4
              ) +
     geom_point(aes(y = water_potential, x = lwa_g_cm2_new, shape = time, color = "df_3"), 
               data = df_3) +
    geom_line(aes(y = water_potential, x = lwa_g_cm2_new, group = tree_week, color = "df_3"),
              data = df_3
              ) +
    scale_colour_manual(name = 'Week', 
        values =c("df_1" ="#48578E", 
                   "df_2"="#528791",
                   "df_3"="#8E9D30",
                   "df_4"="#CC5265"
                   ), 
         labels = c('10',
                    '14',
                    '17',
                    '37')) +
    labs(y = "Water Potential (MPa)", 
         #y = " ",
       x = "Leaf Water Content (g)", 
       color= "Week", 
       shape = "Time"
       #caption = "triangles = predawn, circles = midday"
       ) +
  theme(
    legend.position = c(.8, .2),
   # legend.position = "none",
  strip.background = element_blank(),
  legend.key = element_rect(fill = NA),
  #strip.text.y = element_blank(), 
  #strip.text.x = element_text(size = 16), 
  # axis.text.y = element_text(size = 12), 
   axis.text.x = element_text(size = 12), 
   axis.text.y = element_text(size = 12),
    # axis.text.x  = element_blank(),
  #axis.ticks.y = element_blank(), 
  #axis.title.y = element_blank()
  axis.title.x = element_blank(),
  axis.title.y = element_blank(),
  legend.title = element_text(size = 16), 
  legend.text = element_text(size = 13), 
  plot.margin = unit(c(5.5, 5.5, 0, 5.5), units = "pt")
# legend.position = "none"
  ) +
   #ylim(-5,0)+
   #xlim(.5,2) +
  #scale_x_reverse() +
scale_y_continuous(labels = scales::number_format(accuracy = 0.1)) +
  guides(color = "none")
nice_hydration_layered_1
```
#####Space

```{r}
#scaled
space_nice_1 <- wc_wk_centered_df %>% 
  drop_na(site) %>% 
  ggplot(aes(y = water_potential, 
            x = lwa_g_cm2_new,
             color = as.factor(week)
             )) +
  geom_point(alpha = .5, 
            # aes (shape = site)
             ) +
  theme(legend.position="none",
       # legend.position= c(.7, .2),
      strip.background = element_blank(),
      strip.text.x = element_text(size = 12),
      plot.title = element_text(size=13),
     # axis.title = element_text(size = 16),
     axis.text.y = element_text(size = 12),
     axis.text.x = element_text(size = 12),
    # axis.text.x  = element_blank(),
     legend.key=element_blank(), 
      axis.title.x = element_blank(),
     axis.title.y = element_blank(),
     legend.text = element_text(size = 13), 
     legend.title = element_text(size = 16),
     legend.margin=margin(0,0,0,0),
      legend.box.margin=margin(-5,-8,-8,-8),
    plot.margin = unit(c(5.5, 5.5, 0, 5.5), units = "pt")
    ) +
  geom_smooth(method = "lm", se = F,size = .5) + 
 # geom_vline(xintercept = 0, linetype = "dotted") +
  labs(#y= "Middays (MPa)", 
       y = " ",
       x = "Leaf Water per Area", 
       color = "Week", 
       shape = "Site") +
  #color_many_2 +
  color_week_assigned +
 # xlim(.5,2) +
 # ylim(-5,0)+
  guides(color = guide_legend(nrow = 2)) +
  #scale_x_reverse() +
scale_y_continuous(labels = scales::number_format(accuracy = 0.1))
space_nice_1


```
#####Time:

```{r}
time_1 <- wc_qudo_df %>% 
  mutate(time = case_when(
    time == "md" ~ "Midday", 
    time == "pd" ~ "Predawn"
  )) %>% 
  filter(time == "Midday") %>% 
  drop_na(week) %>% 
 # filter(time == "md") %>% 
  ggplot(aes(y = water_potential, 
             x = lwa_g_cm2_new, 
             color = time_season, 
             size = week,
             shape = as.factor(tree),
             show.legend = FALSE)) +
  geom_point(size = 1,
             alpha = .1)+
  geom_smooth(method = "lm", se = F, size = .5, alpha = .4) +
 # theme(legend.position = "none") +
  labs(x = "lwa_g_cm2_new", 
       y = "Midday MPa") +
  #color_two_grey +
  theme(
      legend.position="none",
      strip.background = element_blank(),
     # strip.text.x = element_text(size = 12),
      strip.text.x = element_text(size = 12),
      plot.title = element_text(size=13),
      #axis.title = element_text(size = 15),
     axis.title = element_blank(),
     axis.text = element_text(size = 12),
     axis.text.x = element_text(size = 12),
     legend.key=element_blank(),
     legend.text = element_text(size = 13),
     legend.title = element_text(size = 16),
     #legend.title = element_blank(),
     legend.margin=margin(0,0,0,0),
      legend.box.margin=margin(-5,-8,-8,-8),
     plot.margin = unit(c(5.5, 5.5, 0, 5.5), units = "pt")
    ) +
  labs(#y= "Water Potential (MPa)", 
       y = "Water Potential (MPa)",
       x = "Leaf Water Content (g)", 
       color = "Week", 
       ) +
  guides(color = guide_legend(nrow = 2)) +
 # facet_wrap(~time, nrow = 2) +
  # xlim(.5,2) +
 # ylim(-5,0)+
  #scale_x_reverse()+
  color_two_grey +
scale_y_continuous(labels = scales::number_format(accuracy = 0.1))

time_1 
```


#Slopes: 

####Hydration

Bootstrapped: 
880.5797	701.6674	1074.701

```{r}
hyd_slopes <- final_df_all %>% 
   #filter(!week %in% c(9, 19)) %>% 
  filter(analysis == "hydration, ind. trees") %>% 
  ggplot(aes(color= as.factor(week))
    ) + 
   geom_rect(aes(ymin = 701.6674, ymax = 1074.701, 
          xmin = -Inf, xmax = Inf), 
          fill = "#969696",
          alpha = .01, 
          color = NA) +
  geom_hline(aes(yintercept = 880.5797), color = "#525252") +
  geom_point(aes(y = slope ,x = as.factor(week)), size =2) +
  geom_point(aes(y = upr, x = as.factor(week)), size = 1) +
  geom_point(aes(y = lwr, x = as.factor(week)), size = 1) +
  #geom_line(aes(group = week, y = week, x = upr)) +
  geom_segment(aes(y= lwr, 
                   yend = upr, 
                   x = as.factor(week), 
                   xend = as.factor(week))) +
  labs(x = "Week", 
       color = "week", 
       y = "Slope") +
  theme(legend.position="none",
      strip.background = element_blank(),
      strip.text.x = element_text(size = 12),
      plot.title = element_text(size=13),
      axis.title = element_blank(),
     axis.text = element_text(size = 12),
     legend.key=element_blank(), 
     legend.text = element_text(size = 13), 
     legend.title = element_blank(),
     legend.margin=margin(0,0,0,0),
      legend.box.margin=margin(-5,-8,-8,-8),
     plot.margin = unit(c(5.5, 5.5, 0, 5.5), units = "pt")
    ) +
  geom_hline(yintercept = 0, linetype="dotted") + 
   #scale_x_continuous(breaks = c(10, 12, 14,15, 17, 19, 21, 29, 33, 37)) +
  theme(legend.position = "none") +
  color_week_assigned
hyd_slopes
```

####Space

Bootstrapped: 
```{r}
space_slope_bs <-  final_df_all %>% 
   #filter(!week %in% c(9, 18)) %>% 
  filter(analysis == "space") %>% 
 # mutate(week = as.numeric(week)) %>% 
   ggplot(aes(color= as.factor(week))
    ) + 
   geom_rect(aes(ymin = 701.6674, ymax = 1074.701, 
          xmin = -Inf, xmax = Inf), 
          fill = "#969696",
          alpha = .01, 
          color = NA) +
  geom_hline(aes(yintercept = 880.5797), color = "#525252") +
  geom_point(aes(y = slope ,x = as.factor(week)), size =2) +
    geom_point(aes(y = upr, x = as.factor(week)), size = 1) +
  geom_point(aes(y = lwr, x = as.factor(week)), size = 1) +
  #geom_line(aes(group = week, y = week, x = upr)) +
  geom_segment(aes(y= lwr, 
                   yend = upr, 
                   x = as.factor(week), 
                   xend = as.factor(week), 
                   color = as.factor(week)))+
  labs(x = "Week", 
       color = "Week", 
       y = " ") +
  theme(legend.position="none",
      strip.background = element_blank(),
      strip.text.x = element_text(size = 12),
      plot.title = element_text(size=13),
      axis.title.y = element_blank(),
      axis.title.x = element_blank(),
     axis.text = element_text(size = 12),
     legend.key=element_blank(), 
     legend.text = element_text(size = 13), 
    # legend.title = element_text(size = 16),
    legend.title = element_blank(),
     legend.margin=margin(0,0,0,0),
      legend.box.margin=margin(-5,-8,-8,-8),
    plot.margin = unit(c(5.5, 5.5, 0, 5.5), units = "pt")
    ) +
  #scale_x_continuous(breaks = c(10, 12, 14,15, 17, 19, 21, 29, 33, 37)
                  #   breaks = c(37, 33, 29, 21, 19, 17, 15, 14, 12, 10)
                      # seq(10, 30, by = 1)
  #                   ) +
  geom_hline(yintercept = 0, linetype="dotted") + 
  color_week_assigned+
 # color_many_2 +
  guides(color = guide_legend(nrow = 2)) #+
  #ylim(-8, 35)
 # scale_x_reverse()
space_slope_bs
```


####Time

BS:

```{r}
time_summary <- final_df_all %>% 
   filter(!week %in% c(9)) %>% 
  filter(analysis == "time, ind. tree") %>% 
   mutate(week = as.factor(week), 
          time_season = case_when(
            time_season %in% c("before leafout") ~ "During", 
             time_season %in% c("after leafout") ~ "After", 
            
          )) %>% 
ggplot(aes(color = time_season)) + 
   geom_rect(aes(ymin = 701.6674, ymax = 1074.701, 
          xmin = -Inf, xmax = Inf), 
          fill = "#969696",
          alpha = .01, 
          color = NA) +
  geom_hline(aes(yintercept = 880.5797), color = "#525252") +
  geom_point(aes(y = slope, x = time_season), size =3) +
  geom_point(aes(y = upr, x = time_season), size = 1) +
  geom_point(aes(y = lwr, x = time_season), size = 1) +
  geom_segment(aes(y= lwr, 
                   yend = upr, 
                   x = time_season, 
                   xend = time_season)) +
  labs(y = "",
       x = "Time") +
  color_two_grey +
  geom_hline(yintercept = 0, linetype="dotted") + 
  theme(legend.position="none",
      strip.background = element_blank(),
      strip.text.x = element_text(size = 12),
      plot.title = element_text(size=13),
      axis.title.y = element_blank(),
      #axis.title.x = element_text(size = 13),
      axis.title.x = element_blank(),
     axis.text = element_text(size = 12),
     legend.key=element_blank(), 
     legend.text = element_text(size = 13), 
     legend.title = element_text(size = 16),
     legend.margin=margin(0,0,0,0),
      legend.box.margin=margin(-5,-8,-8,-8),
     plot.margin = unit(c(5.5, 5.5, 0, 5.5), units = "pt")
    ) +
  scale_x_discrete(limits = rev)

time_summary
```

##PV curves

```{r, fig.width=2, fig.height=2}
mean_slope_all <- read_csv(here::here("processed-data", "pv_slopes.csv"))

pv_boot_df <- mean_slope_all %>% 
  select(pre_post, mean_slope, lower_bound, upper_bound) %>% 
  distinct() %>% 
  mutate(analysis = "pv curves") %>% 
  mutate(slope = mean_slope, 
         upr = as.numeric(upper_bound), 
         lwr = as.numeric(lower_bound))


pv_slope <-  pv_boot_df %>% 
   #filter(!pre_post %in% c(9)) %>% 
  #filter(analysis == "space") %>% 
 # mutate(pre_post = as.numeric(pre_post)) %>% 
   ggplot(aes(color= as.factor(pre_post))
    ) + 
  geom_point(aes(y = slope ,x = as.factor(pre_post)), size =2) +
    geom_point(aes(y = upr, x = as.factor(pre_post)), size = 1) +
  geom_point(aes(y = lwr, x = as.factor(pre_post)), size = 1) +
  #geom_line(aes(group = pre_post, y = pre_post, x = upr)) +
  geom_segment(aes(y= lwr, 
                   yend = upr, 
                   x = as.factor(pre_post), 
                   xend = as.factor(pre_post), 
                   color = as.factor(pre_post)))+
  labs(x = "PV", 
       color = "pre_post", 
      # y = "Slope"
       ) +
  theme(legend.position="none",
      strip.background = element_blank(),
      strip.text.x = element_text(size = 12),
      axis.text.x = element_blank(),
      plot.title = element_text(size=13),
     # axis.title.y = element_text(size = 15),
      axis.title.y = element_blank(),
    #  axis.title.x = element_text(size = 15),
     # axis.title.x = element_blank(),
     axis.text = element_text(size = 12),
     legend.key=element_blank(), 
     legend.text = element_text(size = 13), 
    # legend.title = element_text(size = 16),
    legend.title = element_blank(),
     legend.margin=margin(0,0,0,0),
      legend.box.margin=margin(-5,-8,-8,-8),
    plot.margin = unit(c(5.5, 5.5, 0, 5.5), units = "pt")
    ) +
  geom_hline(yintercept = 0, linetype="dotted") +
 # color_pre_post_assigned+
 # color_many_2 +
  guides(color = guide_legend(nrow = 2)) +
 scale_color_manual(values = c("#525252")) 
 # scale_x_reverse()
pv_slope

ggsave(here("figures", "qudo figures", "LWC", "pv_slope_plot_qudo"), pv_slope, device = "jpg", width = 1, height = 3, dpi = 300)
```


```{r, fig.width=2, fig.height=2}
empty_plot <-  pv_boot_df %>% 
   filter(pre_post = NA) %>% #we want an empty plot
  #filter(analysis == "space") %>% 
 # mutate(pre_post = as.numeric(pre_post)) %>% 
   ggplot(aes(color= as.factor(pre_post))
    ) + 
  geom_point(aes(y = slope ,x = as.factor(pre_post)), size =2) +
    geom_point(aes(y = upr, x = as.factor(pre_post)), size = 1) +
  geom_point(aes(y = lwr, x = as.factor(pre_post)), size = 1) +
  #geom_line(aes(group = pre_post, y = pre_post, x = upr)) +
  geom_segment(aes(y= lwr, 
                   yend = upr, 
                   x = as.factor(pre_post), 
                   xend = as.factor(pre_post), 
                   color = as.factor(pre_post)))+
  labs(x = "PV", 
       color = "pre_post", 
      # y = "Slope"
       ) +
  theme(legend.position="none",
      strip.background = element_blank(),
      strip.text.x = element_text(size = 12),
      axis.text.x = element_blank(),
      plot.title = element_text(size=13),
     # axis.title.y = element_text(size = 15),
      axis.title.y = element_blank(),
      axis.title.x = element_text(size = 15),
     # axis.title.x = element_blank(),
     axis.text = element_text(size = 12),
     legend.key=element_blank(), 
     legend.text = element_text(size = 13), 
    # legend.title = element_text(size = 16),
    legend.title = element_blank(),
     legend.margin=margin(0,0,0,0),
      legend.box.margin=margin(-5,-8,-8,-8)
    ) +
  geom_hline(yintercept = 0, linetype="dotted") +
 # color_pre_post_assigned+
 # color_many_2 +
  guides(color = guide_legend(nrow = 2)) +
 scale_color_manual(values = c("darkgrey", "darkgrey")) 
 # scale_x_reverse()
empty_plot
```

######Legend:

```{r}
legend_plot <- wk_figs %>% 
  drop_na(site) %>% 
  mutate(time = case_when(
    time %in% c("md") ~ "Midday",
    time %in% c("pd") ~ "Predawn"
  )) %>% 
  ggplot(aes(y = water_potential,
             x = lwa_g_cm2_new,
             color = as.factor(week), 
            # shape = as.factor(time),
             )) +
  geom_point() +
  theme(legend.position="top",
       # legend.position= c(.7, .2),
      strip.background = element_blank(),
      strip.text.x = element_text(size = 12),
      plot.title = element_text(size=13),
     # axis.title = element_text(size = 16),
     axis.text = element_text(size = 12),
     legend.key=element_blank(), 
     axis.title.x = element_blank(),
     axis.title.y = element_blank(),
     legend.text = element_text(size = 13), 
     legend.title = element_text(size = 16),
     legend.margin=margin(0,0,0,0),
    legend.box.margin=margin(-5,-8,-8,-8)
    ) +
  geom_smooth(method = "lm", se = F, size = .5) + 
  labs(color = "Week", 
       shape = "Time") +
  color_week_assigned +
 # ylim(0,2) +
  guides(color = guide_legend(nrow = 1))
legend_plot

legend_scatterplots <- cowplot::get_legend(legend_plot)
legend_scatterplots
```

#COMBINE

```{r}
library(grid)
#labels of slopes: 
y.grob1 <- textGrob("Slope", 
                   gp=gpar(fontface="bold", #col="blue", 
                           fontsize=15), rot=90)

x.grob1 <- textGrob("Time", 
                   gp=gpar(fontface="bold", #col="blue", 
                           fontsize=15))

combined_slopes_wide <- cowplot::plot_grid(space_slope_bs, NULL, time_summary, NULL, hyd_slopes,
                                           # NULL, pv_slope,
                                   nrow = 1, 
                                  rel_widths = c(1, -.02, 1, -.02, 1 #, -.02, .5
                                                 ), 
                                  labels = c("B", "", "D", "", "F"#,"", "G"
                                             ),
                                  label_x = .2,
                                  hjust= c(3, 0, 3, 0, 3),
                                  label_y = .98
                                  # rel_heights = c(.15, 1, .3), 
                                   #rel_widths = c(1, 1, 1), 
                                  # rel_widths = c(.7, .8, .7, .3)
                                   ) 

combined_slopes_wide

combined_slopes_wide_labels <- grid.arrange(arrangeGrob(combined_slopes_wide, 
                            left = y.grob1, 
                            bottom = x.grob1),
                            heights = c(1, .05))

#labels of scatterplots: 
y.grob2 <- textGrob("Water Potential (MPa)", 
                   gp=gpar(fontface="bold", #col="blue", 
                           fontsize=15), rot=90)

x.grob2 <- textGrob("Leaf Water per Area (g/cm2)", 
                   gp=gpar(fontface="bold", #col="blue",
                           fontsize=15))


combined_plots_wide_nolegend <- cowplot::plot_grid(space_nice_1, NULL, 
                                      time_1, NULL, 
                                   nice_hydration_layered_1,
                                   labels=c("A","", "C", "", "E"),
                                   label_x = .112,
                                   label_y = .98,
                                   nrow = 1,
                                   rel_widths = c(1, -.015, 1, -.015, 1)
                                   #rel_widths = c(1, 1, 1)
                                   ) 
combined_plots_wide_nolegend

combined_plots_wide <- cowplot::plot_grid(legend_scatterplots,
                                      combined_plots_wide_nolegend,
                                          nrow = 2, 
                                          rel_heights = c(.1, 1))

combined_plots_wide_labels <- grid.arrange(arrangeGrob(combined_plots_wide, 
                                left = y.grob2, 
                                bottom = x.grob2,
                                heights = c(1, 0)))

combined_plots_all <- cowplot::plot_grid(
  combined_plots_wide_labels, 
  combined_slopes_wide_labels, 
          nrow = 2, 
          rel_heights = c(.75, .75))

combined_plots_all

ggsave(here("figures", "qudo figures", "LWA", "combined_plots_all_qudo_new"), combined_plots_all, device = "jpg", width = 12, height = 8, dpi = 300)
```
