==> scripts/5.2_ewt_firstpass.Rmd <==
---
title: "Adding in EWT analysis!"
author: "Indra Boving"
date: "3/14/2023"
output: html_document
---

##setup
```{r setup, include=FALSE}
#install.packages("MuMIn")
#devtools::install_github("an-bui/calecopal")
library(data.table)
library(janitor)
library(here)
library(tidyverse)
library(lubridate)
library(readxl)
library(gridExtra)
library(calecopal)
library(rstatix)
library(vars)
library(MuMIn)
library(lme4) #for mixed effects models
library(cowplot)
select = dplyr::select
library('variancePartition')
source(here::here("scripts", "scripts_functions", "plotPercentBars_IB.R"))
source(here::here("scripts", "scripts_functions", "figure_info.R"))

datver_ewt <- "20230220"
dataversion_ewt  <- paste0("Data_", datver_ewt)
```

```{r}
ewt_allen <- read_csv(here("data", "ewt_all_firstpass.csv"), show_col_types = F) %>% 
  clean_names() %>% 
  filter(sample_type == "tree") %>% 
  mutate(date = mdy(date), 
         tree = as.numeric(tree)) %>% 
  drop_na(tree) %>% 
  select(-row_nr, -date_str)

#Attach to water pot. data: 

ewt_lwc_raw <- read_csv(here("processed-data", paste0("wp_wc_rwc_",datver_ewt ,".csv")), show_col_types = F) %>% 
  mutate(lwc_mean = lwc_mean) %>% 
  # filter(lwc_mean > 0, 
  #        lwc_mean < 3, 
  #        time == "md") %>% 
  select(tree, week, mpa_mean, lwc_mean, lwa_g_cm2)

#Combine: 

ewt_lwc_df <- merge(ewt_allen, ewt_lwc_raw, by = c("tree", "week"))
```

vizzz

```{r}
ewt_lwc_df %>% 
  ggplot(aes(y = ewt, x = lwc_mean, color = species
             )) +
  geom_point()

ewt_lwc_df %>% 
  ggplot(aes(y = ewt, x = lwa_g_cm2, color = species
             )) +
  geom_point()

ewt_lwc_df %>% 
  ggplot(aes(y = ewt, x = mpa_mean, color = species
             )) +
  geom_point()
```
